% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\name{add_missing_stan_blocks}
\alias{add_missing_stan_blocks}
\title{Add Missing Stan Blocks}
\usage{
add_missing_stan_blocks(x, stan_blocks = STAN_BLOCKS)
}
\arguments{
\item{x}{(\code{list})\cr list of Stan code blocks}

\item{stan_blocks}{(\code{list})\cr reference list of stan blocks.}
}
\value{
Amended list \code{x} such that all blocks in the global variable
\code{STAN_BLOCKS} are contained.
}
\description{
Add Missing Stan Blocks
}
\keyword{internal}
