% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimJointData.R
\name{add_pfs}
\alias{add_pfs}
\title{Add PFS events at Tumour Progression to Data}
\usage{
add_pfs(
  object,
  relative_threshold = 1.2,
  absolute_threshold = 5,
  from_time = 0,
  observed_after = FALSE
)
}
\arguments{
\item{object}{(\code{SimJointData}) where the PFS events should be added.}

\item{relative_threshold}{(\code{number}) a multiplicative threshold for the change in SLD compared to the \code{min(SLD)}.
Default is 1.2 meaning a 20\% increase.}

\item{absolute_threshold}{(\code{number}) an absolute threshold for the change in SLD compared to the minimum.
Default is 5.}

\item{from_time}{(\code{number}) Ignore observations before this time for determining SLD minimum.}

\item{observed_after}{(\code{logical}) If \code{FALSE} set longitudinal observations after the progression time to
\code{observed = FALSE}}
}
\value{
The input \code{SimJointData} object with progression-free survival columns
\code{pfs_time} and \code{pfs_event} added.
}
\description{
Adds new columns \code{pfs_time} and \code{pfs_event} based on observed changes to SLD.
}
\details{
Both thresholds must be met for a progression to be declared.
}
\examples{
data <- SimJointData(
  survival = SimSurvivalExponential(lambda = 1/10),
  longitudinal = SimLongitudinalSteinFojo()
  )
data <- add_pfs(data)
data@survival # now has pfs_time and pfs_event columns

}
