% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{addLink}
\alias{addLink}
\title{\code{addLink}}
\usage{
addLink(x, y, ...)
}
\arguments{
\item{x}{the longitudinal model.}

\item{y}{the link to be added.}

\item{...}{additional arguments.}
}
\description{
Add a link to a longitudinal model.
}
