% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/SurvivalSamples.R
\name{aggregate}
\alias{aggregate}
\alias{aggregate,SurvivalSamples-method}
\title{Aggregation Methods for Different Classes}
\usage{
\S4method{aggregate}{SurvivalSamples}(x, groups, ...)
}
\arguments{
\item{x}{what to aggregate.}

\item{groups}{(\code{list})\cr defining into which groups to aggregate
individual samples, where the names are the new group labels and
the character vectors are the old individual sample labels.}

\item{...}{other arguments passed to aggregation methods.}
}
\description{
These aggregation methods allow to group samples of different objects.
}
