% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_model.R
\name{as.character,JointModel-method}
\alias{as.character,JointModel-method}
\title{as character joint model TODO}
\usage{
\S4method{as.character}{JointModel}(x)
}
\arguments{
\item{x}{A \code{JointModel} object}
}
\description{
as character joint model TODO
}
