% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_module.R
\name{as.character,StanModule-method}
\alias{as.character,StanModule-method}
\title{Convert a StanModule object into stan code}
\usage{
\S4method{as.character}{StanModule}(x)
}
\arguments{
\item{x}{A \code{StanModule} object}
}
\description{
Collapses a StanModule object down into a single string inserting the required block
fences
i.e. \verb{data \{ ... \}}
}
