% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModel.R
\name{as.character.JointModel}
\alias{as.character.JointModel}
\title{\code{JointModel} -> \code{character}}
\usage{
\method{as.character}{JointModel}(x, ...)
}
\arguments{
\item{x}{(\code{JointModel}) Joint model specification.}

\item{...}{Not Used.}
}
\value{
A character vector.
}
\description{
Renders a \code{\link{JointModel}} object to a stan program
}
\seealso{
Other JointModel:
\code{\link{JointModel-class}},
\code{\link[=as.StanModule.JointModel]{as.StanModule.JointModel()}}
}
\concept{JointModel}
