% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{as.character.Parameter}
\alias{as.character.Parameter}
\title{\code{Parameter} -> \code{Character}}
\usage{
\method{as.character}{Parameter}(x, ...)
}
\arguments{
\item{x}{(\code{Parameter}) a prior Distribution}

\item{...}{Not Used.}
}
\value{
A character vector.
}
\description{
Converts a \code{\link{Parameter}} object to a character vector
}
\seealso{
Other Parameter:
\code{\link{Parameter-Getter-Methods}},
\code{\link{Parameter-class}},
\code{\link[=as.StanModule.Parameter]{as.StanModule.Parameter()}},
\code{\link[=as_stan_list.Parameter]{as_stan_list.Parameter()}}
}
\concept{Parameter}
