% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterList.R
\name{as.character.ParameterList}
\alias{as.character.ParameterList}
\title{\code{ParameterList} -> \code{character}}
\usage{
\method{as.character}{ParameterList}(x, ...)
}
\arguments{
\item{x}{(\code{ParameterList}) \cr A List of \code{\link{Parameter}} Objects.}

\item{...}{Not Used.}
}
\description{
Converts a \code{\link{ParameterList}} to a Stan model-block indented by 4 spaces
}
\seealso{
Other ParameterList: 
\code{\link{ParameterList-Getter-Methods}},
\code{\link{ParameterList-class}},
\code{\link{as.StanModule.ParameterList}()},
\code{\link{as.list.ParameterList}()}
}
\concept{ParameterList}
