% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{as.character.Prior}
\alias{as.character.Prior}
\title{\code{Prior} -> \code{Character}}
\usage{
\method{as.character}{Prior}(x, ...)
}
\arguments{
\item{x}{(\code{Prior}) a prior Distribution}

\item{...}{Not Used.}
}
\value{
A character vector.
}
\description{
Converts a \code{\link{Prior}} object to a character vector
}
\seealso{
Other Prior-internal:
\code{\link{Prior-Getter-Methods}},
\code{\link{Prior-class}},
\code{\link[=as.StanModule.Prior]{as.StanModule.Prior()}},
\code{\link[=as_stan_list.Prior]{as_stan_list.Prior()}}
}
\concept{Prior-internal}
