% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Prior.R, R/Parameter.R,
%   R/ParameterList.R, R/StanModule.R, R/JointModel.R
\name{as.character}
\alias{as.character}
\alias{as.character,Prior-method}
\alias{coerce-Prior-character-method}
\alias{coerce,Prior,character-method}
\alias{as.character,Parameter-method}
\alias{coerce-Parameter-character-method}
\alias{coerce,Parameter,character-method}
\alias{coerce-ParameterList-character-method}
\alias{coerce,ParameterList,character-method}
\alias{as.character,ParameterList-method}
\alias{as.character,StanModule-method}
\alias{as.character,JointModel-method}
\title{Character Conversion of Different Classes}
\usage{
\S4method{as.character}{Prior}(x)

\S4method{as.character}{Parameter}(x)

\S4method{as.character}{ParameterList}(x)

\S4method{as.character}{StanModule}(x)

\S4method{as.character}{JointModel}(x)
}
\arguments{
\item{x}{what to convert.}

\item{...}{not used.}
}
\description{
These methods convert objects of different classes to \code{character}
representation.
}
