% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{as.data.frame,SurvivalQuantities-method}
\alias{as.data.frame,SurvivalQuantities-method}
\title{\code{as.data.frame}}
\usage{
\S4method{as.data.frame}{SurvivalQuantities}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{(\code{\link{SurvivalQuantities}}) \cr Survival Quantities}

\item{row.names}{not used.}

\item{optional}{not used.}

\item{...}{Not used}
}
\description{
\code{as.data.frame}
}
\seealso{
Other \code{SurvivalQuantities}: 
\code{\link{SurvivalQuantities-class}},
\code{\link{autoplot,SurvivalQuantities-method}},
\code{\link{summary,SurvivalQuantities-method}}
}
\concept{\code{SurvivalQuantities}}
\concept{\code{as.data.frame}}
