% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLongitudinal.R
\name{as.data.frame.DataLongitudinal}
\alias{as.data.frame.DataLongitudinal}
\title{\code{DataLongitudinal} -> \code{data.frame}}
\usage{
\method{as.data.frame}{DataLongitudinal}(x, ...)
}
\arguments{
\item{x}{(\code{DataLongitudinal}) Longitudinal Data.}

\item{...}{Not Used.}
}
\value{
A \code{data.frame} representation of the object.
}
\description{
Converts a \code{\link{DataLongitudinal}} object into a \code{data.frame}.
The subject variable is cast to factor.
}
\seealso{
Other DataLongitudinal:
\code{\link{DataLongitudinal-class}},
\code{\link[=as_print_string.DataLongitudinal]{as_print_string.DataLongitudinal()}},
\code{\link[=as_stan_list.DataSubject]{as_stan_list.DataSubject()}},
\code{\link[=extractVariableNames.DataLongitudinal]{extractVariableNames.DataLongitudinal()}}
}
\concept{DataLongitudinal}
