% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\name{as.data.frame.DataSurvival}
\alias{as.data.frame.DataSurvival}
\title{\code{DataSurvival} -> \code{data.frame}}
\usage{
\method{as.data.frame}{DataSurvival}(x, ...)
}
\arguments{
\item{x}{(\code{DataSurvival}) Survival Data.}

\item{...}{Not Used.}
}
\value{
A \code{data.frame} representation of the object.
}
\description{
Converts a \code{\link{DataSurvival}} object into a \code{data.frame}.
The subject variable is cast to factor.
}
\seealso{
Other DataSurvival:
\code{\link{DataSurvival-class}},
\code{\link[=as_print_string.DataSurvival]{as_print_string.DataSurvival()}},
\code{\link[=as_stan_list.DataSubject]{as_stan_list.DataSubject()}},
\code{\link[=extractVariableNames.DataSurvival]{extractVariableNames.DataSurvival()}}
}
\concept{DataSurvival}
