% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalQuantities.R
\name{as.data.frame.LongitudinalQuantities}
\alias{as.data.frame.LongitudinalQuantities}
\title{\code{as.data.frame}}
\usage{
\method{as.data.frame}{LongitudinalQuantities}(x, ...)
}
\arguments{
\item{x}{(\code{LongitudinalQuantities}) longitudinal quantities.}

\item{...}{not used.}
}
\value{
A \code{data.frame} representation of the object.
}
\description{
\code{as.data.frame}
}
\seealso{
Other LongitudinalQuantities:
\code{\link{LongitudinalQuantities-class}},
\code{\link[=autoplot.LongitudinalQuantities]{autoplot.LongitudinalQuantities()}},
\code{\link[=summary.LongitudinalQuantities]{summary.LongitudinalQuantities()}}
}
\concept{LongitudinalQuantities}
