% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\name{as.data.frame.Quantities}
\alias{as.data.frame.Quantities}
\title{\code{Quantities} -> \code{data.frame}}
\usage{
\method{as.data.frame}{Quantities}(x, ...)
}
\arguments{
\item{x}{(\code{Quantities}) generated quantities.}

\item{...}{not used.}
}
\value{
A \code{data.frame} representation of the object.
}
\description{
\code{Quantities} -> \code{data.frame}
}
\seealso{
Other Quantities:
\code{\link{Quantities-class}},
\code{\link[=as_print_string.Quantities]{as_print_string.Quantities()}},
\code{\link[=summary.Quantities]{summary.Quantities()}}
}
\concept{Quantities}
\keyword{internal}
