% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/DataLongitudinal.R,
%   R/DataSurvival.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,DataLongitudinal-method}
\alias{as.data.frame,DataSurvival-method}
\alias{coerce-DataSurvival-data.frame-method}
\alias{coerce,DataSurvival,data.frame-method}
\title{Data Frame Conversion of Different Classes}
\usage{
\S4method{as.data.frame}{DataLongitudinal}(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.data.frame}{DataSurvival}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{what to convert.}

\item{row.names}{not used.}

\item{optional}{not used.}

\item{...}{not used.}
}
\description{
These methods convert objects of different classes to \code{data.frame}
representation.
}
