% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{as.data.frame.SurvivalQuantities}
\alias{as.data.frame.SurvivalQuantities}
\title{\code{as.data.frame}}
\usage{
\method{as.data.frame}{SurvivalQuantities}(x, ...)
}
\arguments{
\item{x}{(\code{SurvivalQuantities}) longitudinal quantities.}

\item{...}{not used.}
}
\value{
A \code{data.frame} representation of the object.
}
\description{
\code{as.data.frame}
}
\seealso{
Other SurvivalQuantities:
\code{\link{SurvivalQuantities-class}},
\code{\link[=autoplot.SurvivalQuantities]{autoplot.SurvivalQuantities()}},
\code{\link[=brierScore.SurvivalQuantities]{brierScore.SurvivalQuantities()}},
\code{\link[=summary.SurvivalQuantities]{summary.SurvivalQuantities()}}
}
\concept{SurvivalQuantities}
