% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\name{as.list.DataJoint}
\alias{as.list.DataJoint}
\title{\code{DataJoint} -> \code{list}}
\usage{
\method{as.list}{DataJoint}(x, ...)
}
\arguments{
\item{x}{(\code{\link{DataJoint}}) \cr Survival and Longitudinal Data.}

\item{...}{Not Used.}
}
\description{
Coerces  \code{\link{DataJoint}} into a \code{list} of data components required
for fitting a \code{\link{JointModel}}. See the vignette (TODO) for more details.
}
\seealso{
\code{\link[=as.list.DataSurvival]{as.list.DataSurvival()}}, \code{\link[=as.list.DataLongitudinal]{as.list.DataLongitudinal()}}

Other DataJoint: 
\code{\link{DataJoint-class}},
\code{\link{subset.DataJoint}()}
}
\concept{DataJoint}
