% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLongitudinal.R
\name{as.list.DataLongitudinal}
\alias{as.list.DataLongitudinal}
\title{\code{DataLongitudinal} -> \code{list}}
\usage{
\method{as.list}{DataLongitudinal}(x, ...)
}
\arguments{
\item{x}{(\code{\link{DataLongitudinal}}) \cr Longitudinal Data.}

\item{...}{Not Used.}
}
\description{
Coerces  \code{\link{DataLongitudinal}} into a \code{list} of data components required
for fitting a \code{\link{JointModel}}. See the vignette (TODO) for more details.
}
\seealso{
Other DataLongitudinal: 
\code{\link{DataLongitudinal-class}},
\code{\link{as.data.frame.DataLongitudinal}()},
\code{\link{extractVariableNames.DataLongitudinal}()}
}
\concept{DataLongitudinal}
