% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link.R
\name{as.list.Link}
\alias{as.list.Link}
\title{\code{Link} -> \code{list}}
\usage{
\method{as.list}{Link}(x, ...)
}
\arguments{
\item{x}{(\code{Link}) a link object.}

\item{...}{Not Used.}
}
\value{
A named \code{list} representation of the object.
}
\description{
Returns a named list where each element of the list corresponds
to a Stan modelling block e.g. \code{data}, \code{model}, etc.
}
\seealso{
Other Link:
\code{\link{Link-class}},
\code{\link[=as.StanModule.Link]{as.StanModule.Link()}},
\code{\link[=length.Link]{length.Link()}}
}
\concept{Link}
