% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkComponent.R
\name{as.list.LinkComponent}
\alias{as.list.LinkComponent}
\title{\code{LinkComponent} -> \code{list}}
\usage{
\method{as.list}{LinkComponent}(x, ...)
}
\arguments{
\item{x}{(\code{LinkComponent}) a link component.}

\item{...}{Not Used.}
}
\value{
A named \code{list} representation of the object.
}
\description{
Returns a named list where each element of the list corresponds
to a Stan modelling block e.g. \code{data}, \code{model}, etc.
}
\seealso{
Other LinkComponent:
\code{\link{LinkComponent-class}},
\code{\link[=as.StanModule.LinkComponent]{as.StanModule.LinkComponent()}},
\code{\link[=getParameters]{getParameters()}},
\code{\link[=initialValues]{initialValues()}}
}
\concept{LinkComponent}
