% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/DataLongitudinal.R,
%   R/DataSurvival.R, R/DataJoint.R, R/ParameterList.R, R/StanModule.R,
%   R/StanModel.R
\name{as.list}
\alias{as.list}
\alias{as.list,DataLongitudinal-method}
\alias{as.list,DataSurvival-method}
\alias{as.list,DataJoint-method}
\alias{coerce-DataJoint-list-method}
\alias{coerce,DataJoint,list-method}
\alias{as.list,ParameterList-method}
\alias{as.list,StanModule-method}
\alias{as.list,StanModel-method}
\title{List Conversion of Different Classes}
\usage{
\S4method{as.list}{DataLongitudinal}(x)

\S4method{as.list}{DataSurvival}(x, ...)

\S4method{as.list}{DataJoint}(x)

\S4method{as.list}{ParameterList}(x)

\S4method{as.list}{StanModule}(x)

\S4method{as.list}{StanModel}(x)
}
\arguments{
\item{x}{(\code{DataJoint}) \cr A \link[=DataJoint-class]{DataJoint} object created by \code{\link[=DataJoint]{DataJoint()}}}

\item{...}{not used.}
}
\description{
These methods convert objects of different classes to \code{list}
representation.
}
