% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModel.R
\name{as.StanModule.JointModel}
\alias{as.StanModule.JointModel}
\title{\code{JointModel} -> \code{StanModule}}
\usage{
\method{as.StanModule}{JointModel}(object, ...)
}
\arguments{
\item{object}{(\code{JointModel}) Joint model specification.}

\item{...}{Not Used.}
}
\value{
A \code{StanModule} object.
}
\description{
Converts a \code{\link{JointModel}} object to a \code{\link{StanModule}} object
}
\seealso{
Other JointModel:
\code{\link{JointModel-class}},
\code{\link[=as.character.JointModel]{as.character.JointModel()}}

Other as.StanModule:
\code{\link[=as.StanModule]{as.StanModule()}},
\code{\link[=as.StanModule.Link]{as.StanModule.Link()}},
\code{\link[=as.StanModule.LinkComponent]{as.StanModule.LinkComponent()}},
\code{\link[=as.StanModule.Parameter]{as.StanModule.Parameter()}},
\code{\link[=as.StanModule.ParameterList]{as.StanModule.ParameterList()}},
\code{\link[=as.StanModule.Prior]{as.StanModule.Prior()}}
}
\concept{JointModel}
\concept{as.StanModule}
