% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkComponent.R
\name{as.StanModule.LinkComponent}
\alias{as.StanModule.LinkComponent}
\title{\code{LinkComponent} -> \code{StanModule}}
\usage{
\method{as.StanModule}{LinkComponent}(object, ...)
}
\arguments{
\item{object}{(\code{LinkComponent}) a link component.}

\item{...}{Not Used.}
}
\value{
A \code{StanModule} object.
}
\description{
Converts a \code{\link{LinkComponent}} object to a \code{\link{StanModule}} object
}
\seealso{
Other LinkComponent:
\code{\link{LinkComponent-class}},
\code{\link[=as.list.LinkComponent]{as.list.LinkComponent()}},
\code{\link[=getParameters]{getParameters()}},
\code{\link[=initialValues]{initialValues()}}

Other as.StanModule:
\code{\link[=as.StanModule]{as.StanModule()}},
\code{\link[=as.StanModule.JointModel]{as.StanModule.JointModel()}},
\code{\link[=as.StanModule.Link]{as.StanModule.Link()}},
\code{\link[=as.StanModule.Parameter]{as.StanModule.Parameter()}},
\code{\link[=as.StanModule.ParameterList]{as.StanModule.ParameterList()}},
\code{\link[=as.StanModule.Prior]{as.StanModule.Prior()}}
}
\concept{LinkComponent}
\concept{as.StanModule}
