% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{as.StanModule.Parameter}
\alias{as.StanModule.Parameter}
\title{\code{Parameter} -> \code{StanModule}}
\usage{
\method{as.StanModule}{Parameter}(object, ...)
}
\arguments{
\item{object}{(\code{Parameter}) a prior Distribution}

\item{...}{Not Used.}
}
\value{
A \code{StanModule} object.
}
\description{
Converts a \code{\link{Parameter}} object to a \code{\link{StanModule}} object
}
\seealso{
Other Parameter:
\code{\link{Parameter-Getter-Methods}},
\code{\link{Parameter-class}},
\code{\link[=as.character.Parameter]{as.character.Parameter()}},
\code{\link[=as_stan_list.Parameter]{as_stan_list.Parameter()}}

Other as.StanModule:
\code{\link[=as.StanModule]{as.StanModule()}},
\code{\link[=as.StanModule.JointModel]{as.StanModule.JointModel()}},
\code{\link[=as.StanModule.Link]{as.StanModule.Link()}},
\code{\link[=as.StanModule.LinkComponent]{as.StanModule.LinkComponent()}},
\code{\link[=as.StanModule.ParameterList]{as.StanModule.ParameterList()}},
\code{\link[=as.StanModule.Prior]{as.StanModule.Prior()}}
}
\concept{Parameter}
\concept{as.StanModule}
