% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_formula}
\alias{as_formula}
\title{As Formula}
\usage{
as_formula(x, ...)
}
\arguments{
\item{x}{(\code{ANY}) object to convert to a formula.}

\item{...}{Not used.}
}
\value{
A \code{formula} object.
}
\description{
Utility wrapper function to convert an object to a formula.
}
\examples{
as_formula("response ~ time")
}
