% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLongitudinal.R
\name{as_print_string.DataLongitudinal}
\alias{as_print_string.DataLongitudinal}
\title{\code{DataLongitudinal} -> Printable \code{Character}}
\usage{
\method{as_print_string}{DataLongitudinal}(object, indent = 1, ...)
}
\arguments{
\item{object}{(\code{DataLongitudinal}) Longitudinal Data.}

\item{indent}{(\code{numeric}) how much white space to prefix the print string with.}

\item{...}{Not Used.}
}
\value{
A character vector suitable for printing.
}
\description{
Converts \code{\link{DataLongitudinal}} object into a printable string.
}
\seealso{
Other DataLongitudinal:
\code{\link{DataLongitudinal-class}},
\code{\link[=as.data.frame.DataLongitudinal]{as.data.frame.DataLongitudinal()}},
\code{\link[=as_stan_list.DataSubject]{as_stan_list.DataSubject()}},
\code{\link[=extractVariableNames.DataLongitudinal]{extractVariableNames.DataLongitudinal()}}
}
\concept{DataLongitudinal}
\keyword{internal}
