% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSubject.R
\name{as_print_string.DataSubject}
\alias{as_print_string.DataSubject}
\title{\code{DataSubject} -> Printable \code{Character}}
\usage{
\method{as_print_string}{DataSubject}(object, indent = 1, ...)
}
\arguments{
\item{object}{(\code{DataSubject}) subject-level data.}

\item{indent}{(\code{numeric}) how much white space to prefix the print string with.}

\item{...}{Not Used.}
}
\value{
A character vector suitable for printing.
}
\description{
Converts \code{\link{DataSubject}} object into a printable string.
}
\seealso{
Other DataSubject:
\code{\link{DataSubject-class}},
\code{\link[=as.data.frame.DataSubject]{as.data.frame.DataSubject()}},
\code{\link[=as_stan_list.DataSubject]{as_stan_list.DataSubject()}},
\code{\link[=extractVariableNames.DataSubject]{extractVariableNames.DataSubject()}}
}
\concept{DataSubject}
\keyword{internal}
