% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\name{as_print_string.DataSurvival}
\alias{as_print_string.DataSurvival}
\title{\code{DataSurvival} -> Printable \code{Character}}
\usage{
\method{as_print_string}{DataSurvival}(object, indent = 1, ...)
}
\arguments{
\item{object}{(\code{DataSurvival}) Survival Data.}

\item{indent}{(\code{numeric}) how much white space to prefix the print string with.}

\item{...}{Not Used.}
}
\value{
A character vector suitable for printing.
}
\description{
Converts \code{\link{DataSurvival}} object into a printable string.
}
\seealso{
Other DataSurvival:
\code{\link{DataSurvival-class}},
\code{\link[=as.data.frame.DataSurvival]{as.data.frame.DataSurvival()}},
\code{\link[=as_stan_list.DataSubject]{as_stan_list.DataSubject()}},
\code{\link[=extractVariableNames.DataSurvival]{extractVariableNames.DataSurvival()}}
}
\concept{DataSurvival}
\keyword{internal}
