% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModelSamples.R
\name{as_print_string.JointModelSamples}
\alias{as_print_string.JointModelSamples}
\title{\code{JointModelSamples} -> Printable \code{Character}}
\usage{
\method{as_print_string}{JointModelSamples}(object, indent = 1, ...)
}
\arguments{
\item{object}{(\code{JointModelSamples}) samples as drawn from a \code{\link{JointModel}}.}

\item{indent}{(\code{numeric}) how much white space to prefix the print string with.}
}
\value{
A character vector suitable for printing.
}
\description{
Converts \code{\link{JointModelSamples}} object into a printable string.
}
\seealso{
Other JointModelSamples:
\code{\link[=LongitudinalRandomEffects]{LongitudinalRandomEffects()}}
}
\concept{JointModelSamples}
\keyword{internal}
