% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalQuantities.R
\name{as_print_string.LongitudinalQuantities}
\alias{as_print_string.LongitudinalQuantities}
\title{\code{LongitudinalQuantities} -> Printable \code{Character}}
\usage{
\method{as_print_string}{LongitudinalQuantities}(object, indent = 1, ...)
}
\arguments{
\item{object}{(\code{\link{LongitudinalQuantities}}) \cr longitudinal quantities.}

\item{indent}{(\code{numeric})\cr how much white space to prefix the print string with.}

\item{...}{not used.}
}
\description{
Converts \code{\link{LongitudinalQuantities}} object into a printable string.
}
\seealso{
Other LongitudinalQuantities: 
\code{\link{LongitudinalQuantities-class}},
\code{\link{as.data.frame.LongitudinalQuantities}()},
\code{\link{autoplot.LongitudinalQuantities}()},
\code{\link{summary.LongitudinalQuantities}()}
}
\concept{LongitudinalQuantities}
\keyword{internal}
