% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterList.R
\name{as_print_string.ParameterList}
\alias{as_print_string.ParameterList}
\title{\code{ParameterList} -> Printable \code{Character}}
\usage{
\method{as_print_string}{ParameterList}(object, ...)
}
\arguments{
\item{object}{(\code{ParameterList}) A List of \code{\link{Parameter}} Objects.}

\item{...}{Not Used.}
}
\value{
A character vector suitable for printing.
}
\description{
Converts \code{\link{ParameterList}} object into a printable string.
}
\seealso{
Other ParameterList:
\code{\link{ParameterList-Getter-Methods}},
\code{\link{ParameterList-class}},
\code{\link[=as.StanModule.ParameterList]{as.StanModule.ParameterList()}},
\code{\link[=as.list.ParameterList]{as.list.ParameterList()}},
\code{\link[=as_stan_list.ParameterList]{as_stan_list.ParameterList()}}
}
\concept{ParameterList}
\keyword{internal}
