% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/SimLongitudinal.R,
%   R/SimLongitudinalClaretBruno.R, R/SimLongitudinalClaretBrunoCov.R,
%   R/SimLongitudinalGSF.R, R/SimLongitudinalGSFCov.R,
%   R/SimLongitudinalRandomSlope.R, R/SimLongitudinalRandomSlopeCov.R,
%   R/SimLongitudinalSteinFojo.R, R/SimLongitudinalSteinFojoCov.R,
%   R/SimSurvival.R
\name{as_print_string}
\alias{as_print_string}
\alias{as_print_string.SimLongitudinal}
\alias{as_print_string.SimLongitudinalClaretBruno}
\alias{as_print_string.SimLongitudinalClaretBrunoCov}
\alias{as_print_string.SimLongitudinalGSF}
\alias{as_print_string.SimLongitudinalGSFCov}
\alias{as_print_string.SimLongitudinalRandomSlope}
\alias{as_print_string.SimLongitudinalRandomSlopeCov}
\alias{as_print_string.SimLongitudinalSteinFojo}
\alias{as_print_string.SimLongitudinalSteinFojoCov}
\alias{as_print_string.SimSurvival}
\title{\code{as_print_string}}
\usage{
as_print_string(object, ...)

\method{as_print_string}{SimLongitudinal}(object, ...)

\method{as_print_string}{SimLongitudinalClaretBruno}(object, ...)

\method{as_print_string}{SimLongitudinalClaretBrunoCov}(object, ...)

\method{as_print_string}{SimLongitudinalGSF}(object, ...)

\method{as_print_string}{SimLongitudinalGSFCov}(object, ...)

\method{as_print_string}{SimLongitudinalRandomSlope}(object, ...)

\method{as_print_string}{SimLongitudinalRandomSlopeCov}(object, ...)

\method{as_print_string}{SimLongitudinalSteinFojo}(object, ...)

\method{as_print_string}{SimLongitudinalSteinFojoCov}(object, ...)

\method{as_print_string}{SimSurvival}(object, ...)
}
\arguments{
\item{object}{to be converted to string.}

\item{...}{additional options.}
}
\value{
A character vector suitable for printing.
}
\description{
Returns the character representation of an object which is suitable
for printing to the console
}
\concept{as_print_string}
\keyword{internal}
