% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\name{as_print_string.StanModule}
\alias{as_print_string.StanModule}
\title{\code{StanModule} -> Printable \code{Character}}
\usage{
\method{as_print_string}{StanModule}(object, indent = 1, ...)
}
\arguments{
\item{object}{(\code{StanModule}) A stan program}

\item{indent}{(\code{numeric}) how much white space to prefix the print string with.}
}
\value{
A character vector suitable for printing.
}
\description{
Converts \code{\link{StanModule}} object into a printable string.
}
\seealso{
Other StanModule:
\code{\link{StanModule-class}},
\code{\link[=as.character.StanModule]{as.character.StanModule()}},
\code{\link[=as.list.StanModule]{as.list.StanModule()}}
}
\concept{StanModule}
\keyword{internal}
