% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{as_print_string.SurvivalQuantities}
\alias{as_print_string.SurvivalQuantities}
\title{\code{SurvivalQuantities} -> Printable \code{Character}}
\usage{
\method{as_print_string}{SurvivalQuantities}(object, indent = 1, ...)
}
\arguments{
\item{object}{(\code{\link{SurvivalQuantities}}) \cr survival quantities.}

\item{indent}{(\code{numeric})\cr how much white space to prefix the print string with}

\item{...}{not used.}
}
\description{
Converts \code{\link{SurvivalQuantities}} object into a printable string.
}
\seealso{
Other SurvivalQuantities: 
\code{\link{SurvivalQuantities-class}},
\code{\link{as.data.frame.SurvivalQuantities}()},
\code{\link{autoplot.SurvivalQuantities}()},
\code{\link{summary.SurvivalQuantities}()}
}
\concept{SurvivalQuantities}
\keyword{internal}
