% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\name{as_stan_fragments}
\alias{as_stan_fragments}
\title{Conversion of Character Vector into Stan Code Block List}
\usage{
as_stan_fragments(x)
}
\arguments{
\item{x}{(\code{character})\cr the single Stan code vector.}
}
\value{
A list with the Stan code blocks.
}
\description{
Conversion of Character Vector into Stan Code Block List
}
\details{
Function only works if code is in format

\if{html}{\out{<div class="sourceCode">}}\preformatted{data \{
    <code>
\}
model \{
    <code>
\}
}\if{html}{\out{</div>}}

That is to say we do not support code in inline format i.e.

\if{html}{\out{<div class="sourceCode">}}\preformatted{data \{ <code> \}
model \{ <code> \}
}\if{html}{\out{</div>}}
}
\keyword{internal}
