% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\name{as_stan_list.DataJoint}
\alias{as_stan_list.DataJoint}
\alias{as.list.DataJoint}
\title{\code{DataJoint} -> \code{list}}
\usage{
\method{as_stan_list}{DataJoint}(object, ...)

\method{as.list}{DataJoint}(x, ...)
}
\arguments{
\item{object}{(\code{\link{DataJoint}}) \cr Survival and Longitudinal Data.}

\item{...}{Not Used.}

\item{x}{(\code{\link{DataJoint}}) \cr Survival and Longitudinal Data.}
}
\description{
Coerces  \code{\link{DataJoint}} into a \code{list} of data components required
for fitting a \code{\link{JointModel}}. See the vignette (TODO) for more details.
}
\seealso{
Other DataJoint: 
\code{\link{DataJoint-class}},
\code{\link{subset.DataJoint}()}

Other as_stan_list: 
\code{\link{as_stan_list.DataLongitudinal}()},
\code{\link{as_stan_list.DataSurvival}()}
}
\concept{DataJoint}
\concept{as_stan_list}
