% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLongitudinal.R
\name{as_stan_list.DataLongitudinal}
\alias{as_stan_list.DataLongitudinal}
\title{\code{DataLongitudinal} -> \code{list}}
\usage{
\method{as_stan_list}{DataLongitudinal}(object, subject_var, ...)
}
\arguments{
\item{object}{(\code{\link{DataLongitudinal}}) \cr Longitudinal Data.}

\item{...}{Not Used.}
}
\description{
Coerces  \code{\link{DataLongitudinal}} into a \code{list} of data components required
for fitting a \code{\link{JointModel}}. See the vignette (TODO) for more details.
}
\seealso{
Other DataLongitudinal: 
\code{\link{DataLongitudinal-class}},
\code{\link{as.data.frame.DataLongitudinal}()},
\code{\link{extractVariableNames.DataLongitudinal}()}

Other as_stan_list: 
\code{\link{as_stan_list.DataJoint}()},
\code{\link{as_stan_list.DataSurvival}()}
}
\concept{DataLongitudinal}
\concept{as_stan_list}
