% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSubject.R, R/DataLongitudinal.R,
%   R/DataSurvival.R, R/DataJoint.R
\name{as_stan_list.DataSubject}
\alias{as_stan_list.DataSubject}
\alias{as.list.DataSubject}
\alias{as_stan_list.DataLongitudinal}
\alias{as.list.DataLongitudinal}
\alias{model.matrix.DataSurvival}
\alias{as_stan_list.DataSurvival}
\alias{as.list.DataSurvival}
\alias{as_stan_list.DataObject}
\alias{as_stan_list.DataJoint}
\alias{as.list.DataJoint}
\title{Data Object -> \code{list}}
\usage{
\method{as_stan_list}{DataSubject}(object, ...)

\method{as.list}{DataSubject}(x, ...)

\method{as_stan_list}{DataLongitudinal}(object, subject_var, ...)

\method{as.list}{DataLongitudinal}(x, ...)

\method{model.matrix}{DataSurvival}(
  object,
  df = as.data.frame(object),
  vars = extractVariableNames(object),
  ...
)

\method{as_stan_list}{DataSurvival}(object, ...)

\method{as.list}{DataSurvival}(x, ...)

\method{as_stan_list}{DataJoint}(object, model = NULL, ...)

\method{as.list}{DataJoint}(x, ...)
}
\arguments{
\item{object}{(\code{DataSubject} or \code{DataLongitudinal} or \code{DataSurvival}) data object to convert to a \code{list}.}

\item{...}{not used.}

\item{x}{(\code{DataSubject} or \code{DataLongitudinal} or \code{DataSurvival}) data object to convert to a \code{list}.}

\item{subject_var}{(\code{character}) the name of the variable
containing the subject identifier.}

\item{df}{(\code{data.frame})\cr The data frame to be used for generating the design matrix.}

\item{vars}{(\code{list})\cr The list of variable names to be used for generating the design matrix.}

\item{model}{(\code{JointModel} or \code{NULL}) Optional model specification used to
add model-specific data components.}
}
\value{
A named \code{list} suitable for use as Stan data.
}
\description{
Coerces a data object into a \code{list} of data components required
for fitting a \code{\link{JointModel}}. See the "Extending jmpost" vignette for more details.
}
\seealso{
Other DataSubject:
\code{\link{DataSubject-class}},
\code{\link[=as.data.frame.DataSubject]{as.data.frame.DataSubject()}},
\code{\link[=as_print_string.DataSubject]{as_print_string.DataSubject()}},
\code{\link[=extractVariableNames.DataSubject]{extractVariableNames.DataSubject()}}

Other DataLongitudinal:
\code{\link{DataLongitudinal-class}},
\code{\link[=as.data.frame.DataLongitudinal]{as.data.frame.DataLongitudinal()}},
\code{\link[=as_print_string.DataLongitudinal]{as_print_string.DataLongitudinal()}},
\code{\link[=extractVariableNames.DataLongitudinal]{extractVariableNames.DataLongitudinal()}}

Other DataSurvival:
\code{\link{DataSurvival-class}},
\code{\link[=as.data.frame.DataSurvival]{as.data.frame.DataSurvival()}},
\code{\link[=as_print_string.DataSurvival]{as_print_string.DataSurvival()}},
\code{\link[=extractVariableNames.DataSurvival]{extractVariableNames.DataSurvival()}}

Other as_stan_list:
\code{\link[=as_stan_list]{as_stan_list()}},
\code{\link[=as_stan_list.Parameter]{as_stan_list.Parameter()}},
\code{\link[=as_stan_list.ParameterList]{as_stan_list.ParameterList()}},
\code{\link[=as_stan_list.Prior]{as_stan_list.Prior()}}

Other DataJoint:
\code{\link{DataJoint-class}},
\code{\link[=subset.DataJoint]{subset.DataJoint()}}
}
\concept{DataJoint}
\concept{DataLongitudinal}
\concept{DataSubject}
\concept{DataSurvival}
\concept{as_stan_list}
