% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\name{as_stan_list.DataSurvival}
\alias{as_stan_list.DataSurvival}
\title{\code{DataSurvival} -> \code{list}}
\usage{
\method{as_stan_list}{DataSurvival}(object, ...)
}
\arguments{
\item{object}{(\code{\link{DataSurvival}}) \cr Survival Data.}

\item{...}{Not Used.}
}
\description{
Coerces  \code{\link{DataSurvival}} into a \code{list} of data components required
for fitting a \code{\link{JointModel}}. See the vignette (TODO) for more details.
}
\seealso{
Other DataSurvival: 
\code{\link{DataSurvival-class}},
\code{\link{as.data.frame.DataSurvival}()},
\code{\link{extractVariableNames.DataSurvival}()}

Other as_stan_list: 
\code{\link{as_stan_list.DataJoint}()},
\code{\link{as_stan_list.DataLongitudinal}()}
}
\concept{DataSurvival}
\concept{as_stan_list}
