% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{as_stan_list.Parameter}
\alias{as_stan_list.Parameter}
\title{\code{Parameter} -> \code{list}}
\usage{
\method{as_stan_list}{Parameter}(object, ...)
}
\arguments{
\item{object}{(\code{Parameter}) a prior Distribution}

\item{...}{Not Used.}
}
\value{
A named \code{list} suitable for use as Stan data.
}
\description{
Converts a Parameter object to a list of parameter data values
for a Stan model.
}
\seealso{
Other as_stan_list:
\code{\link[=as_stan_list]{as_stan_list()}},
\code{\link[=as_stan_list.DataSubject]{as_stan_list.DataSubject()}},
\code{\link[=as_stan_list.ParameterList]{as_stan_list.ParameterList()}},
\code{\link[=as_stan_list.Prior]{as_stan_list.Prior()}}

Other Parameter:
\code{\link{Parameter-Getter-Methods}},
\code{\link{Parameter-class}},
\code{\link[=as.StanModule.Parameter]{as.StanModule.Parameter()}},
\code{\link[=as.character.Parameter]{as.character.Parameter()}}
}
\concept{Parameter}
\concept{as_stan_list}
