% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterList.R
\name{as_stan_list.ParameterList}
\alias{as_stan_list.ParameterList}
\title{\code{ParameterList} -> \code{list}}
\usage{
\method{as_stan_list}{ParameterList}(object, ...)
}
\arguments{
\item{object}{(\code{ParameterList}) A List of \code{\link{Parameter}} Objects.}

\item{...}{Not Used.}
}
\value{
A named \code{list} suitable for use as Stan data.
}
\description{
Converts a ParameterList object to a list of parameter data values
for a Stan model.
}
\seealso{
Other as_stan_list:
\code{\link[=as_stan_list]{as_stan_list()}},
\code{\link[=as_stan_list.DataSubject]{as_stan_list.DataSubject()}},
\code{\link[=as_stan_list.Parameter]{as_stan_list.Parameter()}},
\code{\link[=as_stan_list.Prior]{as_stan_list.Prior()}}

Other ParameterList:
\code{\link{ParameterList-Getter-Methods}},
\code{\link{ParameterList-class}},
\code{\link[=as.StanModule.ParameterList]{as.StanModule.ParameterList()}},
\code{\link[=as.list.ParameterList]{as.list.ParameterList()}},
\code{\link[=as_print_string.ParameterList]{as_print_string.ParameterList()}}
}
\concept{ParameterList}
\concept{as_stan_list}
