% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R, R/QuantityGeneratorPopulation.R,
%   R/QuantityGeneratorPrediction.R, R/QuantityGeneratorSubject.R
\name{as_stan_list.QuantityGenerator}
\alias{as_stan_list.QuantityGenerator}
\alias{as_stan_list.QuantityGeneratorPopulation}
\alias{as_stan_list.QuantityGeneratorPrediction}
\alias{as_stan_list.QuantityGeneratorSubject}
\title{\code{QuantityGenerator} -> \code{list}}
\usage{
\method{as_stan_list}{QuantityGenerator}(object, data, ...)

\method{as_stan_list}{QuantityGeneratorPopulation}(object, data, model = NULL, ...)

\method{as_stan_list}{QuantityGeneratorPrediction}(object, data, model, ...)

\method{as_stan_list}{QuantityGeneratorSubject}(object, data, ...)
}
\arguments{
\item{object}{(\code{QuantityGenerator}) object to convert to a list.}

\item{data}{(\code{DataJoint}) Survival and Longitudinal Data.}

\item{...}{Not currently used.}

\item{model}{(\code{JointModel} or \code{NULL}) optional model used to add model-specific generated-quantities data.}
}
\value{
Will always throw an error.
}
\description{
Just a placeholder which will error when used.
}
\keyword{internal}
