% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_stan_list}
\alias{as_stan_list}
\alias{as_stan_list.default}
\title{\code{as_stan_list}}
\usage{
as_stan_list(object, ...)

\method{as_stan_list}{default}(object, ...)
}
\arguments{
\item{object}{to be converted.}

\item{...}{additional options.}
}
\value{
A named \code{list} suitable for use as Stan data.
}
\description{
Extracts a list of data elements from an object to be used as input
to a Stan Model
}
\examples{
as_stan_list(Parameter(prior_normal(0, 1), "beta"))
}
\seealso{
Other as_stan_list:
\code{\link[=as_stan_list.DataSubject]{as_stan_list.DataSubject()}},
\code{\link[=as_stan_list.Parameter]{as_stan_list.Parameter()}},
\code{\link[=as_stan_list.ParameterList]{as_stan_list.ParameterList()}},
\code{\link[=as_stan_list.Prior]{as_stan_list.Prior()}}
}
\concept{as_stan_list}
