% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{as_vcov}
\alias{as_vcov}
\title{Construct a Variance-Covariance Matrix from Standard Deviations and Correlations}
\usage{
as_vcov(sd, cor)
}
\arguments{
\item{sd}{A numeric vector containing the standard deviations of the variables.}

\item{cor}{A numeric vector containing the pairwise correlations between the variables.
The vector should be in the order of the upper triangular part of the matrix, excluding
the diagonal elements. e.g. for a 3x3 matrix c(0.2, 0.3, 0.4) would be the values for
x_12, x_13 and x_23 of the correlation matrix respectively}
}
\value{
A symmetric matrix representing the variance-covariance matrix calculated from
the input standard deviations and correlations. The matrix is guaranteed to be positive
semi-definite.
}
\description{
This function creates a variance-covariance matrix based on the input standard deviations
and correlations. Note that the values may be altered using \code{\link[Matrix:nearPD]{Matrix::nearPD()}}
in order to ensure that it is a valid positive semi-definite matrix.
}
\examples{
sd <- c(1, 2, 3)
cor <- c(0.5, 0.6, 0.7)
as_vcov(sd, cor)

}
