% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{autoplot,SurvivalQuantities-method}
\alias{autoplot,SurvivalQuantities-method}
\title{Automatic Plotting for SurvivalSamples}
\usage{
\S4method{autoplot}{SurvivalQuantities}(object, add_km = FALSE, add_ci = TRUE, add_wrap = TRUE, ...)
}
\arguments{
\item{object}{(\code{\link{SurvivalQuantities}}) \cr Survival Quantities}

\item{add_km}{(\code{logical}) \cr If \code{TRUE} Kaplan-Meier curves will be added to the plot for
each group/patient as defined by \code{patients}}

\item{add_ci}{(\code{logical}) \cr If \code{TRUE} 95\% CI will be added to the plot for
each group/patient as defined by \code{patients}}

\item{add_wrap}{(\code{logical}) \cr If \code{TRUE} will apply a \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} to the plot
by each group/patient as defined by \code{patients}}

\item{...}{other arguments passed to plotting methods.}
}
\description{
Automatic Plotting for SurvivalSamples
}
\seealso{
Other autoplot: 
\code{\link{autoplot}()}

Other SurvivalQuantities: 
\code{\link{SurvivalQuantities-class}},
\code{\link{as.data.frame,SurvivalQuantities-method}},
\code{\link{summary,SurvivalQuantities-method}}
}
\concept{SurvivalQuantities}
\concept{autoplot}
