% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalSamples.R
\name{autoplot,SurvivalSamples-method}
\alias{autoplot,SurvivalSamples-method}
\title{Automatic Plotting for SurvivalSamples}
\usage{
\S4method{autoplot}{SurvivalSamples}(
  object,
  patients,
  time_grid = NULL,
  type = c("surv", "haz", "loghaz", "cumhaz"),
  add_km = FALSE,
  add_ci = TRUE,
  add_wrap = TRUE,
  ...
)
}
\arguments{
\item{patients}{(\code{character} or \code{list} or \code{NULL})\cr which patients to calculate the desired
quantities for.
See "Patient Specification" for more details.}

\item{time_grid}{(\code{numeric})\cr a vector of time points to calculate the desired quantity at.}

\item{type}{(\code{character})\cr The quantity to be generated.
Must be one of \code{surv}, \code{haz}, \code{loghaz}, \code{cumhaz}.}

\item{add_km}{(\code{logical}) \cr If \code{TRUE} kaplan-Meier curves will be added to the plot for
each group/patient as defined by \code{patients}}

\item{add_ci}{(\code{logical}) \cr If \code{TRUE} 95\% CI will be added to the plot for
each group/patient as defined by \code{patients}}

\item{add_wrap}{(\code{logical}) \cr If \code{TRUE} will apply a \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} to the plot
by each group/patient as defined by \code{patients}}
}
\description{
Automatic Plotting for SurvivalSamples
}
\section{Patient Specification}{

If \code{patients} is a character vector then quantities / summary statistics
will only be calculated for those specific patients

If \code{patients} is a list then any elements with more than 1 patient ID will be grouped together
and their quantities / summary statistics (as selected by \code{type})
will be calculated by taking the point-wise average. For example:
\code{patients = list("g1" = c("pt1", "pt2"), "g2" = c("pt3", "pt4"))} would result
in 2 groups being created whose values are the pointwise average
of \code{c("pt1", "pt2")} and \code{c("pt3", "pt4")} respectively.

If \code{patients=NULL} then all patients from original dataset will be selected
}

\seealso{
Other autoplot: 
\code{\link{autoplot}()}

Other SurvivalSamples: 
\code{\link{SurvivalSamples-class}},
\code{\link{SurvivalSamples}},
\code{\link{predict,SurvivalSamples-method}}
}
\concept{SurvivalSamples}
\concept{autoplot}
