% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalSamples.R
\name{autoplot.LongitudinalSamples}
\alias{autoplot.LongitudinalSamples}
\title{Plot Longitudinal Samples}
\usage{
\method{autoplot}{LongitudinalSamples}(object, ...)
}
\arguments{
\item{object}{(\code{\link{LongitudinalSamples}}) \cr Longitudinal Samples}

\item{...}{Not used.}
}
\description{
Plot Longitudinal Samples
}
