% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-exports.R
\name{autoplot}
\alias{autoplot}
\title{Create a complete \code{ggplot} appropriate to a particular data type}
\description{
See \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} for details.
}
\examples{
\dontrun{
autoplot(survival_quantities)
}
}
\seealso{
Other autoplot:
\code{\link[=autoplot.LongitudinalQuantities]{autoplot.LongitudinalQuantities()}},
\code{\link[=autoplot.SurvivalQuantities]{autoplot.SurvivalQuantities()}}
}
\concept{autoplot}
