% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{autoplot.SurvivalQuantities}
\alias{autoplot.SurvivalQuantities}
\title{Automatic Plotting for `SurvivalQuantities``}
\usage{
\method{autoplot}{SurvivalQuantities}(object, conf.level = 0.95, add_km = FALSE, add_wrap = TRUE, ...)
}
\arguments{
\item{object}{(\code{SurvivalQuantities}) survival quantities.}

\item{conf.level}{(\code{numeric}) confidence level of the interval. If values of \code{FALSE},
\code{NULL} or \code{0} are provided then confidence regions will not be added to the plot}

\item{add_km}{(\code{logical}) if \code{TRUE} Kaplan-Meier curves will be added to the plot for
each group/subject.}

\item{add_wrap}{(\code{logical}) if \code{TRUE} will apply a \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} to the plot
by each group/subject.}

\item{...}{not used.}
}
\value{
A \code{ggplot} object.
}
\description{
Automatic Plotting for `SurvivalQuantities``
}
\seealso{
Other SurvivalQuantities:
\code{\link{SurvivalQuantities-class}},
\code{\link[=as.data.frame.SurvivalQuantities]{as.data.frame.SurvivalQuantities()}},
\code{\link[=brierScore.SurvivalQuantities]{brierScore.SurvivalQuantities()}},
\code{\link[=summary.SurvivalQuantities]{summary.SurvivalQuantities()}}

Other autoplot:
\code{\link{autoplot}},
\code{\link[=autoplot.LongitudinalQuantities]{autoplot.LongitudinalQuantities()}}
}
\concept{SurvivalQuantities}
\concept{autoplot}
