% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{brierScore}
\alias{brierScore}
\title{\code{brierScore}}
\usage{
brierScore(object, ...)
}
\arguments{
\item{object}{to calculate Brier Score for.}

\item{...}{additional options.}
}
\value{
A numeric Brier score or a \code{data.frame} of scores over time.
}
\description{
Returns the Brier Score for a given model
}
\examples{
\dontrun{
brierScore(survival_quantities)
}
}
\seealso{
Other brierScore:
\code{\link[=brierScore.SurvivalQuantities]{brierScore.SurvivalQuantities()}}
}
\concept{brierScore}
