% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalClaretBruno.R
\name{clbr_sld}
\alias{clbr_sld}
\alias{clbr_ttg}
\alias{clbr_dsld}
\title{Claret-Bruno Functionals}
\usage{
clbr_sld(t, b, g, c, p)

clbr_ttg(t, b, g, c, p)

clbr_dsld(t, b, g, c, p)
}
\arguments{
\item{t}{(\code{numeric}) time grid.}

\item{b}{(\code{number}) baseline sld.}

\item{g}{(\code{number}) growth rate.}

\item{c}{(\code{number}) resistance rate.}

\item{p}{(\code{number}) growth inhibition.}
}
\value{
The function results.
}
\description{
Claret-Bruno Functionals
}
\keyword{internal}
