% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\name{collapse_quantities}
\alias{collapse_quantities}
\title{Create Grouped Quantities}
\usage{
collapse_quantities(quantities_raw, collapser)
}
\arguments{
\item{quantities_raw}{(\code{matrix}) of samples with 1 row per sample and 1 column per
distinct quantity.}

\item{collapser}{(\code{QuantityCollapser}) specifies which columns to combine together.}
}
\value{
An array of collapsed posterior quantities.
}
\description{
This function takes a matrix of quantity samples and aggregates them by calculating the
pointwise average.
}
\details{
This function essentially implements the group wise average by collapsing multiple columns
together based on the specification provided by the \code{QuantityCollapser} object.
The \link{Grid-Dev} page provides an example of what this function implements
}
\keyword{internal}
