% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{covariates}
\alias{covariates}
\title{Extract Covariate Names}
\usage{
covariates(object, ...)
}
\arguments{
\item{object}{(\code{ANY}) the object to extract covariate names from.}

\item{...}{additional arguments added by methods.}
}
\value{
A character vector containing the covariate names.
}
\description{
Extract Covariate Names
}
\examples{
surv_data <- DataSurvival(os_data, Surv(os_time, os_event) ~ age + sex)
covariates(surv_data)
}
\seealso{
Other covariates:
\code{\link[=covariates.DataSurvival]{covariates.DataSurvival()}}
}
\concept{covariates}
