% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimJointData.R
\name{cut_data}
\alias{cut_data}
\title{Cut Study Data}
\usage{
cut_data(object, cut_time)
}
\arguments{
\item{object}{(\code{SimJointData}) where the cut times should be applied.}

\item{cut_time}{(\code{numeric}) A vector of cut off times, either length 1 for all patients or
\code{nrow(object@survival)} for a time per patient.}
}
\value{
The input \code{SimJointData} object restricted to the requested follow-up time.
}
\description{
Cut Study Data
}
\details{
All observations after this time are remove. Survival is censored at this time and any longitudinal
values are removed.
}
\examples{
data <- SimJointData(
  survival = SimSurvivalExponential(lambda = 1/10),
  longitudinal = SimLongitudinalSteinFojo()
  )
data <- cut_data(data, 5)
data@survival
# Now max time is 5
max(data@survival$time)

}
