% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\docType{class}
\name{DataJoint-class}
\alias{DataJoint-class}
\alias{.DataJoint}
\alias{DataJoint}
\title{Joint Data Object and Constructor Function}
\usage{
DataJoint(subject, survival = NULL, longitudinal = NULL)
}
\arguments{
\item{subject}{(\code{DataSubject}) object created by \code{\link[=DataSubject]{DataSubject()}}.}

\item{survival}{(\code{DataSurvival}) object created by \code{\link[=DataSurvival]{DataSurvival()}}.}

\item{longitudinal}{(\code{DataLongitudinal}) object created by \code{\link[=DataLongitudinal]{DataLongitudinal()}}.}
}
\value{
A \code{DataJoint} object.
}
\description{
The \code{DataJoint} class handles combining data from a \code{\link{DataSurvival}} object and a
\code{\link{DataLongitudinal}} object.
}
\section{Slots}{

\describe{
\item{\code{subject}}{(\code{DataSubject})\cr See Argument for details.}

\item{\code{survival}}{(\code{DataSurvival})\cr See Argument for details.}

\item{\code{longitudinal}}{(\code{DataLongitudinal})\cr See Argument for details.}
}}

\examples{
subjects <- data.frame(id = c("1", "2"), arm = "A", study = "S")
survival <- data.frame(id = c("1", "2"), time = c(5, 8), event = c(1, 0))
longitudinal <- data.frame(
  id = rep(c("1", "2"), each = 2),
  time = rep(c(0, 1), 2),
  response = c(10, 9, 12, 11)
)
DataJoint(
  DataSubject(subjects, "id", "arm", "study"),
  DataSurvival(survival, Surv(time, event) ~ 1),
  DataLongitudinal(longitudinal, response ~ time)
)
}
\seealso{
Other DataObjects:
\code{\link{DataLongitudinal-class}},
\code{\link{DataSubject-class}},
\code{\link{DataSurvival-class}}

Other DataJoint:
\code{\link[=as_stan_list.DataSubject]{as_stan_list.DataSubject()}},
\code{\link[=subset.DataJoint]{subset.DataJoint()}}
}
\concept{DataJoint}
\concept{DataObjects}
