% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\docType{class}
\name{DataJoint-class}
\alias{DataJoint-class}
\alias{.DataJoint}
\alias{DataJoint}
\alias{as.list,DataJoint-method}
\title{\code{DataJoint}}
\usage{
DataJoint(survival, longitudinal)
}
\arguments{
\item{survival}{A \linkS4class{DataSurvival} object created by \code{\link[=DataSurvival]{DataSurvival()}}}

\item{longitudinal}{A \linkS4class{DataLongitudinal} object created by \code{\link[=DataLongitudinal]{DataLongitudinal()}}}
}
\description{
The \code{\link{DataJoint}} class handles combining data from a \linkS4class{DataSurvival} object and a
\linkS4class{DataLongitudinal} object
}
\details{
\subsection{Coercion}{
\itemize{
\item \code{as.list(x)}, \code{as(x, "list")}: Coerces x into a list of data components required
for fitting a \code{\link[=JointModel]{JointModel()}}.
See the vignette (TODO) for more details
}
}
}
