% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLongitudinal.R
\docType{class}
\name{DataLongitudinal-class}
\alias{DataLongitudinal-class}
\alias{.DataLongitudinal}
\alias{DataLongitudinal}
\title{Longitudinal Data Object and Constructor Function}
\usage{
DataLongitudinal(data, formula, threshold = NULL)
}
\arguments{
\item{data}{(\code{data.frame}) containing the observed longitudinal data.}

\item{formula}{(\code{formula}) of the form \code{outcome ~ time}, and cannot contain any additional covariates.}

\item{threshold}{(\code{numeric}) cut-off value to be used to declare an observation as censored
(below detection limit).}
}
\value{
A \code{DataLongitudinal} object.
}
\description{
The \code{\link{DataLongitudinal}} class handles the processing of the longitudinal data for fitting a Joint Model.
}
\section{Slots}{

\describe{
\item{\code{data}}{(\code{data.frame})\cr See Arguments for details; Note that
observations that contain missing values in the required variables are removed.}

\item{\code{formula}}{(\code{formula})\cr See Arguments for details}

\item{\code{threshold}}{(\code{numeric})\cr See Arguments for details}
}}

\examples{
DataLongitudinal(tumor_data, sld ~ year, threshold = 5)
}
\seealso{
Other DataObjects:
\code{\link{DataJoint-class}},
\code{\link{DataSubject-class}},
\code{\link{DataSurvival-class}}

Other DataLongitudinal:
\code{\link[=as.data.frame.DataLongitudinal]{as.data.frame.DataLongitudinal()}},
\code{\link[=as_print_string.DataLongitudinal]{as_print_string.DataLongitudinal()}},
\code{\link[=as_stan_list.DataSubject]{as_stan_list.DataSubject()}},
\code{\link[=extractVariableNames.DataLongitudinal]{extractVariableNames.DataLongitudinal()}}
}
\concept{DataLongitudinal}
\concept{DataObjects}
