% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLongitudinal.R
\name{DataLongitudinal-Shared}
\alias{DataLongitudinal-Shared}
\title{Re-used documentation for \code{DataLongitudinal}}
\arguments{
\item{object}{(\code{\link{DataLongitudinal}}) \cr Longitudinal Data.}

\item{x}{(\code{\link{DataLongitudinal}}) \cr Longitudinal Data.}

\item{...}{Not Used.}
}
\description{
Re-used documentation for \code{DataLongitudinal}
}
\keyword{internal}
