% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLongitudinal.R
\docType{class}
\name{DataLongitudinal-class}
\alias{DataLongitudinal-class}
\alias{.DataLongitudinal}
\alias{DataLongitudinal}
\title{\code{DataLongitudinal}}
\usage{
DataLongitudinal(data, formula, subject, threshold = NULL)
}
\arguments{
\item{data}{A \code{data.frame} object containing the observed longitudinal data}

\item{formula}{A two sided \code{formula} of the form \code{outcome ~ time}. Cannot contain any additional covariates}

\item{subject}{A length 1 \code{character} vector specifying the name of the subject identifier variable}

\item{threshold}{A length 1 \code{numeric} that specifies what cut-off value should be used to declare
an observation as censored / below detection limit}
}
\description{
The \code{\link{DataLongitudinal}} class handles the processing of the longitudinal data for fitting a Joint Model.
}
\details{
\subsection{Coercion}{
\itemize{
\item \code{as.list(x)}, \code{as(x, "list")}: Coerces x into a list of data components required
for fitting a \code{\link{JointModel}}.
See the vignette (TODO) for more details
}
}
}
\seealso{
\code{\link{DataJoint}}, \code{\link{DataSurvival}}
}
