% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSubject.R
\name{DataSubject-Shared}
\alias{DataSubject-Shared}
\title{Re-used documentation for \code{DataSubject}}
\arguments{
\item{object}{(\code{DataSubject}) subject-level data.}

\item{x}{(\code{DataSubject}) subject-level data.}

\item{...}{Not Used.}
}
\description{
Re-used documentation for \code{DataSubject}
}
\keyword{internal}
