% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\name{DataSurvival-Shared}
\alias{DataSurvival-Shared}
\title{Re-used documentation for \code{DataSurvival}}
\arguments{
\item{object}{(\code{DataSurvival}) Survival Data.}

\item{x}{(\code{DataSurvival}) Survival Data.}

\item{...}{Not Used.}
}
\description{
Re-used documentation for \code{DataSurvival}
}
\keyword{internal}
