% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\docType{class}
\name{DataSurvival-class}
\alias{DataSurvival-class}
\alias{.DataSurvival}
\alias{DataSurvival}
\title{\code{DataSurvival}}
\usage{
DataSurvival(data, formula, subject, arm, study)
}
\arguments{
\item{data}{A \code{data.frame} object containing the observed survival/time to event data}

\item{formula}{A two sided \code{formula} of the form \code{Surv(time, event) ~ cov1 + cov2 + ...}. See \code{\link[survival:Surv]{survival::Surv()}} for more details though note that this package only supports right censoring}

\item{subject}{A length 1 \code{character} vector specifying the name of the subject identifier variable}

\item{arm}{A length 1 \code{character} vector specifying the name of the treatment arm variable}

\item{study}{A length 1 \code{character} vector specifying the name of the study variable}
}
\description{
The \code{\link{DataSurvival}} class handles the processing of the survival data for fitting a Joint Model.
}
\details{
\subsection{Coercion}{
\itemize{
\item \code{as.list(x)}, \code{as(x, "list")}: Coerces x into a list of data components required
for fitting a \code{\link{JointModel}}.
See the vignette (TODO) for more details
}
}
}
\seealso{
\code{\link{DataJoint}}, \code{\link{DataLongitudinal}}
}
