% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{decompose_subjects}
\alias{decompose_subjects}
\title{Decompose subjects into Relevant Components}
\usage{
decompose_subjects(subjects, all_subjects)
}
\arguments{
\item{subjects}{(\code{character} or \code{list})\cr subject identifiers. If \code{NULL} will be set to \code{all_subjects}.}

\item{all_subjects}{(\code{character})\cr the set of allowable subject identifiers.
Will cause an error if any value of \code{subjects} is not in this vector.}
}
\value{
A list containing three components:
\itemize{
\item \code{groups}: (\code{list})\cr each element of the list is a character vector
specifying which subjects belong to a given "group" where the "group" is the element name
\item \code{unique_values}: (\code{character})\cr vector of the unique subjects within \code{subjects}
\item \code{indexes}: (\code{list})\cr each element is a named and is a numeric index vector
that maps the values of \code{grouped} to \code{unique_values}
}
}
\description{
This function takes in a character vector or list of subjects and decomposes it into a
structured format.
}
\details{
The primary use of this function is to correctly setup indexing variables for
predicting survival quantities (see \code{\link[=SurvivalQuantities]{SurvivalQuantities()}})
}
\examples{
\dontrun{
result <- decompose_subjects(c("A", "B"), c("A", "B", "C", "D"))
result <- decompose_subjects(
    list("g1" = c("A", "B"), "g2" = c("B", "C")),
    c("A", "B", "C", "D")
)
}
}
\seealso{
\code{\link[=expand_subjects]{expand_subjects()}}, \code{\link[=SurvivalQuantities]{SurvivalQuantities()}}
}
\keyword{internal}
