% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{decorated_render}
\alias{decorated_render}
\title{\code{decorated_render}}
\usage{
decorated_render(...)
}
\arguments{
\item{...}{Arguments passed onto \code{\link[jinjar:render]{jinjar::render()}}}
}
\value{
See \code{\link[jinjar:render]{jinjar::render()}}
}
\description{
Simple wrapper around \code{\link[jinjar:render]{jinjar::render()}} that provides some additional default
variables about the system (avoids each call to jinjar having to specify them)
}
\keyword{internal}
