% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\name{dim.Quantities}
\alias{dim.Quantities}
\title{Dimensions of \code{Quantitites}}
\usage{
\method{dim}{Quantities}(x)
}
\arguments{
\item{x}{(\code{\link{Quantities}}) \cr generated quantities.}
}
\description{
TODO
}
\keyword{internal}
