% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalSteinFojoCov.R
\name{.stein_fojo_cov_population_stan_data}
\alias{.stein_fojo_cov_population_stan_data}
\title{Create covariate Stein-Fojo population-prediction Stan data}
\usage{
.stein_fojo_cov_population_stan_data(model, subject, newdata)
}
\value{
A named list of population generated-quantities design matrices.
}
\description{
Create covariate Stein-Fojo population-prediction Stan data
}
\keyword{internal}
