% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalSteinFojoCov.R
\name{.stein_fojo_cov_population_variables}
\alias{.stein_fojo_cov_population_variables}
\title{Variables required for covariate Stein-Fojo population predictions}
\usage{
.stein_fojo_cov_population_variables(model)
}
\value{
A character vector.
}
\description{
Variables required for covariate Stein-Fojo population predictions
}
\keyword{internal}
