% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalSteinFojoCov.R
\name{.stein_fojo_cov_stan_data}
\alias{.stein_fojo_cov_stan_data}
\title{Create covariate Stein-Fojo Stan data}
\usage{
.stein_fojo_cov_stan_data(model, subject)
}
\value{
A named list of model-specific Stan data components.
}
\description{
Create covariate Stein-Fojo Stan data
}
\keyword{internal}
