% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtsld.R
\name{dtsld}
\alias{dtsld}
\title{Derivative of SLD}
\usage{
dtsld(time, psi_bsld, psi_ks, psi_kg, psi_phi)
}
\arguments{
\item{time}{Times to be predicted}

\item{psi_bsld}{Baseline parameter}

\item{psi_ks}{The srinkage rate parameter}

\item{psi_kg}{The growth rate parameter}

\item{psi_phi}{The percentage of cells responded to the treatment}
}
\description{
Derivative of SLD
}
