% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{enableGQ}
\alias{enableGQ}
\title{Enable Generated Quantities Generic}
\usage{
enableGQ(object, ...)
}
\arguments{
\item{object}{(\code{StanModel}) to enable generated quantities for.}

\item{...}{Not used.

Optional hook method that is called on a \code{\link{StanModel}} if attempting to use
either \code{\link{LongitudinalQuantities}} or \code{\link{SurvivalQuantities}}}
}
\value{
\code{\link{StanModule}} object
}
\description{
Enable Generated Quantities Generic
}
\examples{
enableGQ(LongitudinalGSF())
}
