% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{enableLink}
\alias{enableLink}
\title{Enable Link Generic}
\usage{
enableLink(object, ...)
}
\arguments{
\item{object}{(\code{LongitudinalModel}) to enable link for.}

\item{...}{Not used.

Optional hook method that is called on a \code{\link{LongitudinalModel}} only if a link method
is provided to \code{\link{JointModel}}. This can be used to allow the model to include any
optional stan code that is only required if there are links present.}
}
\value{
\code{\link{LongitudinalModel}} object
}
\description{
Enable Link Generic
}
\examples{
enableLink(LongitudinalGSF())
}
