% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_initial_values}
\alias{expand_initial_values}
\title{Replicate single values in a list based on specified sizes}
\usage{
expand_initial_values(initial_values, sizes)
}
\arguments{
\item{initial_values}{A named list of initial values.}

\item{sizes}{A named list of sizes, where each size corresponds to an element in initial_values.}
}
\value{
A named list of values, with any single values in the initial_values list
replicated according to the corresponding values in the sizes list.
}
\description{
This function takes a list of initial values and a corresponding list of sizes
and replicates any single values in the initial values list according to the
corresponding values in the sizes list. The resulting list has the same names
as the original lists.
}
\keyword{internal}
