% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_patients}
\alias{expand_patients}
\title{\code{expand_patients}}
\usage{
expand_patients(patients, all_pts)
}
\arguments{
\item{patients}{(\code{character} or \code{NULL})\cr Character vector representing the patients.
If NULL, it will be set to the value of \code{all_pts}.}

\item{all_pts}{(\code{character})\cr Character vector representing all possible patients.}
}
\value{
Returns the expanded \code{patients} vector.
}
\description{
This function checks and expands a given patients vector.
The input vector must be unique and contain only values
as specified by \code{all_pts}
}
\examples{
expand_patients(
    patients = c("patient1", "patient2"),
    all_pts = c("patient1", "patient2", "patient3")
)
expand_patients(
    patients = NULL,
    all_pts = c("patient1", "patient2", "patient3")
)
}
\keyword{internal}
