% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_subjects}
\alias{expand_subjects}
\title{\code{expand_subjects}}
\usage{
expand_subjects(subjects, all_subjects)
}
\arguments{
\item{subjects}{(\code{character} or \code{NULL})\cr Character vector representing the subjects.
If NULL, it will be set to the value of \code{all_subjects}.}

\item{all_subjects}{(\code{character})\cr Character vector representing all possible subjects.}
}
\value{
Returns the expanded \code{subjects} vector.
}
\description{
This function checks and expands a given subjects vector.
The input vector must be unique and contain only values
as specified by \code{all_subjects}
}
\keyword{internal}
