% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_time_grid}
\alias{expand_time_grid}
\title{\code{expand_time_grid}}
\usage{
expand_time_grid(time_grid, max_time)
}
\arguments{
\item{time_grid}{(\code{numeric} or \code{NULL})\cr A vector of times which quantities will be
evaluated at. If NULL, a default grid will be created as a length 201 vector spanning
from 0 to \code{max_time}.}

\item{max_time}{(`numeric``)\cr Specifies the maximum time to be used in creating the default grid.}
}
\value{
Returns the expanded time_grid.
}
\description{
This function expands a given time grid by setting a default grid if one hasn't been provided
and then verifying it's properties.
The grid must be finite, sorted, and contain unique values.
}
\keyword{internal}
