% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_long_model.R
\name{exponential_long_model}
\alias{exponential_long_model}
\title{Exponential longitudinal model creator for joint modeling analysis}
\usage{
exponential_long_model(
  functions = long_fun,
  data = long_data,
  parameters = long_parameters,
  transformed_parameters = long_tr_parameters,
  prior = long_prior(),
  generated_quantities = long_gen_quantities,
  inits = long_inits
)
}
\arguments{
\item{functions}{Stan object with the functions of the longitudinal model}

\item{data}{Stan object with the data of the longitudinal model}

\item{parameters}{Stan object with the parameters of the longitudinal model}

\item{transformed_parameters}{Stan object with the transformed parameters of the longitudinal model}

\item{prior}{list with the priors of the longitudinal model}

\item{generated_quantities}{Stan object with the generated quantities of the longitudinal model}

\item{inits}{list}
}
\description{
Exponential longitudinal model creator for joint modeling analysis
}
