% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/stan_module.R, R/hazard_link.R,
%   R/longmodel.R, R/OsModel.R
\name{inits<-}
\alias{inits<-}
\alias{inits<-,StanModule-method}
\alias{inits<-,HazardLink-method}
\alias{inits<-,LongModel-method}
\alias{inits<-,OsModel-method}
\title{inits() replacement method for \code{LongModel}, \code{HazardLink} , \code{StanModule} or \code{OsModel} object}
\usage{
inits(object) <- value

\S4method{inits}{StanModule}(object) <- value

\S4method{inits}{HazardLink}(object) <- value

\S4method{inits}{LongModel}(object) <- value

\S4method{inits}{OsModel}(object) <- value
}
\arguments{
\item{object}{A \code{LongModel}, \code{HazardLink} , \code{StanModule} and \code{OsModel} object}

\item{value}{the character strings of the inits information for replacement}
}
\description{
inits() replacement method for \code{LongModel}, \code{HazardLink} , \code{StanModule} or \code{OsModel} object
}
\examples{
inits(object)<-"mean_mu_ks = 60"
}
