% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/stan_module.R, R/hazard_link.R,
%   R/longmodel.R, R/OsModel.R
\name{priors<-}
\alias{priors<-}
\alias{priors<-,StanModule-method}
\alias{priors<-,HazardLink-method}
\alias{priors<-,LongModel-method}
\alias{priors<-,OsModel-method}
\title{priors() replacement method for \code{LongModel}, \code{HazardLink} , \code{StanModule} or \code{OsModel} object}
\usage{
priors(object) <- value

\S4method{priors}{StanModule}(object) <- value

\S4method{priors}{HazardLink}(object) <- value

\S4method{priors}{LongModel}(object) <- value

\S4method{priors}{OsModel}(object) <- value
}
\arguments{
\item{object}{A \code{LongModel}, \code{HazardLink} , \code{StanModule} and \code{OsModel} object}

\item{value}{the character strings of the prior information for replacement}
}
\description{
priors() replacement method for \code{LongModel}, \code{HazardLink} , \code{StanModule} or \code{OsModel} object
}
\examples{
priors(object)<-"lognormal(0,1);"
}
