% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\name{extract_observed_values}
\alias{extract_observed_values}
\title{Extract Observed Longitudinal Values}
\usage{
extract_observed_values(object)
}
\arguments{
\item{object}{(\code{\link{DataJoint}})\cr The data used to fit a \code{\link{JointModel}}}
}
\value{
A data.frame with the following columns
\itemize{
\item \code{subject} (\code{character})\cr The subject identifier
\item \code{time} (\code{numeric})\cr The time at which the observation occured
\item \code{Yob} (\code{numeric})\cr The observed value
}
}
\description{
Utility function to extract the observed longitudinal values from a \code{\link{DataJoint}} object
}
\keyword{internal}
