% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\name{extract_quantities}
\alias{extract_quantities}
\title{Extract Survival Quantities}
\usage{
extract_quantities(
  gq,
  type = c("surv", "haz", "loghaz", "cumhaz", "lm_identity")
)
}
\arguments{
\item{gq}{(\code{CmdStanGQ}) a \link[cmdstanr:CmdStanGQ]{cmdstanr::CmdStanGQ} object created by \code{\link[=generateQuantities]{generateQuantities()}}.}

\item{type}{(\code{character}) quantity to be generated.
Must be one of \code{surv}, \code{haz}, \code{loghaz}, \code{cumhaz}, \code{lm_identity}.}
}
\value{
A \link[posterior:draws_matrix]{posterior::draws_matrix} containing the requested generated quantities.
}
\description{
Utility function to extract generated quantities from a \link[cmdstanr:CmdStanGQ]{cmdstanr::CmdStanGQ} object.
Multiple quantities are generated by default so this is a convenience function to extract
the desired ones and return them them as a user friendly \link[posterior:draws_matrix]{posterior::draws_matrix} object
}
\keyword{internal}
