% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModelSamples.R
\name{extract_survival_quantities}
\alias{extract_survival_quantities}
\title{Extract Survival Quantities}
\usage{
extract_survival_quantities(gq, type = c("surv", "haz", "loghaz", "cumhaz"))
}
\arguments{
\item{gq}{(\code{CmdStanGQ}) \cr A \link[cmdstanr:CmdStanGQ]{cmdstanr::CmdStanGQ} object created by \link{generateQuantities}}

\item{type}{(\code{character})\cr The quantity to be generated.
Must be one of \code{surv}, \code{haz}, \code{loghaz}, \code{cumhaz}.}
}
\description{
Utility function to extract generated quantities from a \link[cmdstanr:CmdStanGQ]{cmdstanr::CmdStanGQ} object.
Multiple quantities are generated by default so this is a convenience function to extract
the desired ones and return them them as a user friendly \link[posterior:draws_matrix]{posterior::draws_matrix} object
}
\keyword{internal}
