% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalSamples.R
\name{extractQuantities,SurvivalSamples-method}
\alias{extractQuantities,SurvivalSamples-method}
\title{extractQuantities}
\usage{
\S4method{extractQuantities}{SurvivalSamples}(
  object,
  patients = NULL,
  time_grid = NULL,
  type = c("surv", "haz", "loghaz", "cumhaz")
)
}
\arguments{
\item{object}{(\code{SurvivalSamples}) \cr A \link[=SurvivalSamples-class]{SurvivalSamples}
object created by \code{\link[=SurvivalSamples]{SurvivalSamples()}}}

\item{patients}{(\code{character} or \code{list} or \code{NULL})\cr which patients to calculate the desired
quantities for.
See "Patient Specification" for more details.}

\item{time_grid}{(\code{numeric})\cr a vector of time points to calculate the desired quantity at.}

\item{type}{(\code{character})\cr The quantity to be generated.
Must be one of \code{surv}, \code{haz}, \code{loghaz}, \code{cumhaz}.}
}
\description{
This method returns a \code{list} of samples of key quantities (survival / log-hazard / etc)
for selected patients at a given set of time points.
}
\section{Patient Specification}{

If \code{patients} is a character vector then quantities / summary statistics
will only be calculated for those specific patients

If \code{patients} is a list then any elements with more than 1 patient ID will be grouped together
and their quantities / summary statistics (as selected by \code{type})
will be calculated by taking the point-wise average. For example:
\code{patients = list("g1" = c("pt1", "pt2"), "g2" = c("pt3", "pt4"))} would result
in 2 groups being created whose values are the pointwise average
of \code{c("pt1", "pt2")} and \code{c("pt3", "pt4")} respectively.

If \code{patients=NULL} then all patients from original dataset will be selected
}

\seealso{
Other SurvivalSamples: 
\code{\link{SurvivalSamples-class}},
\code{\link{autoplot,SurvivalSamples-method}}
}
\concept{SurvivalSamples}
\concept{extractQuantities}
