% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extractQuantities}
\alias{extractQuantities}
\title{\code{extractQuantities}}
\usage{
extractQuantities(object, ...)
}
\arguments{
\item{object}{object to obtain quantities from}

\item{...}{additional options.}
}
\description{
Extracts quantity samples from a stan model
}
\concept{extractSamples}
