% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLongitudinal.R
\name{extractVariableNames.DataLongitudinal}
\alias{extractVariableNames.DataLongitudinal}
\title{Extract Mapping to Standardised Variable Names}
\usage{
\method{extractVariableNames}{DataLongitudinal}(object)
}
\arguments{
\item{object}{(\code{\link{DataLongitudinal}}) \cr Longitudinal Data.}
}
\value{
A list with the following named elements:
\itemize{
\item \code{subject} (\code{character})\cr The name of the variable containing the subject identifier
\item \code{frm} (\code{formula})\cr of the form \code{outcome ~ time}
\item \code{time} (\code{character})\cr The name of the variable containing the outcome time
\item \code{outcome} (\code{character})\cr The name of the variable containing the outcome values
\item \code{threshold} (\code{numeric})\cr cut-off value to be used to declare an observation as censored
(below detection limit).
}
}
\description{
Extract a \code{list} that maps the variable names in a user-defined
\code{data.frame} to standardised values.
}
\seealso{
Other DataLongitudinal: 
\code{\link{DataLongitudinal-class}},
\code{\link{as.data.frame.DataLongitudinal}()},
\code{\link{as_print_string.DataLongitudinal}()},
\code{\link{as_stan_list.DataSubject}()}

Other extractVariableNames: 
\code{\link{extractVariableNames}()},
\code{\link{extractVariableNames.DataSubject}()},
\code{\link{extractVariableNames.DataSurvival}()}
}
\concept{DataLongitudinal}
\concept{extractVariableNames}
