% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSubject.R
\name{extractVariableNames.DataSubject}
\alias{extractVariableNames.DataSubject}
\title{Extract Mapping to Standardised Variable Names}
\usage{
\method{extractVariableNames}{DataSubject}(object)
}
\arguments{
\item{object}{(\code{\link{DataSubject}}) \cr subject-level data.}
}
\value{
A list with the following named elements:
\itemize{
\item \code{subject} (\code{character})\cr the name of the variable containing the subject identifier.
\item \code{arm} (\code{character})\cr the name of the variable containing the arm identifier.
\item \code{study} (\code{character}) \cr the name of the variable containing the study identifier.
}
}
\description{
Extract a \code{list} that maps the variable names in a user-defined
\code{data.frame} to standardised values.
}
\seealso{
Other DataSubject: 
\code{\link{DataSubject-class}},
\code{\link{as.data.frame.DataSubject}()},
\code{\link{as_print_string.DataSubject}()},
\code{\link{as_stan_list.DataSubject}()}

Other extractVariableNames: 
\code{\link{extractVariableNames.DataLongitudinal}()},
\code{\link{extractVariableNames.DataSurvival}()},
\code{\link{extractVariableNames}()}
}
\concept{DataSubject}
\concept{extractVariableNames}
\keyword{internal}
