% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\name{extractVariableNames.DataSurvival}
\alias{extractVariableNames.DataSurvival}
\title{Extract Mapping to Standardised Variable Names}
\usage{
\method{extractVariableNames}{DataSurvival}(object)
}
\arguments{
\item{object}{(\code{\link{DataSurvival}}) \cr Survival Data.}
}
\value{
A list with the following named elements:
\itemize{
\item \code{arm} (\code{character})\cr The name of the variable containing the treatment arm
\item \code{study} (\code{character})\cr The name of the variable containing the study identifier
\item \code{pt} (\code{character})\cr The name of the variable containing the subject identifier
\item \code{frm} (\code{formula})\cr a symbolic description of the survival model to be fitted
\item \code{time} (\code{character})\cr  The name of the variable containing the event time
\item \code{event} (\code{character}) \cr  The name of the variable containing the event status
}
}
\description{
Extract a \code{list} that maps the variable names in a user-defined
\code{data.frame} to standardised values.
}
\seealso{
Other DataSurvival: 
\code{\link{DataSurvival-class}},
\code{\link{as.data.frame.DataSurvival}()},
\code{\link{as.list.DataSurvival}()}

Other extractVariableNames: 
\code{\link{extractVariableNames.DataLongitudinal}()},
\code{\link{extractVariableNames}()}
}
\concept{DataSurvival}
\concept{extractVariableNames}
\keyword{internal}
