% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extractVariableNames}
\alias{extractVariableNames}
\title{Extract Mapping to Standardised Variable Names}
\usage{
extractVariableNames(object)
}
\arguments{
\item{object}{the data object.}
}
\value{
A named list mapping source variables to standard names.
}
\description{
Extract a \code{list} that maps the variable names in a user-defined
\code{data.frame} to standardised values.
}
\seealso{
Other extractVariableNames:
\code{\link[=extractVariableNames.DataLongitudinal]{extractVariableNames.DataLongitudinal()}},
\code{\link[=extractVariableNames.DataSubject]{extractVariableNames.DataSubject()}},
\code{\link[=extractVariableNames.DataSurvival]{extractVariableNames.DataSurvival()}}
}
\concept{extractVariableNames}
\keyword{internal}
