% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_missing_rownumbers}
\alias{get_missing_rownumbers}
\title{\code{get_missing_rownumbers}}
\usage{
get_missing_rownumbers(df, formula = NULL)
}
\arguments{
\item{df}{A data.frame}

\item{formula}{A formula specifying which variables to inspect for missingness, if null
all variables are considered}
}
\description{
Returns which row numbers contain at least 1 missing observation any variables
}
\keyword{internal}
