% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{get_timepoints}
\alias{get_timepoints}
\title{Construct Time Intervals}
\usage{
get_timepoints(x)
}
\arguments{
\item{x}{(\code{numeric})\cr grid of time points.}
}
\value{
A \code{tibble} with \code{lower}, \code{upper}, \code{time} and \code{width}.
}
\description{
Construct Time Intervals
}
\keyword{internal}
