% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longmodel.R
\name{getLink}
\alias{getLink}
\alias{getLinkTTG}
\alias{getLinkDSLD}
\title{Access Hazard Link}
\usage{
getLink(object, selection)

getLinkTTG(object)

getLinkDSLD(object)
}
\arguments{
\item{object}{A \code{LongModel} object.}
}
\value{
A \code{HazardLink} object
}
\description{
Access the hazard link for a given longditudinal hazard model.
\itemize{
\item \code{\link[=getLinkTTG]{getLinkTTG()}} returns the TTG hazard link
\item \code{\link[=getLinkDSLD]{getLinkDSLD()}} returns the dSLD hazard link
\item \code{\link[=getLink]{getLink()}} is a convience function to return a hazard link created by merging
other hazard link functions together
}
}
