% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModel.R,
%   R/LongitudinalClaretBruno.R, R/LongitudinalClaretBrunoCov.R,
%   R/LongitudinalGSF.R, R/LongitudinalGSFCov.R, R/LongitudinalRandomSlope.R,
%   R/LongitudinalRandomSlopeCov.R, R/LongitudinalSteinFojo.R,
%   R/LongitudinalSteinFojoCov.R
\name{getRandomEffectsNames}
\alias{getRandomEffectsNames}
\alias{getRandomEffectsNames.default}
\alias{getRandomEffectsNames.JointModel}
\alias{getRandomEffectsNames.LongitudinalGSF}
\alias{getRandomEffectsNames.LongitudinalClaretBrunoCov}
\alias{getRandomEffectsNames.LongitudinalGSFCov}
\alias{getRandomEffectsNames.LongitudinalRandomSlope}
\alias{getRandomEffectsNames.LongitudinalRandomSlopeCov}
\alias{getRandomEffectsNames.LongitudinalSteinFojo}
\alias{getRandomEffectsNames.LongitudinalSteinFojoCov}
\title{Get Random Effects Names}
\usage{
getRandomEffectsNames(object, ...)

\method{getRandomEffectsNames}{default}(object, ...)

\method{getRandomEffectsNames}{JointModel}(object, ...)

\method{getRandomEffectsNames}{LongitudinalGSF}(object, ...)

\method{getRandomEffectsNames}{LongitudinalClaretBrunoCov}(object, ...)

\method{getRandomEffectsNames}{LongitudinalGSF}(object, ...)

\method{getRandomEffectsNames}{LongitudinalGSFCov}(object, ...)

\method{getRandomEffectsNames}{LongitudinalRandomSlope}(object, ...)

\method{getRandomEffectsNames}{LongitudinalRandomSlopeCov}(object, ...)

\method{getRandomEffectsNames}{LongitudinalSteinFojo}(object, ...)

\method{getRandomEffectsNames}{LongitudinalSteinFojoCov}(object, ...)
}
\arguments{
\item{object}{(\code{LongitudinalModel}) A longitudinal model object}

\item{...}{Not used.}
}
\value{
A named character vector of random-effect parameter names, or \code{NULL}.
}
\description{
Utility function that returns the names of the random effects parameters.
The main use for this is to allow the \code{\link{LongitudinalRandomEffects}} function
to know which parameters it needs to extract and to what common names
it should map the parameters to.
}
\examples{
getRandomEffectsNames(LongitudinalGSF())
}
