% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalGSF.R
\name{gsf_sld}
\alias{gsf_sld}
\alias{gsf_ttg}
\alias{gsf_dsld}
\title{Generalized Stein-Fojo Functionals}
\usage{
gsf_sld(time, b, s, g, phi)

gsf_ttg(time, b, s, g, phi)

gsf_dsld(time, b, s, g, phi)
}
\arguments{
\item{time}{(\code{numeric}) time grid.}

\item{b}{(\code{number}) baseline.}

\item{s}{(\code{number}) shrinkage.}

\item{g}{(\code{number}) growth.}

\item{phi}{(\code{number}) shrinkage proportion.}
}
\value{
The function results.
}
\description{
Generalized Stein-Fojo Functionals
}
\keyword{internal}
