% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/stan_module.R
\name{h_bracket}
\alias{h_bracket}
\alias{h_bracket,character-method}
\alias{h_bracket,list-method}
\title{h_bracket}
\usage{
h_bracket(x)

\S4method{h_bracket}{character}(x)

\S4method{h_bracket}{list}(x)
}
\arguments{
\item{x}{An Object}
}
\description{
Generic function that creates a character string oriented by curly brackets
}
