% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSubject.R, R/generics.R,
%   R/DataLongitudinal.R, R/DataSurvival.R
\name{harmonise.DataSubject}
\alias{harmonise.DataSubject}
\alias{harmonise}
\alias{harmonise.default}
\alias{harmonise.DataLongitudinal}
\alias{harmonise.DataSurvival}
\title{Prepare Data Object}
\usage{
\method{harmonise}{DataSubject}(object, ...)

harmonise(object, ...)

\method{harmonise}{default}(object, ...)

\method{harmonise}{DataLongitudinal}(object, subject_var, subject_ord, ...)

\method{harmonise}{DataSurvival}(object, subject_var, subject_ord, ...)
}
\arguments{
\item{object}{(\code{DataSubject} or \code{DataLongitudinal} or \code{DataSurvival}) \cr data object to "harmonise"}

\item{...}{not used.}

\item{subject_var}{(\code{character}) \cr the name of the variable containing the subject identifier.}

\item{subject_ord}{(\code{character}) \cr the expected levels (in order) of the subject identifier.}
}
\value{
Returns the original object but with the data standardised (see details)
}
\description{
Prepare Data Object
}
\details{
This utility function prepares the datasets in the data objects in order to ensure they
are consistent and compatible with each other.

In particular it ensures that the \code{subject} variable, as specified by \code{DataSubject},
is available in \code{DataLongitudinal} and \code{DataSurvival} and that all levels are present
in all 3 data objects.

It also sorts the datasets to ensure that indexes are consistent e.g. index 1 for
\code{DataSubject@data} corresponds to the same subject as index 1 for \code{DataSurvival@data}.
For \code{DataLongitudinal} the data is additionally sorted by time and outcome value.
}
\seealso{
\code{\link{DataJoint}}, \code{\link{DataSurvival}}, \code{\link{DataSubject}}, \code{\link{DataLongitudinal}}
}
\keyword{internal}
