% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_link.R
\docType{class}
\name{HazardLink-class}
\alias{HazardLink-class}
\alias{HazardLink}
\title{Hazard Link Class}
\description{
Creates a Hazard Link object which combines a longditudinal model to
an OS model
}
\section{Slots}{

\describe{
\item{\code{stan}}{A \code{StanModule} object as created by \code{\link[=StanModule]{StanModule()}} that specifies any Stan code
required for the link / contribution}

\item{\code{contribution}}{Stan code for what this link contriutes to the log hazard in the OS model}

\item{\code{parameter}}{The name of link parameter}
}}

\examples{
HazardLink(
    parameter = "beta_ttg",
    contribution = "beta_ttg * ttg(phi)",
    stan = StanModule(
        functions = "real ttg(real phi) {phi^2 };"
    )
)
}
