% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{hazardWindows}
\alias{hazardWindows}
\title{Generate time windows for evaluating a hazard function}
\usage{
hazardWindows(object, ...)
}
\arguments{
\item{object}{(\code{SurvivalModel}) object to generate time windows for.}

\item{...}{Not used.}
}
\value{
A \code{data.frame} describing the hazard-evaluation intervals.
}
\description{
Generate time windows for evaluating a hazard function
}
