% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HazardLink.R
\name{initialize,HazardLink-method}
\alias{initialize,HazardLink-method}
\title{Title - TODO}
\usage{
\S4method{initialize}{HazardLink}(.Object, ..., stan, contribution, parameter)
}
\description{
Description
TODO - Find way to merge roxygen page between constructor and initaliser
}
