% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OsModel.R
\name{initialize,LogLogisticOs-method}
\alias{initialize,LogLogisticOs-method}
\title{Title - TODO}
\usage{
\S4method{initialize}{LogLogisticOs}(.Object, ..., stan = LogLogisticModule(), templated)
}
\description{
Description
}
