% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/stan_module.R, R/hazard_link.R,
%   R/longmodel.R, R/OsModel.R
\name{inits}
\alias{inits}
\alias{inits,StanModule-method}
\alias{inits,HazardLink-method}
\alias{inits,LongModel-method}
\alias{inits,OsModel-method}
\title{inits() get method for \code{LongModel}, \code{HazardLink} , \code{StanModule} or \code{OsModel}}
\usage{
inits(object)

\S4method{inits}{StanModule}(object)

\S4method{inits}{HazardLink}(object)

\S4method{inits}{LongModel}(object)

\S4method{inits}{OsModel}(object)
}
\arguments{
\item{object}{A \code{LongModel}, \code{HazardLink} , \code{StanModule} or \code{OsModel} object}
}
\description{
inits() get method for \code{LongModel}, \code{HazardLink} , \code{StanModule} or \code{OsModel}
}
\examples{
inits(object)
}
