% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits_prep.R
\name{inits_prep,JMModel,JMdata-method}
\alias{inits_prep,JMModel,JMdata-method}
\title{Initial values preperation}
\usage{
\S4method{inits_prep}{JMModel,JMdata}(object, data)
}
\description{
Initial values preperation
}
