% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_contained.R
\name{is.contained}
\alias{is.contained}
\title{is.contained}
\usage{
is.contained(vec1, vec2)
}
\arguments{
\item{vec1}{A list}

\item{vec2}{A list}
}
\description{
a function to check if a list is included in a predefined list
}
