% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jm_data.R
\name{jm_data}
\alias{jm_data}
\title{JMdata class helper function}
\usage{
jm_data(sld, os, vars, shared_treatment, censoring_threshold)
}
\arguments{
\item{sld}{Longitudinal data}

\item{os}{Overall survival data}

\item{vars}{List with the names of columns important for the preparation of data for the stan object}

\item{shared_treatment}{Character, Identifies the treatment group}

\item{censoring_threshold}{Numeric, defines the threshold for the accuracy of the sld measurements}
}
\description{
JMdata class helper function
}
