% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heidel.R
\name{jm_post_check}
\alias{jm_post_check}
\title{Heidelberger and Welch's convergence diagnostic}
\usage{
jm_post_check(object)
}
\arguments{
\item{object}{A JMpost object}
}
\description{
Heidelberger and Welch's convergence diagnostic
}
