% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jm_data.R
\docType{class}
\name{JMdata-class}
\alias{JMdata-class}
\alias{jm_data}
\title{JMdata class}
\arguments{
\item{data_sld}{Longitudinal data}

\item{data_os}{Overall survival data}

\item{data}{Modified data and prepared for the mcmc run}

\item{vars}{List with the names of columns important for the preparation of data for the stan object}

\item{shared_treatment}{Character, Identifies the treatment group}

\item{censoring_threshold}{Numeric, defines the threshold for the accuracy of the sld measurements}
}
\description{
JMdata class
}
