% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{jmpost-settings}
\alias{jmpost-settings}
\alias{set_options}
\title{jmpost settings}
\usage{
set_options()
}
\description{
Define settings that modify the behaviour of the \code{jmpost} package

Each of the following are the name of options that can be set via:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(<option_name> = <value>)
}\if{html}{\out{</div>}}
\subsection{\code{jmpost.prior_shrinkage}}{

Default = \code{0.5}

By default all initial values are drawn as random sample from the respective prior
distribution with a shrinkage factor towards the mean. That is:

\if{html}{\out{<div class="sourceCode">}}\preformatted{initial_value = prior_mean * prior_shrinkage + (1 - prior_shrinkage) * prior_sample
}\if{html}{\out{</div>}}

This setting controls the shrinkage factor. A value of 0 means no shrinkage (i.e.
pure random draw) whilst a value of 1 means the initial value is just the mean.
}

\subsection{\code{jmpost.cache_dir}}{

Default = \code{tempfile()}

Directory to store compiled stan models in. If not set, a temporary directory is used for
the given R session. Can also be set via the environment variable \code{JMPOST_CACHE_DIR}.
}

\subsection{\code{jmpost.gauss_quad_n}}{

Default = 15

In most cases the survival function of the joint model does not have a closed form
and as such it is calculated by integrating the hazard function. \code{jmpost} estimates this
via Gaussian Quadrature, in particular it uses \code{\link[statmod:gauss.quad]{statmod::gauss.quad}} with
\code{kind = "legendre"} to create the nodes and weights.

This option specifies the \code{n} argument in the call to \code{\link[statmod:gauss.quad]{statmod::gauss.quad}}. In general
higher values of \code{n} lead to better accuracy of the approximation but at the cost of
increased computational time.
}

\subsection{\code{jmpost.double_eps} and \code{jmpost.double_neg_eps}}{

Default = \code{0} and \code{sqrt(.Machine$double.neg.eps)}

These are small floating-point numbers \code{x} which are used in parameter
definitions as lower boundaries to prevent numerical issues with the MCMC
sampler in the beginning of the sampling process.
}
}
\examples{
\dontrun{
options(jmpost.prior_shrinkage = 0.5)
}
}
