% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModel.R
\name{JointModel-Shared}
\alias{JointModel-Shared}
\title{Re-used documentation for \code{JointModel-Shared}}
\arguments{
\item{object}{(\code{\link{JointModel}}) \cr Joint model specification.}

\item{x}{(\code{\link{JointModel}}) \cr Joint model specification.}

\item{...}{Not Used.}
}
\description{
Re-used documentation for \code{JointModel-Shared}
}
\keyword{internal}
