% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModelQuantities.R
\docType{class}
\name{JointModelQuantities-class}
\alias{JointModelQuantities-class}
\alias{.JointModelQuantities}
\alias{JointModelQuantities}
\title{\code{JointModelQuantities}}
\usage{
JointModelQuantities(joint_model, generated_quantities)
}
\arguments{
\item{joint_model}{(\code{JointModelSamples})\cr the Model Samples Object.}

\item{generated_quantities}{(\code{CmdStanGQ})\cr A \code{CmdStanGQ} object.}
}
\description{
Contains samples and generated quantities from a \code{\link{JointModel}}.
}
\section{Slots}{

\describe{
\item{\code{model}}{(\code{StanModule})\cr the original model.}

\item{\code{data}}{(\code{list})\cr data input.}

\item{\code{init}}{(\code{list})\cr initial values.}

\item{\code{results}}{(\code{CmdStanMCMC})\cr the results from \code{\link[=sampleStanModel]{sampleStanModel()}}.}

\item{\code{quantites}}{(\code{CmdStanGQ})\cr the results from \code{\link[=generateQuantities]{generateQuantities()}}.}
}}

