% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link.R
\name{length.Link}
\alias{length.Link}
\title{\code{Link} -> \code{list}}
\usage{
\method{length}{Link}(x)
}
\arguments{
\item{x}{(\code{Link}) a link object.}
}
\value{
A single integer giving the number of link components.
}
\description{
Returns the number of link components within the \code{\link{Link}} object
}
\seealso{
Other Link:
\code{\link{Link-class}},
\code{\link[=as.StanModule.Link]{as.StanModule.Link()}},
\code{\link[=as.list.Link]{as.list.Link()}}
}
\concept{Link}
