% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link.R
\docType{class}
\name{Link-class}
\alias{Link-class}
\alias{.Link}
\alias{Link}
\title{\code{Link}}
\usage{
Link(...)
}
\arguments{
\item{...}{(\code{\link{LinkComponent}} or \code{\link{PromiseLinkComponent}}) an arbitrary number of link components.}
}
\value{
A \code{Link} object.
}
\description{
Simple container class to enable the use of multiple link components in a joint model.
Note that the constructor of this object is idempotent e.g. \code{Link(Link(x)) == Link(x)}
}
\section{Slots}{

\describe{
\item{\code{components}}{(\code{list})\cr a list of \code{\link{LinkComponent}} or \code{\link{PromiseLinkComponent}} objects.}

\item{\code{resolved}}{(\code{logical})\cr indicates if all the \code{components} have been resolved.}
}}

\examples{
Link(
    linkDSLD(),
    linkTTG()
  )

}
\seealso{
Other Link:
\code{\link[=as.StanModule.Link]{as.StanModule.Link()}},
\code{\link[=as.list.Link]{as.list.Link()}},
\code{\link[=length.Link]{length.Link()}}
}
\concept{Link}
