% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link.R
\name{Link-Shared}
\alias{Link-Shared}
\title{\code{Link} Function Arguments}
\arguments{
\item{x}{(\code{\link{Link}})\cr a link object.}

\item{object}{(\code{\link{Link}})\cr a link object.}

\item{...}{Not Used.}
}
\description{
This exists just to contain all the common arguments for \code{\link{Link}} methods.
}
\keyword{internal}
