% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkGSF.R
\docType{class}
\name{link_gsf_dsld-class}
\alias{link_gsf_dsld-class}
\alias{.link_gsf_dsld}
\alias{link_gsf_dsld}
\title{\code{link_gsf_dsld}}
\usage{
link_gsf_dsld(beta = prior_normal(0, 5, init = 0))
}
\arguments{
\item{beta}{(\code{Prior})\cr prior for the link coefficient \code{beta}.}
}
\description{
This class extends the general \code{\link{link_gsf_abstract}} for the derivative of the
sum of longest diameters (\code{dsld}) link contribution.
}
