% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkGSF.R
\docType{class}
\name{link_gsf_identity-class}
\alias{link_gsf_identity-class}
\alias{.link_gsf_identity}
\alias{link_gsf_identity}
\title{\code{link_gsf_identity}}
\usage{
link_gsf_identity(tau = prior_normal(0, 5, init = 0))
}
\arguments{
\item{tau}{(\code{Prior})\cr prior for the link coefficient \code{tau}.}
}
\description{
This class extends the general \code{\link{link_gsf_abstract}} for the identity of the
sum of longest diameters (\code{sld}) link contribution.
}
