% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkGSF.R
\docType{class}
\name{link_gsf_ttg-class}
\alias{link_gsf_ttg-class}
\alias{.link_gsf_ttg}
\alias{link_gsf_ttg}
\title{\code{link_gsf_ttg}}
\usage{
link_gsf_ttg(gamma = prior_normal(0, 5, init = 0))
}
\arguments{
\item{gamma}{(\code{Prior})\cr prior for the link coefficient \code{gamma}.}
}
\description{
This class extends the general \code{\link{link_gsf_abstract}} for the time-to-growth
(\code{ttg}) link contribution.
}
