% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkSF.R
\docType{class}
\name{link_sf_dsld-class}
\alias{link_sf_dsld-class}
\alias{.link_sf_dsld}
\alias{link_sf_dsld}
\title{\code{link_sf_dsld}}
\usage{
link_sf_dsld(beta = prior_normal(0, 5))
}
\arguments{
\item{beta}{(\code{Prior})\cr prior for the link coefficient \code{beta}.}
}
\description{
This class extends the general \code{\link{link_sf_abstract}} for the derivative of the
sum of longest diameters (\code{dsld}) link contribution.
}
