% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkSF.R
\docType{class}
\name{link_sf_ttg-class}
\alias{link_sf_ttg-class}
\alias{.link_sf_ttg}
\alias{link_sf_ttg}
\title{\code{link_sf_ttg}}
\usage{
link_sf_ttg(gamma = prior_normal(0, 5))
}
\arguments{
\item{gamma}{(\code{Prior})\cr prior for the link coefficient \code{gamma}.}
}
\description{
This class extends the general \code{\link{link_sf_abstract}} for the time-to-growth
(\code{ttg}) link contribution.
}
