% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkComponent.R
\docType{class}
\name{LinkComponent-class}
\alias{LinkComponent-class}
\alias{.LinkComponent}
\alias{LinkComponent}
\title{\code{LinkComponent}}
\usage{
LinkComponent(stan, prior, key, ...)
}
\arguments{
\item{stan}{(\code{StanModule}) Stan code. See Details.}

\item{prior}{(\code{Prior}) The prior for the scaling coeficient.}

\item{key}{(\code{character}) Link identifier. See Details.}

\item{...}{additional arguments for \code{\link[=StanModel]{StanModel()}}.}
}
\value{
A \code{LinkComponent} object.
}
\description{
\code{LinkComponent}
}
\details{
This object provides key information needed to construct a link contribution in the
survival model based on the parameters of the longitudinal model.

Each link component defines a stan function of the longitudinal model parameters which is
multiplied by a model coefficient and added to the survival models hazard function.

For full details about the specification of a \code{LinkComponent} please see
\code{vignette("extending-jmpost", package = "jmpost")}.
}
\section{Slots}{

\describe{
\item{\code{stan}}{(\code{StanModule})\cr See Arguments.}

\item{\code{name}}{(\code{character})\cr See Arguments.}

\item{\code{parameters}}{(\code{ParameterList})\cr The parameter specification.}
}}

\examples{
LinkComponent(
  stan = StanModule(),
  prior = prior_normal(0, 1),
  key = "custom_link"
)
}
\seealso{
Other LinkComponent:
\code{\link[=as.StanModule.LinkComponent]{as.StanModule.LinkComponent()}},
\code{\link[=as.list.LinkComponent]{as.list.LinkComponent()}},
\code{\link[=getParameters]{getParameters()}},
\code{\link[=initialValues]{initialValues()}}
}
\concept{LinkComponent}
