% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkComponent.R
\name{LinkComponent-Shared}
\alias{LinkComponent-Shared}
\title{\code{LinkComponent} Function Arguments}
\arguments{
\item{stan}{(\code{StanModule})\cr Stan code.}

\item{x}{(\code{\link{LinkComponent}})\cr a link component.}

\item{object}{(\code{\link{LinkComponent}})\cr a link component.}

\item{...}{Not Used.}
}
\description{
This exists to contain all the common arguments for \code{\link{LinkComponent}} methods.
}
\keyword{internal}
