% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkGSF.R
\docType{class}
\name{LinkGSF-class}
\alias{LinkGSF-class}
\alias{.LinkGSF}
\alias{LinkGSF}
\title{\code{LinkGSF}}
\usage{
LinkGSF(...)
}
\arguments{
\item{...}{(\code{link_gsf_abstract})\cr which link components should be included. If no arguments
are provided then this will be set to \code{\link[=link_gsf_dsld]{link_gsf_dsld()}} and \code{\link[=link_gsf_ttg]{link_gsf_ttg()}}}
}
\description{
This class extends the general \code{\link[=Link-class]{Link}} class for
the \code{\link[=LongitudinalGSF-class]{LongitudinalGSF}}
model.
}
