% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkNone.R
\docType{class}
\name{LinkNone-class}
\alias{LinkNone-class}
\alias{.LinkNone}
\alias{LinkNone}
\title{\code{LinkNone}}
\usage{
LinkNone()
}
\description{
This class extends the general \code{\link{Link}} class to specify no link
between the longitudinal and the time-to-event models.
}
