% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkRandomSlope.R
\docType{class}
\name{LinkRandomSlope-class}
\alias{LinkRandomSlope-class}
\alias{.LinkRandomSlope}
\alias{LinkRandomSlope}
\title{\code{LinkRandomSlope}}
\usage{
LinkRandomSlope(link_lm_phi = prior_normal(0.2, 0.5, init = 0.02))
}
\arguments{
\item{link_lm_phi}{(\code{Prior})\cr prior for the link coefficient for the
random slope link contribution.}
}
\description{
This class extends the general \code{\link{Link}} class for the \code{\link{LongitudinalRandomSlope}}
model.
}
