% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkSF.R
\docType{class}
\name{LinkSF-class}
\alias{LinkSF-class}
\alias{.LinkSF}
\alias{LinkSF}
\title{\code{LinkSF}}
\usage{
LinkSF(...)
}
\arguments{
\item{...}{(\code{link_sf_abstract})\cr which link components should be included. If no arguments
are provided then this will be set to \code{\link[=link_sf_dsld]{link_sf_dsld()}} and \code{\link[=link_sf_ttg]{link_sf_ttg()}}}
}
\description{
This class extends the general \code{\link[=Link-class]{Link}} class for
the \code{\link[=LongitudinalSF-class]{LongitudinalSF}}
model.
}
