% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_h0.R
\name{log_h0}
\alias{log_h0}
\title{Baseline log hazard function}
\usage{
log_h0(time, lambda, p)
}
\arguments{
\item{time}{Times to be predicted}

\item{lambda}{The lambda estimate}

\item{p}{The p estimate}
}
\description{
Baseline log hazard function
}
