% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_haz.R
\name{log_haz}
\alias{log_haz}
\title{Log hazard function}
\usage{
log_haz(
  time,
  lambda,
  p,
  beta,
  gamma,
  psi_bsld,
  psi_ks,
  psi_kg,
  psi_phi,
  beta_os_cov,
  os_cov_design
)
}
\arguments{
\item{psi_ks}{The srinkage rate parameter}

\item{psi_kg}{The growth rate parameter}

\item{psi_phi}{The percentage of cells responded to the treatment}
}
\description{
Log hazard function
}
