% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_surv.R
\name{log_surv}
\alias{log_surv}
\title{Log survival function}
\usage{
log_surv(
  time,
  lambda,
  p,
  beta,
  gamma,
  psi_bsld,
  psi_ks,
  psi_kg,
  psi_phi,
  nodes,
  weights,
  beta_os_cov,
  os_cov_design
)
}
\arguments{
\item{psi_ks}{The srinkage rate parameter}

\item{psi_kg}{The growth rate parameter}

\item{psi_phi}{The percentage of cells responded to the treatment}
}
\description{
Log survival function
}
