% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OsModel.R
\docType{class}
\name{LogLogisticOs-class}
\alias{LogLogisticOs-class}
\alias{LogLogisticOs}
\title{Log-logistic overall survival}
\description{
Description - Log-logistic overall survival model object creator
}
