% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModelSamples.R
\name{longitudinal}
\alias{longitudinal}
\alias{longitudinal,JointModelSamples-method}
\title{\code{longitudinal}}
\usage{
longitudinal(object, ...)

\S4method{longitudinal}{JointModelSamples}(object, patients = NULL, ...)
}
\arguments{
\item{object}{(\code{JointModelSamples})\cr samples to extract the longitudinal fits from.}

\item{...}{not used.}

\item{patients}{(\code{character} or \code{NULL})\cr optional subset of patients for
which the longitudinal fit samples should be extracted, the default \code{NULL}
meaning all patients.}
}
\value{
The longitudinal fit samples.
}
\description{
\code{longitudinal}
}
