% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalQuantities.R
\name{longitudinal_plot}
\alias{longitudinal_plot}
\title{Longitudinal Plot}
\usage{
longitudinal_plot(data, data_obs = NULL, add_ci = FALSE)
}
\arguments{
\item{data}{(\code{data.frame}) summary statistics for longitudinal
value estimates. See details.}

\item{data_obs}{(\code{data.frame}) real observed values to be
overlaid for reference.  See details.}

\item{add_ci}{(\code{logical}) Should confidence intervals be added? Default = \code{TRUE}.}
}
\value{
A \code{ggplot} object.
}
\description{
Internal plotting function to create longitudinal plots
This function predominately exists to extract core logic into its own function
to enable easier unit testing.
}
\details{
\subsection{\code{data}}{

Should contain the following columns:
\itemize{
\item \code{time} (\code{numeric}) \cr time point for the summary statistic.
\item \code{group} (\code{character}) \cr the group in which the observation belongs to.
\item \code{median} (\code{numeric}) \cr the median value for the summary statistic.
\item \code{upper} (\code{numeric}) \cr the upper 95\% CI for the summary statistic.
\item \code{lower} (\code{numeric}) \cr the lower 95\% CI for the summary statistic.
}
}

\subsection{\code{data_obs}}{

Should contain the following columns:
\itemize{
\item \code{time} (\code{numeric}) \cr the time at which the observed value occurred.
\item \code{Yob} (\code{numeric}) \cr the real observed value.
\item \code{group} (\code{character}) \cr which group the event belongs to, should correspond to
values in \code{data$group}.
}
}
}
\keyword{internal}
