% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalClaretBruno.R
\docType{class}
\name{LongitudinalClaretBruno-class}
\alias{LongitudinalClaretBruno-class}
\alias{.LongitudinalClaretBruno}
\alias{LongitudinalClaretBruno}
\title{\code{LongitudinalClaretBruno}}
\usage{
LongitudinalClaretBruno(
  mu_b = prior_normal(log(60), 0.5),
  mu_g = prior_normal(log(1), 0.5),
  mu_c = prior_normal(log(0.4), 0.5),
  mu_p = prior_normal(log(2), 0.5),
  omega_b = prior_lognormal(log(0.2), 0.5),
  omega_g = prior_lognormal(log(0.2), 0.5),
  omega_c = prior_lognormal(log(0.2), 0.5),
  omega_p = prior_lognormal(log(0.2), 0.5),
  sigma = prior_lognormal(log(0.1), 0.5),
  scaled_variance = FALSE,
  centred = FALSE
)
}
\arguments{
\item{mu_b}{(\code{Prior}) for the mean population baseline sld value.}

\item{mu_g}{(\code{Prior}) for the mean population growth rate.}

\item{mu_c}{(\code{Prior}) for the mean population resistance rate.}

\item{mu_p}{(\code{Prior}) for the mean population growth inhibition}

\item{omega_b}{(\code{Prior}) for the population standard deviation for the baseline sld value.}

\item{omega_g}{(\code{Prior}) for the population standard deviation for the growth rate.}

\item{omega_c}{(\code{Prior}) for the population standard deviation for the resistance rate.}

\item{omega_p}{(\code{Prior}) for the population standard deviation for the growth inhibition.}

\item{sigma}{(\code{Prior}) for the variance of the longitudinal values.}

\item{scaled_variance}{(\code{logical}) whether the variance should be scaled by the expected value, corresponding to a multiplicative model.
As a default, the variance is not scaled by the expected value, corresponding to an additive model.
(See the "Statistical Specifications" vignette for more details.)}

\item{centred}{(\code{logical}) whether to use the centred parameterization.}
}
\value{
A \code{LongitudinalClaretBruno} object.
}
\description{
This class extends the general \code{\link{LongitudinalModel}} class for using the
Claret-Bruno model for the longitudinal outcome.
}
\section{Available Links}{

\itemize{
\item \code{\link[=linkDSLD]{linkDSLD()}}
\item \code{\link[=linkTTG]{linkTTG()}}
\item \code{\link[=linkIdentity]{linkIdentity()}}
\item \code{\link[=linkGrowth]{linkGrowth()}}
}
}

\examples{
LongitudinalClaretBruno()
}
