% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalQuantities.R
\name{LongitudinalQuantities-Shared}
\alias{LongitudinalQuantities-Shared}
\title{Re-used documentation for \code{LongitudinalQuantities}}
\arguments{
\item{x}{(\code{LongitudinalQuantities}) longitudinal quantities.}

\item{object}{(\code{LongitudinalQuantities}) longitudinal quantities.}

\item{...}{not used.}
}
\description{
Re-used documentation for \code{LongitudinalQuantities}
}
\keyword{internal}
