% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalSamples.R
\docType{class}
\name{LongitudinalSamples}
\alias{LongitudinalSamples}
\alias{.LongitudinalSamples}
\alias{[,LongitudinalSamples,ANY,ANY,ANY-method}
\title{Longitudinal Samples Storage}
\usage{
\S4method{[}{LongitudinalSamples,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{(\code{LongitudinalSamples})\cr the samples object to subset.}

\item{i}{(\code{vector})\cr the index vector.}
}
\value{
The subsetted \code{LongitudinalSamples} object.
}
\description{
This class is an extension of a standard \code{list} so that we
can define custom methods for it.
}
