% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalSF.R
\docType{class}
\name{LongitudinalSF-class}
\alias{LongitudinalSF-class}
\alias{.LongitudinalSF}
\alias{LongitudinalSF}
\title{\code{LongitudinalSF}}
\usage{
LongitudinalSF(
  mu_bsld = prior_lognormal(log(55), 5, init = 55),
  mu_ks = prior_lognormal(log(0.1), 0.5, init = 0.1),
  mu_kg = prior_lognormal(log(0.1), 1, init = 0.1),
  omega_bsld = prior_lognormal(log(0.1), 1, init = 0.1),
  omega_ks = prior_lognormal(log(0.1), 1, init = 0.1),
  omega_kg = prior_lognormal(log(0.1), 1, init = 0.1),
  sigma = prior_lognormal(log(0.1), 0.8, init = 0.1)
)
}
\arguments{
\item{mu_bsld}{(\code{Prior})\cr for the mean baseline value \code{mu_bsld}.}

\item{mu_ks}{(\code{Prior})\cr for the mean shrinkage rate \code{mu_ks}.}

\item{mu_kg}{(\code{Prior})\cr for the mean growth rate \code{mu_kg}.}

\item{omega_bsld}{(\code{Prior})\cr for the baseline value standard deviation \code{omega_bsld}.}

\item{omega_ks}{(\code{Prior})\cr for the shrinkage rate standard deviation \code{omega_ks}.}

\item{omega_kg}{(\code{Prior})\cr for the growth rate standard deviation \code{omega_kg}.}

\item{sigma}{(\code{Prior})\cr for the variance of the longitudinal values \code{sigma}.}
}
\description{
This class extends the general \code{\link{LongitudinalModel}} class for using the
Stein-Fojo (SF) model for the longitudinal outcome.
}
