% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longmodel.R
\docType{class}
\name{LongModel-class}
\alias{LongModel-class}
\alias{LongModel}
\alias{getLinkTTG,LongModel-method}
\alias{getLinkDSLD,LongModel-method}
\alias{getLink,LongModel-method}
\title{\code{LongModel}}
\usage{
\S4method{getLinkTTG}{LongModel}(object)

\S4method{getLinkDSLD}{LongModel}(object)

\S4method{getLink}{LongModel}(object, selection = c("ttg", "dsld"))
}
\description{
This represents an abstract class for longditudinal models.
This class only exists for other concrete instatiations to inherit off of.
}
\details{
Provides default implementation of the \code{\link[=getLink]{getLink()}} method.
Child classes should define for themselves \code{\link[=getLinkTTG]{getLinkTTG()}} and
\code{\link[=getLinkDSLD]{getLinkDSLD()}} or alternative provide their own implementations
for \code{\link[=getLink]{getLink()}}.
}
\section{Slots}{

\describe{
\item{\code{stan}}{a \code{StanModule} object as created by \code{\link[=StanModule]{StanModule()}}}
}}

