% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StanModule.R, R/ParameterList.R,
%   R/defaults.R
\name{merge}
\alias{merge}
\alias{merge,StanModule,StanModule-method}
\alias{merge,ParameterList,ParameterList-method}
\alias{merge,StanModel,NULL-method}
\alias{merge,NULL,StanModel-method}
\alias{merge,StanModule,NULL-method}
\alias{merge,NULL,StanModule-method}
\alias{merge,ParameterList,NULL-method}
\alias{merge,NULL,ParameterList-method}
\alias{merge,NULL,NULL-method}
\title{\code{merge}}
\usage{
merge(x, y, ...)

\S4method{merge}{StanModule,StanModule}(x, y, stan_blocks = STAN_BLOCKS, ...)

\S4method{merge}{ParameterList,ParameterList}(x, y)

\S4method{merge}{StanModel,NULL}(x, y, ...)

\S4method{merge}{NULL,StanModel}(x, y, ...)

\S4method{merge}{StanModule,NULL}(x, y, ...)

\S4method{merge}{NULL,StanModule}(x, y, ...)

\S4method{merge}{ParameterList,NULL}(x, y, ...)

\S4method{merge}{NULL,ParameterList}(x, y, ...)

\S4method{merge}{NULL,NULL}(x, y, ...)
}
\arguments{
\item{x}{first module.}

\item{y}{second module.}

\item{...}{additional arguments.}

\item{stan_blocks}{(\code{list}) reference list of stan blocks.}
}
\value{
A merged object of the same class as \code{x} and \code{y}.
}
\description{
Merge two \code{StanModule} or \code{ParameterList} objects.
}
\examples{
x <- ParameterList(Parameter(prior_normal(0, 1), "x"))
y <- ParameterList(Parameter(prior_normal(0, 1), "y"))
merge(x, y)
}
