% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_module.R
\name{model_prep}
\alias{model_prep}
\title{model_prep. Populates the model section of a StanModule}
\usage{
model_prep(x)
}
\arguments{
\item{x}{A StanModule object}
}
\description{
model_prep. Populates the model section of a StanModule
}
