% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Parameter.R, R/ParameterList.R
\name{names}
\alias{names}
\alias{names,Parameter-method}
\alias{names,ParameterList-method}
\title{Names of Different Classes}
\usage{
\S4method{names}{Parameter}(x)

\S4method{names}{ParameterList}(x)
}
\arguments{
\item{x}{where to get the names from.}
}
\description{
These methods return the names of objects of different classes.
}
