% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{os_prior}
\alias{os_prior}
\title{os_prior A list of prior densities for the overall survival model.}
\usage{
os_prior(
  p = "gamma(2, 0.5)",
  `1/lambda` = "lognormal(0, 5)",
  beta_os_cov = "normal(0, 5)"
)
}
\description{
os_prior A list of prior densities for the overall survival model.
}
