% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OsModel.R
\docType{class}
\name{OsModel-class}
\alias{OsModel-class}
\alias{OsModel}
\title{Overall survival generic object for jmpost package.}
\description{
Creates an object containing all the stan code relevant to the overall survival part of a joint model.
}
\section{Slots}{

\describe{
\item{\code{stan}}{StanModule. Object containing all the nessesary stan parts.}
}}

