% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\docType{class}
\name{Parameter-class}
\alias{Parameter-class}
\alias{.Parameter}
\alias{Parameter}
\title{\code{Parameter}}
\usage{
Parameter(prior, name, size = 1)
}
\arguments{
\item{prior}{(\code{Prior}) for the parameter.}

\item{name}{(\code{string}) of the parameter.}

\item{size}{(\code{numeric} or \code{string}) dimension of the parameter.}
}
\value{
A \code{Parameter} object.
}
\description{
Stores the name, the prior distribution and the size of a parameter.
If \code{size} is a string then this indicates the name of the variable
within the stan data object that specifies the size of this parameter.
}
\section{Slots}{

\describe{
\item{\code{name}}{(\code{string})\cr of the parameter.}

\item{\code{prior}}{(\code{Prior})\cr for the parameter.}

\item{\code{size}}{(\code{numeric} or \code{string})\cr dimension of the parameter.}
}}

\examples{
Parameter(prior = prior_normal(0, 1), name = "beta")
}
\seealso{
Other Parameter:
\code{\link{Parameter-Getter-Methods}},
\code{\link[=as.StanModule.Parameter]{as.StanModule.Parameter()}},
\code{\link[=as.character.Parameter]{as.character.Parameter()}},
\code{\link[=as_stan_list.Parameter]{as_stan_list.Parameter()}}
}
\concept{Parameter}
