% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{Parameter-Shared}
\alias{Parameter-Shared}
\title{\code{Parameter} Function Arguments}
\arguments{
\item{x}{(\code{\link{Parameter}})\cr a prior Distribution}

\item{object}{(\code{\link{Parameter}})\cr a prior Distribution}

\item{...}{Not Used.}
}
\description{
The documentation lists all the conventional arguments for \code{\link{Parameter}}
constructors.
}
\keyword{internal}
