% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterList.R
\docType{class}
\name{ParameterList-class}
\alias{ParameterList-class}
\alias{.ParameterList}
\alias{ParameterList}
\title{\code{ParameterList}}
\usage{
ParameterList(...)
}
\arguments{
\item{...}{(\code{Parameter}) which parameter specifications to include.}
}
\description{
This class extends the general \code{\link{list}} type for containing \code{\link{Parameter}}
specifications. When converted to a \code{\link{StanModule}}, the list supplies the
Stan declarations and prior statements for those parameters.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{(\code{list}) \cr a list of \code{\link{Parameter}} objects}
}}

\seealso{
Other ParameterList:
\code{\link{ParameterList-Getter-Methods}},
\code{\link[=as.StanModule.ParameterList]{as.StanModule.ParameterList()}},
\code{\link[=as.list.ParameterList]{as.list.ParameterList()}},
\code{\link[=as_print_string.ParameterList]{as_print_string.ParameterList()}},
\code{\link[=as_stan_list.ParameterList]{as_stan_list.ParameterList()}}
}
\concept{ParameterList}
