% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterList.R
\name{ParameterList-Getter-Methods}
\alias{ParameterList-Getter-Methods}
\alias{names.ParameterList}
\alias{initialValues.ParameterList}
\alias{size.ParameterList}
\title{Parameter-List Getter Functions}
\usage{
\method{names}{ParameterList}(x)

\method{initialValues}{ParameterList}(object, n_chains, ...)

\method{size}{ParameterList}(object)
}
\arguments{
\item{x}{(\code{ParameterList}) A List of \code{\link{Parameter}} Objects.}

\item{object}{(\code{ParameterList}) A List of \code{\link{Parameter}} Objects.}

\item{n_chains}{(\code{integer}) the number of chains.}

\item{...}{Not Used.}
}
\value{
The requested parameter-list property: names, initial values, or total size.
}
\description{
Getter functions for the slots of a \code{\link{ParameterList}} object
}
\section{Functions}{
\itemize{
\item \code{names(ParameterList)}: The parameter-list's parameter names

\item \code{initialValues(ParameterList)}: The parameter-list's parameter initial values

\item \code{size(ParameterList)}: The parameter-list's parameter dimensionality

}}
\seealso{
Other ParameterList:
\code{\link{ParameterList-class}},
\code{\link[=as.StanModule.ParameterList]{as.StanModule.ParameterList()}},
\code{\link[=as.list.ParameterList]{as.list.ParameterList()}},
\code{\link[=as_print_string.ParameterList]{as_print_string.ParameterList()}},
\code{\link[=as_stan_list.ParameterList]{as_stan_list.ParameterList()}}
}
\concept{ParameterList}
