% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_module.R
\name{paste_str}
\alias{paste_str}
\title{paste_vector Returns a single string from a vector of strings}
\usage{
paste_str(vec)
}
\arguments{
\item{vec}{A vector of stings representing stan code.}
}
\description{
paste_vector Returns a single string from a vector of strings
}
