% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationHR.R
\name{populationHR}
\alias{populationHR}
\title{Calculate Population Hazard Ratios}
\usage{
populationHR(
  object,
  hr_formula = object@data@survival@formula,
  baseline = ~bs(time, df = 10),
  quantiles = c(0.025, 0.975)
)
}
\arguments{
\item{object}{(\code{JointModelSamples}) samples as drawn from a Joint Model.}

\item{hr_formula}{(\code{formula}) defines the terms to include in the hazard ratio calculation.
By default this uses the right side of the formula used in the survival model.
Set to \code{NULL} not include any terms}

\item{baseline}{(\code{formula}) terms to model baseline hazard using variable \code{time}.
Default is a B-spline from \link{splines}: \code{~bs(time, df = 10)}}

\item{quantiles}{(\code{numeric}) vector of two values in (0, 1) for calculating quantiles from log hazard ratio
distributions.}
}
\value{
A list containing a summary of parameter distributions as a \code{data.frame} and a
matrix containing the parameter estimates for each sample.
}
\description{
Calculates hazard ratios marginalised over subject specific random effects using the
approach proposed by \insertCite{oudenhoven2020marginal}{jmpost}.
}
\examples{
\dontrun{
populationHR(fit, quantiles = c(0.01, 0.99))
}
}
\references{
\insertAllCited{}
}
