% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict}
\alias{predict}
\title{Model Predictions}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{...}{additional arguments affecting the predictions produced.}
}
\description{
NOTE: This man page is for the \code{predict} S4 generic function defined within
jmpost. See \code{\link[stats:predict]{stats::predict()}} for the default method.
}
\seealso{
Other predict: 
\code{\link{predict,SurvivalSamples-method}}
}
\concept{predict}
