% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Prior-Shared}
\alias{Prior-Shared}
\title{\code{Prior} Function Arguments}
\arguments{
\item{centre}{(\code{number}) the central point of distribution to shrink sampled values towards
(for most distributions this is the mean or median if the mean is undefined)}

\item{x}{(\code{Prior}) a prior Distribution}

\item{object}{(\code{Prior}) a prior Distribution}

\item{name}{(\code{character}) the name of the parameter the prior distribution is for}

\item{size}{(\code{numeric} or \code{character}) the parameter size.}

\item{...}{Not Used.}
}
\description{
The documentation lists all the conventional arguments for \code{\link{Prior}}
constructors.
}
\keyword{internal}
