% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{prior_arguments}
\alias{prior_arguments}
\title{\code{Prior} Function Arguments}
\arguments{
\item{init}{(\code{number})\cr initial value.}
}
\description{
The documentation lists all the conventional arguments for \code{\link{Prior}}
constructors.
}
\keyword{internal}
