% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_const_vector}
\alias{prior_const_vector}
\title{Constant Prior Distribution for a Vector}
\usage{
prior_const_vector(values)
}
\arguments{
\item{values}{(\code{numeric}) the fixed parameter values.}
}
\value{
A \code{Prior} object.
}
\description{
Constant Prior Distribution for a Vector
}
\examples{
prior_const_vector(c(0, 1))
}
\seealso{
Other Prior:
\code{\link[=prior_beta]{prior_beta()}},
\code{\link[=prior_cauchy]{prior_cauchy()}},
\code{\link[=prior_const]{prior_const()}},
\code{\link[=prior_gamma]{prior_gamma()}},
\code{\link[=prior_horseshoe]{prior_horseshoe()}},
\code{\link[=prior_init_only]{prior_init_only()}},
\code{\link[=prior_invgamma]{prior_invgamma()}},
\code{\link[=prior_logistic]{prior_logistic()}},
\code{\link[=prior_loglogistic]{prior_loglogistic()}},
\code{\link[=prior_lognormal]{prior_lognormal()}},
\code{\link[=prior_normal]{prior_normal()}},
\code{\link[=prior_normal_vector]{prior_normal_vector()}},
\code{\link[=prior_std_normal]{prior_std_normal()}},
\code{\link[=prior_student_t]{prior_student_t()}},
\code{\link[=prior_uniform]{prior_uniform()}}
}
\concept{Prior}
