% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_loglogistic}
\alias{prior_loglogistic}
\title{Log-Logistic Prior Distribution}
\usage{
prior_loglogistic(alpha, beta)
}
\arguments{
\item{alpha}{(\code{number}) Scale parameter.}

\item{beta}{(\code{number}) Shape parameter.}
}
\value{
A \code{Prior} object.
}
\description{
Log-Logistic Prior Distribution
}
\examples{
prior_loglogistic(2, 3)
}
\seealso{
Other Prior:
\code{\link[=prior_beta]{prior_beta()}},
\code{\link[=prior_cauchy]{prior_cauchy()}},
\code{\link[=prior_const]{prior_const()}},
<<<<<<< HEAD
\code{\link[=prior_const_vector]{prior_const_vector()}},
=======
>>>>>>> main
\code{\link[=prior_gamma]{prior_gamma()}},
\code{\link[=prior_horseshoe]{prior_horseshoe()}},
\code{\link[=prior_init_only]{prior_init_only()}},
\code{\link[=prior_invgamma]{prior_invgamma()}},
\code{\link[=prior_logistic]{prior_logistic()}},
\code{\link[=prior_lognormal]{prior_lognormal()}},
\code{\link[=prior_normal]{prior_normal()}},
\code{\link[=prior_normal_vector]{prior_normal_vector()}},
\code{\link[=prior_std_normal]{prior_std_normal()}},
\code{\link[=prior_student_t]{prior_student_t()}},
\code{\link[=prior_uniform]{prior_uniform()}}
}
\concept{Prior}
