% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_none}
\alias{prior_none}
\title{Only Initial Values Specification}
\usage{
prior_none(init = 1e-05)
}
\arguments{
\item{init}{(\code{number})\cr initial value.}
}
\description{
Only Initial Values Specification
}
