% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_normal_vector}
\alias{prior_normal_vector}
\title{Normal Prior for a Vector Distribution}
\usage{
prior_normal_vector(mus, sigmas)
}
\arguments{
\item{mus}{(\code{numeric}) means.}

\item{sigmas}{(\code{numeric}) standard deviations.}
}
\value{
A \code{Prior} object.
}
\description{
Normal Prior for a Vector Distribution
}
\examples{
prior_normal_vector(c(0, 1), c(1, 2))
}
\seealso{
Other Prior:
\code{\link[=prior_beta]{prior_beta()}},
\code{\link[=prior_cauchy]{prior_cauchy()}},
\code{\link[=prior_const]{prior_const()}},
<<<<<<< HEAD
\code{\link[=prior_const_vector]{prior_const_vector()}},
=======
>>>>>>> main
\code{\link[=prior_gamma]{prior_gamma()}},
\code{\link[=prior_horseshoe]{prior_horseshoe()}},
\code{\link[=prior_init_only]{prior_init_only()}},
\code{\link[=prior_invgamma]{prior_invgamma()}},
\code{\link[=prior_logistic]{prior_logistic()}},
\code{\link[=prior_loglogistic]{prior_loglogistic()}},
\code{\link[=prior_lognormal]{prior_lognormal()}},
\code{\link[=prior_normal]{prior_normal()}},
\code{\link[=prior_std_normal]{prior_std_normal()}},
\code{\link[=prior_student_t]{prior_student_t()}},
\code{\link[=prior_uniform]{prior_uniform()}}
}
\concept{Prior}
