% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_link.R
\name{priors<-,HazardLink-method}
\alias{priors<-,HazardLink-method}
\title{setter method for Hazardlink:}
\usage{
\S4method{priors}{HazardLink}(object) <- value
}
\description{
setter method for Hazardlink:
}
