% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_module.R
\name{priors<-,StanModule-method}
\alias{priors<-,StanModule-method}
\title{setter method for StanModule:}
\usage{
\S4method{priors}{StanModule}(object) <- value
}
\description{
setter method for StanModule:
}
