% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Promise.R
\docType{class}
\name{PromiseLinkComponent-class}
\alias{PromiseLinkComponent-class}
\alias{.PromiseLinkComponent}
\alias{PromiseLinkComponent}
\title{Promise of a \code{LinkComponent}}
\usage{
PromiseLinkComponent(fun, prior, key)
}
\arguments{
\item{fun}{(\code{function}) a function that returns a \code{LinkComponent}. See details.}

\item{prior}{(\code{Prior}) The prior for the scaling coeficient.}

\item{key}{(\code{character}) Link identifier. See Details.}
}
\value{
A \code{PromiseLinkComponent} object.
}
\description{
An object that promises to resolve to a \code{\link{LinkComponent}} object.
Inheriting from \code{\link{Promise}} and \code{\link{LinkComponent}}.
}
\details{
The \code{fun} slot should be a function of signature \verb{function(prior, model)} and should return
a \code{\link{LinkComponent}} object. An error will be thrown if the returned \code{\link{LinkComponent}} object
does not have the same \code{key} slot value as the original \code{PromiseLinkComponent}.
}
\section{Slots}{

\describe{
\item{\code{fun}}{(\code{function}) \cr a function that returns a \code{LinkComponent}. See details.}
}}

\examples{
promise <- PromiseLinkComponent(linkDSLD, prior_normal(0, 2), "link_dsld")
resolvePromise(promise, LongitudinalGSF())
}
