% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Grid.R, R/GridEven.R,
%   R/GridEvent.R, R/GridFixed.R, R/GridGrouped.R, R/GridManual.R,
%   R/GridObserved.R, R/GridPopulation.R, R/GridPrediction.R,
%   R/QuantityGeneratorPopulation.R, R/QuantityGeneratorPrediction.R,
%   R/QuantityGeneratorSubject.R
\docType{class}
\name{as.QuantityGenerator}
\alias{as.QuantityGenerator}
\alias{as.QuantityCollapser}
\alias{Quant-Dev}
\alias{QuantityGenerator-class}
\alias{.QuantityGenerator}
\alias{QuantityCollapser-class}
\alias{.QuantityCollapser}
\alias{QuantityCollapser}
\alias{as.QuantityGenerator.GridEven}
\alias{as.QuantityCollapser.GridEven}
\alias{as.QuantityGenerator.GridEvent}
\alias{as.QuantityCollapser.GridEvent}
\alias{as.QuantityGenerator.GridFixed}
\alias{as.QuantityCollapser.GridFixed}
\alias{as.QuantityGenerator.GridGrouped}
\alias{as.QuantityCollapser.GridGrouped}
\alias{as.QuantityGenerator.GridManual}
\alias{as.QuantityCollapser.GridManual}
\alias{as.QuantityGenerator.GridObserved}
\alias{as.QuantityCollapser.GridObserved}
\alias{as.QuantityGenerator.GridPopulation}
\alias{as.QuantityCollapser.GridPopulation}
\alias{as.QuantityGenerator.GridPrediction}
\alias{as.QuantityCollapser.GridPrediction}
\alias{QuantityGeneratorPopulation-class}
\alias{.QuantityGeneratorPopulation}
\alias{QuantityGeneratorPopulation}
\alias{QuantityGeneratorPrediction-class}
\alias{.QuantityGeneratorPrediction}
\alias{QuantityGeneratorPrediction}
\alias{QuantityGeneratorSubject-class}
\alias{.QuantityGeneratorSubject}
\alias{QuantityGeneratorSubject}
\title{Quantity Developer Notes}
\usage{
as.QuantityGenerator(object, ...)

as.QuantityCollapser(object, ...)

QuantityCollapser(times, groups, indexes)

\method{as.QuantityGenerator}{GridEven}(object, data, ...)

\method{as.QuantityCollapser}{GridEven}(object, data, ...)

\method{as.QuantityGenerator}{GridEvent}(object, data, ...)

\method{as.QuantityCollapser}{GridEvent}(object, data, ...)

\method{as.QuantityGenerator}{GridFixed}(object, data, ...)

\method{as.QuantityCollapser}{GridFixed}(object, data, ...)

\method{as.QuantityGenerator}{GridGrouped}(object, data, ...)

\method{as.QuantityCollapser}{GridGrouped}(object, data, ...)

\method{as.QuantityGenerator}{GridManual}(object, data, ...)

\method{as.QuantityCollapser}{GridManual}(object, data, ...)

\method{as.QuantityGenerator}{GridObserved}(object, data, ...)

\method{as.QuantityCollapser}{GridObserved}(object, data, ...)

\method{as.QuantityGenerator}{GridPopulation}(object, data, model = NULL, ...)

\method{as.QuantityCollapser}{GridPopulation}(object, data, model = NULL, ...)

\method{as.QuantityGenerator}{GridPrediction}(object, data, ...)

\method{as.QuantityCollapser}{GridPrediction}(object, data, ...)

QuantityGeneratorPopulation(times, studies = NULL, arms = NULL, newdata = NULL)

QuantityGeneratorPrediction(times, newdata = NULL, params = NULL)

QuantityGeneratorSubject(times, subjects = NULL)
}
\arguments{
\item{object}{(\code{Grid}) object to convert to a \code{QuantityGenerator} or \code{QuantityCollapser}.}

\item{...}{Not currently used.}

\item{times}{(\code{numeric}) vector of time points to extract quantities at.}

\item{groups}{(\code{character}) vector of labels to apply to the generated quantities.}

\item{indexes}{(\code{list}) list of indexes that specify which observations from a
\code{QuantityGenerator} should be combined to form the desired quantities.}

\item{data}{(\code{DataJoint}) Survival and Longitudinal Data.}

\item{model}{(\code{JointModel} or \code{NULL}) optional fitted-model specification used for model-specific population
profiles.}

\item{newdata}{(\code{data.frame} or \code{NULL}) population profiles carried by a population quantity generator.}

\item{subjects}{(\code{character}) vector of subjects to extract quantities for.}
}
\value{
A \code{QuantityGenerator} or \code{QuantityCollapser} object, according to the function called.
}
\description{
Developer details for \code{QuantityX} objects/methods. This page just outlines the arguments
and slots of these objects/methods. For the full implementation details please see \link{Grid-Dev}
}
\details{
The \code{as.QuantityGenerator} must return a \code{QuantityGenerator} object.
The \code{as.QuantityCollapser} must return a \code{QuantityCollapser} object.
}
\section{Slots}{

\describe{
\item{\code{times}}{(\code{numeric})\cr See Arguments for details.}

\item{\code{subjects}}{(\code{character})\cr See Arguments for details.}

\item{\code{groups}}{(\code{character})\cr See Arguments for details.}

\item{\code{indexes}}{(\code{list})\cr See Arguments for details.}
}}

\examples{
\dontrun{
grid <- GridFixed(times = c(0, 10))
as.QuantityGenerator(grid, joint_data)
as.QuantityCollapser(grid, joint_data)
}
}
\keyword{internal}
