% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\name{Quantities-Shared}
\alias{Quantities-Shared}
\title{Re-used documentation for \code{Quantities}}
\arguments{
\item{x}{(\code{Quantities}) generated quantities.}

\item{object}{(\code{Quantities}) generated quantities.}

\item{type}{(\code{character}) sets the \code{type} variable.}

\item{conf.level}{(\code{numeric}) confidence level of the interval.}

\item{...}{not used.}
}
\description{
Re-used documentation for \code{Quantities}
}
\keyword{internal}
