% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalRandomEffects.R
\docType{class}
\name{RandomEffectQuantities-class}
\alias{RandomEffectQuantities-class}
\alias{.RandomEffectQuantities}
\alias{RandomEffectQuantities}
\title{Random Effects Quantities Container}
\usage{
RandomEffectQuantities(quantities, subject, parameter)
}
\arguments{
\item{quantities}{(\code{matrix}) of random effects values.}

\item{subject}{(\code{character}) labels specifying which subjects the values belong to.}

\item{parameter}{(\code{character}) labels specifying which parameter the value is.}
}
\value{
A \code{RandomEffectQuantities} object.
}
\description{
A simple wrapper around a \code{matrix} to store required metadata for patient level
random effects data
}
\details{
Each row of the matrix represents a sample and each column represents a distinct subject
specific parameter.
As such the number of columns in the matrix should equal the length of \code{subject} and \code{parameter}
which provide metadata for who the parameter corresponds to as well as which parameter it is.
}
\section{Slots}{

\describe{
\item{\code{quantities}}{(\code{matrix})\cr See Arguments for details.}

\item{\code{subject}}{(\code{numeric})\cr See Arguments for details.}

\item{\code{parameter}}{(\code{character})\cr See Arguments for details.}
}}

\seealso{
Other RandomEffectQuantities:
\code{\link[=LongitudinalRandomEffects]{LongitudinalRandomEffects()}},
\code{\link[=as.data.frame.RandomEffectQuantities]{as.data.frame.RandomEffectQuantities()}},
\code{\link[=as_print_string.RandomEffectQuantities]{as_print_string.RandomEffectQuantities()}},
\code{\link[=summary.RandomEffectQuantities]{summary.RandomEffectQuantities()}}
}
\concept{RandomEffectQuantities}
\keyword{internal}
