% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalRandomEffects.R
\name{RandomEffectQuantities-Shared}
\alias{RandomEffectQuantities-Shared}
\title{Re-used documentation for \code{RandomEffectQuantities}}
\arguments{
\item{x}{(\code{\link{RandomEffectQuantities}}) \cr generated quantities.}

\item{object}{(\code{\link{RandomEffectQuantities}}) \cr generated quantities.}

\item{conf.level}{(\code{numeric}) \cr confidence level of the interval.}

\item{...}{not used.}
}
\description{
Re-used documentation for \code{RandomEffectQuantities}
}
\keyword{internal}
