% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_module.R
\name{read_file}
\alias{read_file}
\title{Read entire file as a single string}
\usage{
read_file(filename)
}
\arguments{
\item{filename}{Location of file to read in}
}
\description{
Simple utility function to read in a file as 1
continous string
}
