% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_module.R
\name{remove_blank_strings}
\alias{remove_blank_strings}
\title{Removes blank strings from a string vector}
\usage{
remove_blank_strings(x)
}
\arguments{
\item{x}{a vector of string}
}
\description{
Function removes blank strings from a string vector
If all strings are blank then it will just return a single blank
}
