% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{render_stan_limits.numeric}
\alias{render_stan_limits.numeric}
\title{describeIn render_stan_limits method for numeric vectors.}
\usage{
\method{render_stan_limits}{numeric}(object)
}
\description{
describeIn render_stan_limits method for numeric vectors.
}
