% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{render_stan_limits.Prior}
\alias{render_stan_limits.Prior}
\title{describeIn render_stan_limits method for \link{Prior} objects.}
\usage{
\method{render_stan_limits}{Prior}(object)
}
\description{
describeIn render_stan_limits method for \link{Prior} objects.
}
