% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{render_stan_parameter_declaration}
\alias{render_stan_parameter_declaration}
\title{Render Stan Parameter Declaration}
\usage{
render_stan_parameter_declaration(name, size, limits)
}
\arguments{
\item{name}{(\code{string}) parameter name.}

\item{size}{(\code{numeric_OR_character}) parameter size.}

\item{limits}{(\code{numeric}) lower and upper parameter limits.}
}
\value{
A length-one character vector.
}
\description{
Creates a Stan declaration for a sampled scalar or vector parameter.
}
\keyword{internal}
