% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{render_stan_parameter_limits}
\alias{render_stan_parameter_limits}
\title{Render Stan Parameter Constraints}
\usage{
render_stan_parameter_limits(limits)
}
\arguments{
\item{limits}{(\code{numeric}) length-two vector containing lower and upper parameter limits.}
}
\value{
A length-one character vector.
}
\description{
Converts a lower/upper limit vector from a \code{\link{Prior}} into a Stan declaration
constraint, e.g. \verb{<lower=0.1, upper=1>}.
}
\keyword{internal}
