% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{replace_with_lookup}
\alias{replace_with_lookup}
\title{Replace Character Size by Looked Up Numbers}
\usage{
replace_with_lookup(sizes, data)
}
\arguments{
\item{sizes}{(\code{list})\cr may include character elements that correspond to
names in the data list.}

\item{data}{(\code{list})\cr data containing numeric values.}
}
\value{
A list of sizes with character elements in \code{sizes}
replaced by their corresponding numeric values in \code{data}.
}
\description{
Replace Character Size by Looked Up Numbers
}
\note{
Each element in the final list of sizes must be a single number.
}
\keyword{internal}
