% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LongitudinalClaretBrunoCov.R,
%   R/LongitudinalGSFCov.R, R/LongitudinalRandomSlopeCov.R,
%   R/LongitudinalSteinFojoCov.R
\name{required_longitudinal_covs}
\alias{required_longitudinal_covs}
\alias{required_longitudinal_covs.default}
\alias{required_longitudinal_covs.LongitudinalClaretBrunoCov}
\alias{required_longitudinal_covs.LongitudinalGSFCov}
\alias{required_longitudinal_covs.LongitudinalRandomSlopeCov}
\alias{required_longitudinal_covs.LongitudinalSteinFojoCov}
\title{Required Longitudinal Covariates}
\usage{
required_longitudinal_covs(object, ...)

\method{required_longitudinal_covs}{default}(object, ...)

\method{required_longitudinal_covs}{LongitudinalClaretBrunoCov}(object, ...)

\method{required_longitudinal_covs}{LongitudinalGSFCov}(object, ...)

\method{required_longitudinal_covs}{LongitudinalRandomSlopeCov}(object, ...)

\method{required_longitudinal_covs}{LongitudinalSteinFojoCov}(object, ...)
}
\arguments{
\item{object}{(\code{LongitudinalModel}) A longitudinal model object.}

\item{...}{Not used.}
}
\value{
A character vector of covariate names.
}
\description{
Return the subject-level covariates required to generate longitudinal
population quantities for a model.
}
