% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{resolvePromise}
\alias{resolvePromise}
\alias{resolvePromise.default}
\title{Resolve a Promise}
\usage{
resolvePromise(object, ...)

\method{resolvePromise}{default}(object, ...)
}
\arguments{
\item{object}{(\code{ANY}) an object to resolve.}

\item{...}{(\code{ANY}) additional arguments.

If \code{object} is not a promise will just return itself else will resolve the promise
and return the promised object.}
}
\value{
The input object, or the value obtained by resolving a promise.
}
\description{
Resolve a Promise
}
\examples{
resolvePromise(1)
resolvePromise(linkDSLD(), LongitudinalGSF())
}
