% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_score.R
\name{reverse_km}
\alias{reverse_km}
\alias{reverse_km_event_first}
\alias{reverse_km_cen_first}
\title{Reverse Kaplan-Meier}
\usage{
reverse_km_event_first(t, times, events)

reverse_km_cen_first(t, times, events)
}
\arguments{
\item{t}{(\code{numeric}) timepoints to calculate the desired quantity at.}

\item{times}{(\code{numeric}) observed times.}

\item{events}{(\code{numeric}) event indicator for \code{times}. Either 1 for an event or 0 for censor.}
}
\value{
The reverse Kaplan-Meier estimate at each requested time.
}
\description{
Calculates the survival estimates of the censoring distribution
using the Kaplan-Meier estimate.
This is primarily used in the calculation of the IPCW estimates.
}
\details{
With regards to ties between censor and event times; the standard
approach is to regard events as occurring before censors. However,
when modelling the censoring distribution we are regarding the
censors as "events" so which should come first in the case of ties?

The \code{reverse_km_event_first()} function maintains the rule
that events always come first even if we are regarding the censors
as "events". This matches the implementation of
\code{prodlim::prodlim(..., reverse = TRUE)}.

The \code{reverse_km_cen_first()} function provides the alternative
implementation assuming that in the case of ties the censor "events"
come before the event "censors". This is essentially a thin wrapper
around \code{survival::survfit(Surv(time, 1 - event), ...)}
}
\keyword{internal}
