% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/SimLongitudinalClaretBruno.R,
%   R/SimLongitudinalClaretBrunoCov.R, R/SimLongitudinalGSF.R,
%   R/SimLongitudinalGSFCov.R, R/SimLongitudinalRandomSlope.R,
%   R/SimLongitudinalRandomSlopeCov.R, R/SimLongitudinalSteinFojo.R,
%   R/SimLongitudinalSteinFojoCov.R, R/SimSurvival.R
\name{sampleObservations}
\alias{sampleObservations}
\alias{sampleObservations.SimLongitudinalClaretBruno}
\alias{sampleObservations.SimLongitudinalClaretBrunoCov}
\alias{sampleObservations.SimLongitudinalGSF}
\alias{sampleObservations.SimLongitudinalGSFCov}
\alias{sampleObservations.SimLongitudinalRandomSlope}
\alias{sampleObservations.SimLongitudinalRandomSlopeCov}
\alias{sampleObservations.SimLongitudinalSteinFojo}
\alias{sampleObservations.SimLongitudinalSteinFojoCov}
\alias{sampleObservations.SimSurvival}
\title{Generate Simulated Observations}
\usage{
sampleObservations(object, times_df)

\method{sampleObservations}{SimLongitudinalClaretBruno}(object, times_df)

\method{sampleObservations}{SimLongitudinalClaretBrunoCov}(object, times_df)

\method{sampleObservations}{SimLongitudinalGSF}(object, times_df)

\method{sampleObservations}{SimLongitudinalGSFCov}(object, times_df)

\method{sampleObservations}{SimLongitudinalRandomSlope}(object, times_df)

\method{sampleObservations}{SimLongitudinalRandomSlopeCov}(object, times_df)

\method{sampleObservations}{SimLongitudinalSteinFojo}(object, times_df)

\method{sampleObservations}{SimLongitudinalSteinFojoCov}(object, times_df)

\method{sampleObservations}{SimSurvival}(object, times_df)
}
\arguments{
\item{object}{(\code{SimLongitudinal} or \code{SimSurvival}) object to generate observations from.}

\item{times_df}{(\code{data.frame}) the times at which to generate observations. See details.}
}
\value{
A \code{data.frame} containing simulated observations.
}
\description{
Generate Simulated Observations
}
\details{
The \code{times_df} argument should be a \code{data.frame} as created by \code{sampleSubjects} but
replicated for each time point at which observations are to be generated. That is if you want
to generate observations for times \code{c(0, 1, 2, 3)} then \code{times_df} should be created as:

\if{html}{\out{<div class="sourceCode">}}\preformatted{subject_dat <- sampleSubjects(object, ...)
times_df <- tidyr::expand_grid(
    subject_dat,
    time = c(0, 1, 2, 3)
  )
}\if{html}{\out{</div>}}
}
\examples{
sim <- SimLongitudinalRandomSlope(slope_mu = 0.01, slope_sigma = 0.5)
subjects <- data.frame(study = factor("S"), arm = factor("A"), subject = "1")
subject_parameters <- sampleSubjects(sim, subjects)
sampleObservations(sim, transform(subject_parameters, time = 0))
}
