% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModel.R
\name{sampleStanModel}
\alias{sampleStanModel}
\alias{sampleStanModel.JointModel}
\title{\code{sampleStanModel}}
\usage{
sampleStanModel(object, ...)

\method{sampleStanModel}{JointModel}(object, data, ...)
}
\arguments{
\item{object}{the module.}

\item{...}{additional arguments.}

\item{data}{(\code{DataJoint} or \code{list}) input data.}
}
\value{
A \code{JointModelSamples} object.
}
\description{
Sample from a Stan Module.
}
\examples{
\dontrun{
sampleStanModel(JointModel(LongitudinalGSF()), data = joint_data)
}
}
