% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{saveObject}
\alias{saveObject}
\title{Save Object to File}
\usage{
saveObject(object, file, ...)
}
\arguments{
\item{object}{(\code{ANY}) object to save.}

\item{file}{(\code{character}) file to save object to.}

\item{...}{(\code{ANY}) additional arguments.}
}
\value{
Invisibly returns \code{NULL} after saving the object.
}
\description{
Save Object to File
}
\examples{
\dontrun{
saveObject(fit, "joint-model-fit.rds")
}
}
\seealso{
Other saveObject:
\code{\link[=saveObject.JointModelSamples]{saveObject.JointModelSamples()}}
}
\concept{saveObject}
