% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalSteinFojo.R
\name{sf_sld}
\alias{sf_sld}
\alias{sf_ttg}
\alias{sf_dsld}
\title{Stein-Fojo Functionals}
\usage{
sf_sld(time, b, s, g)

sf_ttg(time, b, s, g)

sf_dsld(time, b, s, g)
}
\arguments{
\item{time}{(\code{numeric}) time grid.}

\item{b}{(\code{number}) baseline.}

\item{s}{(\code{number}) shrinkage.}

\item{g}{(\code{number}) growth.}
}
\value{
The function results.
}
\description{
Stein-Fojo Functionals
}
\keyword{internal}
