% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{show}
\alias{show}
\title{Printing of Different Classes}
\arguments{
\item{object}{what to print.}
}
\description{
These methods print objects of different classes.
}
\examples{
show(prior_normal(0, 1))
}
