% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sf.R
\name{sim_lm_sf}
\alias{sim_lm_sf}
\title{Construct a Simulation Function for Longitudinal Data from SF Model}
\usage{
sim_lm_sf(
  sigma = 0.01,
  mu_s = c(3, 4),
  mu_g = c(0.2, 0.3),
  mu_b = 50,
  omega_b = 0.135,
  omega_s = 0.15,
  omega_g = 0.225,
  link_dsld = 0,
  link_ttg = 0,
  .debug = FALSE
)
}
\arguments{
\item{sigma}{(\code{number})\cr the variance of the longitudinal values.}

\item{mu_s}{(\code{numeric})\cr the mean shrinkage rates for the two treatment arms.}

\item{mu_g}{(\code{numeric})\cr the mean growth rates for the two treatment arms.}

\item{mu_b}{(\code{numeric})\cr the mean baseline values for the two treatment arms.}

\item{omega_b}{(\code{number})\cr the baseline value standard deviation.}

\item{omega_s}{(\code{number})\cr the shrinkage rate standard deviation.}

\item{omega_g}{(\code{number})\cr the growth rate standard deviation.}

\item{link_dsld}{(\code{number})\cr the link coefficient for the derivative contribution.}

\item{link_ttg}{(\code{number})\cr the link coefficient for the time-to-growth contribution.}

\item{.debug}{(\code{flag})\cr whether to enter debug mode such that the function
would only return a subset of columns.}
}
\value{
A function with argument \code{lm_base} that can be used to simulate
longitudinal data from the corresponding SF model.
}
\description{
Construct a Simulation Function for Longitudinal Data from SF Model
}
