% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_os.R
\name{sim_os_exponential}
\alias{sim_os_exponential}
\title{Construct a Log Hazard Function for the Exponential Model}
\usage{
sim_os_exponential(lambda)
}
\arguments{
\item{lambda}{(\code{number})\cr the rate parameter.}
}
\value{
A function of \code{time} returning the log hazard.
}
\description{
Construct a Log Hazard Function for the Exponential Model
}
