% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_os_weibull}
\alias{sim_os_weibull}
\title{Construct a Log Hazard Function for the Weibull Model}
\usage{
sim_os_weibull(lambda, gamma)
}
\arguments{
\item{lambda}{(\code{number})\cr the scale parameter.}

\item{gamma}{(\code{number})\cr the shape parameter.}
}
\value{
A function of \code{time} returning the log hazard.
}
\description{
Construct a Log Hazard Function for the Weibull Model
}
