% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimGroup.R
\docType{class}
\name{SimGroup-class}
\alias{SimGroup-class}
\alias{.SimGroup}
\alias{SimGroup}
\title{Define Simulation Group}
\usage{
SimGroup(n, arm, study)
}
\arguments{
\item{n}{(\code{numeric}) number of subjects in the group.}

\item{arm}{(\code{character}) treatment arm.}

\item{study}{(\code{character}) study name.}
}
\value{
A \code{SimGroup} object.
}
\description{
Specifies a simulation group to be used by \code{\link[=SimJointData]{SimJointData()}}.
}
\section{Slots}{

\describe{
\item{\code{n}}{(\code{numeric})\cr See arguments.}

\item{\code{arm}}{(\code{character})\cr See arguments.}

\item{\code{study}}{(\code{character})\cr See arguments.}
}}

\examples{
SimGroup(n = 50, arm = "Arm-A", study = "Study-1")
}
