% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimJointData.R
\docType{class}
\name{SimJointData-class}
\alias{SimJointData-class}
\alias{.SimJointData}
\alias{SimJointData}
\title{Simulating Joint Longitudinal and Time-to-Event Data}
\usage{
SimJointData(
  design = list(SimGroup(n = 50, study = "Study-1", arm = "Arm-A"), SimGroup(n = 50,
    study = "Study-1", arm = "Arm-B")),
  longitudinal,
  survival,
  .silent = FALSE
)
}
\arguments{
\item{design}{(\code{list}) a list of \code{\link{SimGroup}} objects. See details.}

\item{longitudinal}{(\code{SimLongitudinal}) object specifying how to simulate the longitudinal data}

\item{survival}{(\code{SimSurvival}) object specifying how to simulate the survival data}

\item{.silent}{(\code{flag}) whether to suppress info messages}
}
\value{
A \code{SimJointData} object.
}
\description{
Simulating Joint Longitudinal and Time-to-Event Data
}
\details{
The \code{design} argument is used to specify how many distinct groups should be simulated
including key information such as the number of subjects within the group as well as
which treatment arm and study the group belongs to. The \code{design} argument should be a
list of \code{\link{SimGroup}} objects e.g.

\if{html}{\out{<div class="sourceCode">}}\preformatted{design = list(
    SimGroup(n = 50, study = "Study-1", arm = "Arm-A"),
    SimGroup(n = 50, study = "Study-1", arm = "Arm-B")
  )
}\if{html}{\out{</div>}}
}
\section{Slots}{

\describe{
\item{\code{longitudinal}}{(\code{data.frame})\cr the simulated longitudinal data.}

\item{\code{survival}}{(\code{data.frame})\cr the simulated survival data.}
}}

\examples{
set.seed(1)
SimJointData(
  design = list(SimGroup(3, "A", "Study 1")),
  longitudinal = SimLongitudinalRandomSlope(
    times = c(0, 10),
    slope_mu = 0.01,
    slope_sigma = 0.5
  ),
  survival = SimSurvivalExponential(lambda = 1 / 300)
)
}
