% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimJointData.R
\name{SimJointData}
\alias{SimJointData}
\title{Simulating Joint Longitudinal and Time-to-Event Data}
\usage{
SimJointData(
  design = list(SimGroup(n = 50, study = "Study-1", arm = "Arm-A"), SimGroup(n = 50,
    study = "Study-1", arm = "Arm-B")),
  longitudinal,
  survival,
  .silent = FALSE,
  .debug = FALSE
)
}
\arguments{
\item{design}{(\code{list})\cr a list of \code{\link{SimGroup}} objects. See details.}

\item{longitudinal}{(\code{function})\cr function of \code{lm_base} generating the longitudinal model outcomes.}

\item{survival}{(\code{function})\cr function of \code{lm_base} generating the survival model outcomes.}

\item{.silent}{(\code{flag})\cr whether to suppress info messages}

\item{.debug}{(\code{flag})\cr whether to enter debug mode such that the function
would only return a subset of columns.}

\item{times}{(\code{numeric})\cr time grid, e.g. specifying the days after randomization.}

\item{lambda_cen}{(\code{number})\cr rate of the exponential censoring distribution.}

\item{beta_cont}{(\code{number})\cr coefficient for the continuous covariate.}

\item{beta_cat}{(\code{numeric})\cr coefficients for the categorical covariate levels.}
}
\value{
List with simulated \code{lm} (longitudinal) and \code{os} (survival) data sets.
}
\description{
Simulating Joint Longitudinal and Time-to-Event Data
}
\details{
The \code{design} argument is used to specify how many distinct groups should be simulated
including key information such as the number of subjects within the group as well as
which treatment arm and study the group belongs to. The \code{design} argument should be a
list of \code{\link{SimGroup}} objects e.g.

\if{html}{\out{<div class="sourceCode">}}\preformatted{design = list(
    SimGroup(n = 50, study = "Study-1", arm = "Arm-A"),
    SimGroup(n = 50, study = "Study-1", arm = "Arm-B")
)
}\if{html}{\out{</div>}}
}
