% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalClaretBruno.R
\docType{class}
\name{SimLongitudinalClaretBruno-class}
\alias{SimLongitudinalClaretBruno-class}
\alias{.SimLongitudinalClaretBruno}
\alias{SimLongitudinalClaretBruno}
\title{Simulate Longitudinal Data from a Claret-Bruno Model}
\usage{
SimLongitudinalClaretBruno(
  times = c(-100, -50, 0, 50, 100, 150, 250, 350, 450, 550)/365,
  sigma = 0.01,
  mu_b = log(60),
  mu_g = log(c(0.9, 1.1)),
  mu_c = log(c(0.25, 0.35)),
  mu_p = log(c(1.5, 2)),
  omega_b = 0.2,
  omega_g = 0.2,
  omega_c = 0.2,
  omega_p = 0.2,
  link_dsld = 0,
  link_ttg = 0,
  link_identity = 0,
  link_growth = 0,
  scaled_variance = FALSE
)
}
\arguments{
\item{times}{(\code{numeric}) the times to generate observations at.}

\item{sigma}{(\code{number}) the variance of the longitudinal values.}

\item{mu_b}{(\code{numeric}) the mean population baseline sld value.}

\item{mu_g}{(\code{numeric}) the mean population growth rate.}

\item{mu_c}{(\code{numeric}) the mean population resistance rate.}

\item{mu_p}{(\code{numeric}) the mean population growth inhibition.}

\item{omega_b}{(\code{number}) the population standard deviation for the baseline sld value.}

\item{omega_g}{(\code{number}) the population standard deviation for the growth rate.}

\item{omega_c}{(\code{number}) the population standard deviation for the resistance rate.}

\item{omega_p}{(\code{number}) the population standard deviation for the growth inhibition.}

\item{link_dsld}{(\code{number}) the link coefficient for the derivative contribution.}

\item{link_ttg}{(\code{number}) the link coefficient for the time-to-growth contribution.}

\item{link_identity}{(\code{number}) the link coefficient for the SLD Identity contribution.}

\item{link_growth}{(\code{number}) the link coefficient for the growth parameter contribution.}

\item{scaled_variance}{(\code{logical}) whether the variance should be scaled by the expected value
(see the "Statistical Specifications" vignette for more details)}
}
\value{
A \code{SimLongitudinalClaretBruno} object.
}
\description{
Simulate Longitudinal Data from a Claret-Bruno Model
}
\section{Slots}{

\describe{
\item{\code{sigma}}{(\code{numeric})\cr See arguments.}

\item{\code{mu_b}}{(\code{numeric})\cr See arguments.}

\item{\code{mu_g}}{(\code{numeric})\cr See arguments.}

\item{\code{mu_c}}{(\code{numeric})\cr See arguments.}

\item{\code{mu_p}}{(\code{numeric})\cr See arguments.}

\item{\code{omega_b}}{(\code{numeric})\cr See arguments.}

\item{\code{omega_g}}{(\code{numeric})\cr See arguments.}

\item{\code{omega_c}}{(\code{numeric})\cr See arguments.}

\item{\code{omega_p}}{(\code{numeric})\cr See arguments.}

\item{\code{link_dsld}}{(\code{numeric})\cr See arguments.}

\item{\code{link_ttg}}{(\code{numeric})\cr See arguments.}

\item{\code{link_identity}}{(\code{numeric})\cr See arguments.}

\item{\code{link_growth}}{(\code{numeric})\cr See arguments.}

\item{\code{scaled_variance}}{(\code{logical})\cr See arguments.}
}}

\examples{
SimLongitudinalClaretBruno(times = c(0, 30, 60))
}
\seealso{
Other SimLongitudinal:
\code{\link{SimLongitudinal-class}},
\code{\link{SimLongitudinalGSF-class}},
\code{\link{SimLongitudinalRandomSlope-class}},
\code{\link{SimLongitudinalSteinFojo-class}}
}
\concept{SimLongitudinal}
