% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalGSF.R
\docType{class}
\name{SimLongitudinalGSF-class}
\alias{SimLongitudinalGSF-class}
\alias{.SimLongitudinalGSF}
\alias{SimLongitudinalGSF}
\title{Simulate Longitudinal Data from a GSF Model}
\usage{
SimLongitudinalGSF(
  times = c(-100, -50, 0, 50, 100, 150, 250, 350, 450, 550)/365,
  sigma = 0.01,
  mu_s = log(c(0.6, 0.4)),
  mu_g = log(c(0.25, 0.35)),
  mu_b = log(60),
  mu_phi = qlogis(c(0.4, 0.6)),
  omega_b = 0.2,
  omega_s = 0.2,
  omega_g = 0.2,
  omega_phi = 0.2,
  link_dsld = 0,
  link_ttg = 0,
  link_identity = 0,
  link_growth = 0,
  link_shrinkage = 0,
  scaled_variance = FALSE
)
}
\arguments{
\item{times}{(\code{numeric}) the times to generate observations at.}

\item{sigma}{(\code{number}) the variance of the longitudinal values.}

\item{mu_s}{(\code{numeric}) the mean shrinkage rates.}

\item{mu_g}{(\code{numeric}) the mean growth rates.}

\item{mu_b}{(\code{numeric}) the mean baseline values.}

\item{mu_phi}{(\code{numeric}) the logit-scale location of the proportion of cells affected by the
treatment.}

\item{omega_b}{(\code{number}) the baseline value standard deviation.}

\item{omega_s}{(\code{number}) the shrinkage rate standard deviation.}

\item{omega_g}{(\code{number}) the growth rate standard deviation.}

\item{omega_phi}{(\code{number}) the logit-scale standard deviation of the proportion of cells affected by
the treatment.}

\item{link_dsld}{(\code{number}) the link coefficient for the derivative contribution.}

\item{link_ttg}{(\code{number}) the link coefficient for the time-to-growth contribution.}

\item{link_identity}{(\code{number}) the link coefficient for the SLD Identity contribution.}

\item{link_growth}{(\code{number}) the link coefficient for the log-growth parameter contribution.}

\item{link_shrinkage}{(\code{number}) the link coefficient for the log-shrinkage parameter contribution.}

\item{scaled_variance}{(\code{logical}) whether the variance should be scaled by the expected value
(see the "Statistical Specifications" vignette for more details)}
}
\value{
A \code{SimLongitudinalGSF} object.
}
\description{
Simulate Longitudinal Data from a GSF Model
}
\section{Slots}{

\describe{
\item{\code{sigma}}{(\code{numeric})\cr See arguments.}

\item{\code{mu_s}}{(\code{numeric})\cr See arguments.}

\item{\code{mu_g}}{(\code{numeric})\cr See arguments.}

\item{\code{mu_b}}{(\code{numeric})\cr See arguments.}

\item{\code{mu_phi}}{(\code{numeric})\cr See arguments.}

\item{\code{omega_b}}{(\code{numeric})\cr See arguments.}

\item{\code{omega_s}}{(\code{numeric})\cr See arguments.}

\item{\code{omega_g}}{(\code{numeric})\cr See arguments.}

\item{\code{omega_phi}}{(\code{numeric})\cr See arguments.}

\item{\code{link_dsld}}{(\code{numeric})\cr See arguments.}

\item{\code{link_ttg}}{(\code{numeric})\cr See arguments.}

\item{\code{link_identity}}{(\code{numeric})\cr See arguments.}

\item{\code{link_growth}}{(\code{numeric})\cr See arguments.}

\item{\code{link_shrinkage}}{(\code{numeric})\cr See arguments.}

\item{\code{scaled_variance}}{(\code{logical})\cr See arguments.}
}}

\examples{
SimLongitudinalGSF(times = c(0, 30, 60))
}
\seealso{
Other SimLongitudinal:
\code{\link{SimLongitudinal-class}},
\code{\link{SimLongitudinalClaretBruno-class}},
\code{\link{SimLongitudinalRandomSlope-class}},
\code{\link{SimLongitudinalSteinFojo-class}}
}
\concept{SimLongitudinal}
