% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalRandomSlope.R
\docType{class}
\name{SimLongitudinalRandomSlope-class}
\alias{SimLongitudinalRandomSlope-class}
\alias{.SimLongitudinalRandomSlope}
\alias{SimLongitudinalRandomSlope}
\title{Simulate Longitudinal Data from a Random Slope Model}
\usage{
SimLongitudinalRandomSlope(
  times = c(-100, -50, 0, 50, 100, 150, 250, 350, 450, 550),
  intercept = 50,
  slope_mu = c(0.01, 0.03),
  slope_sigma = c(0.5, 0.6),
  sigma = 2,
  link_dsld = 0,
  link_identity = 0,
  scaled_variance = FALSE
)
}
\arguments{
\item{times}{(\code{numeric}) the times to generate observations at.}

\item{intercept}{(\code{numeric}) the mean baseline value for each study.}

\item{slope_mu}{(\code{numeric}) the population slope for each treatment arm.}

\item{slope_sigma}{(\code{numeric}) the random slope standard deviation for each treatment arm.}

\item{sigma}{(\code{number}) the variance of the longitudinal values.}

\item{link_dsld}{(\code{number}) the link coefficient for the DSLD contribution.}

\item{link_identity}{(\code{number}) the link coefficient for the identity contribution.}

\item{scaled_variance}{(\code{logical}) whether the variance should be scaled by the expected value
(see the "Statistical Specifications" vignette for more details)}
}
\value{
A \code{SimLongitudinalRandomSlope} object.
}
\description{
Simulate Longitudinal Data from a Random Slope Model
}
\section{Slots}{

\describe{
\item{\code{intercept}}{(\code{numeric})\cr See arguments.}

\item{\code{slope_mu}}{(\code{numeric})\cr See arguments.}

\item{\code{slope_sigma}}{(\code{numeric})\cr See arguments.}

\item{\code{sigma}}{(\code{numeric})\cr See arguments.}

\item{\code{link_dsld}}{(\code{numeric})\cr See arguments.}

\item{\code{link_identity}}{(\code{numeric})\cr See arguments.}

\item{\code{scaled_variance}}{(\code{logical})\cr See arguments.}
}}

\examples{
SimLongitudinalRandomSlope(times = c(0, 30, 60))
}
\seealso{
Other SimLongitudinal:
\code{\link{SimLongitudinal-class}},
\code{\link{SimLongitudinalClaretBruno-class}},
\code{\link{SimLongitudinalGSF-class}},
\code{\link{SimLongitudinalSteinFojo-class}}
}
\concept{SimLongitudinal}
