% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSurvival.R
\docType{class}
\name{SimSurvival-class}
\alias{SimSurvival-class}
\alias{.SimSurvival}
\alias{SimSurvival}
\title{Abstract Simulation Class for Survival Data}
\usage{
SimSurvival(
  time_max = 2000,
  time_step = 1,
  lambda_censor = 1/3000,
  beta_cont = 0.2,
  beta_cat = c(A = 0, B = -0.4, C = 0.2),
  loghazard,
  name = "SimSurvival"
)
}
\arguments{
\item{time_max}{(\code{number}) the maximum time to simulate to.}

\item{time_step}{(\code{number}) the time interval between evaluating the log-hazard function.}

\item{lambda_censor}{(\code{number}) the censoring rate.}

\item{beta_cont}{(\code{number}) the continuous covariate coefficient.}

\item{beta_cat}{(\code{numeric}) the categorical covariate coefficients.}

\item{loghazard}{(\code{function}) the log hazard function.}

\item{name}{(\code{character}) the name of the object.}
}
\value{
A \code{SimSurvival} object.
}
\description{
Abstract Simulation Class for Survival Data
}
\section{Slots}{

\describe{
\item{\code{time_max}}{(\code{numeric})\cr See arguments.}

\item{\code{time_step}}{(\code{numeric})\cr See arguments.}

\item{\code{lambda_censor}}{(\code{numeric})\cr See arguments.}

\item{\code{beta_cont}}{(\code{numeric})\cr See arguments.}

\item{\code{beta_cat}}{(\code{numeric})\cr See arguments.}

\item{\code{loghazard}}{(\code{function})\cr See arguments.}

\item{\code{name}}{(\code{character})\cr See arguments.}

\item{\code{beta_os_cov}}{(\code{numeric}) Additional regression coefficients for survival models.}
}}

\section{Hazard Evaluation}{


Event times are simulated by sampling a cumulative hazard limit from a \eqn{U(0, 1)} distribution
for
each subject and then counting how much hazard they've been exposed to by evaluating the
log-hazard function at a set interval. The \code{time_max} argument sets the upper bound for the
number of time points to evaluate the log-hazard function at with subjects who have not had an
event being censored at \code{time_max}. The \code{time_step} argument sets the interval at which to
evaluate the log-hazard function. Setting smaller values for \code{time_step} will increase the
precision of the simulation at the cost of increased computation time. Likewise, setting large
values for \code{time_max} will minimize the number of censored subjects at the cost of
increased computation time.
}

\examples{
SimSurvival(loghazard = function(time, ...) rep(log(1 / 100), length(time)))
}
\seealso{
Other SimSurvival:
\code{\link[=SimSurvivalExponential]{SimSurvivalExponential()}},
\code{\link[=SimSurvivalGamma]{SimSurvivalGamma()}},
\code{\link[=SimSurvivalLogLogistic]{SimSurvivalLogLogistic()}},
\code{\link[=SimSurvivalWeibullPH]{SimSurvivalWeibullPH()}}
}
\concept{SimSurvival}
