% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSurvival.R
\name{SimSurvivalWeibullPH}
\alias{SimSurvivalWeibullPH}
\title{Simulate Survival Data from a Weibull Proportional Hazard Model}
\usage{
SimSurvivalWeibullPH(
  lambda,
  gamma,
  time_max = 2000,
  time_step = 1,
  lambda_censor = 1/3000,
  beta_cont = 0.2,
  beta_cat = c(A = 0, B = -0.4, C = 0.2)
)
}
\arguments{
\item{lambda}{(\code{number}) the scale parameter.}

\item{gamma}{(\code{number}) the shape parameter.}

\item{time_max}{(\code{number}) the maximum time to simulate to.}

\item{time_step}{(\code{number}) the time interval between evaluating the log-hazard function.}

\item{lambda_censor}{(\code{number}) the censoring rate.}

\item{beta_cont}{(\code{number}) the continuous covariate coefficient.}

\item{beta_cat}{(\code{numeric}) the categorical covariate coefficients.}
}
\value{
A \code{SimSurvival} object.
}
\description{
Simulate Survival Data from a Weibull Proportional Hazard Model
}
\section{Hazard Evaluation}{


Event times are simulated by sampling a cumulative hazard limit from a \eqn{U(0, 1)} distribution
for
each subject and then counting how much hazard they've been exposed to by evaluating the
log-hazard function at a set interval. The \code{time_max} argument sets the upper bound for the
number of time points to evaluate the log-hazard function at with subjects who have not had an
event being censored at \code{time_max}. The \code{time_step} argument sets the interval at which to
evaluate the log-hazard function. Setting smaller values for \code{time_step} will increase the
precision of the simulation at the cost of increased computation time. Likewise, setting large
values for \code{time_max} will minimize the number of censored subjects at the cost of
increased computation time.
}

\examples{
SimSurvivalWeibullPH(lambda = 1 / 300, gamma = 1.2)
}
\seealso{
Other SimSurvival:
\code{\link{SimSurvival-class}},
\code{\link[=SimSurvivalExponential]{SimSurvivalExponential()}},
\code{\link[=SimSurvivalGamma]{SimSurvivalGamma()}},
\code{\link[=SimSurvivalLogLogistic]{SimSurvivalLogLogistic()}}
}
\concept{SimSurvival}
