% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_os_sld}
\alias{simulate_os_sld}
\title{Simulate OS and SLD data}
\usage{
simulate_os_sld(
  N = 200,
  lambda_cen = 1/3,
  lambda = 0.9,
  p = 2.1,
  beta = 0.0035,
  gamma = -1,
  psi_mu = c(b = 58, s = 1.5, g = 0.5, phi = 0.3),
  psi_omega = c(b = 0.135, s = 0.15, g = 0.225, phi = 0.7),
  eta_mu = rep(0, 4),
  eta_sigma = diag(rep(2, 4)),
  beta_age = 0.2,
  beta_sex = c(M = 0.1, F = -0.1),
  times = c(0, 60, 120, 180, 300, 420, 660, 900),
  time_max = 6,
  pt_suffix = "suffix"
)
}
\arguments{
\item{N}{Number of patients to simulate}

\item{lambda_cen}{Lambda for censoring distribution (1/mean in years)}

\item{lambda}{Lambda for log-logistic baseline OS distribution}

\item{p}{p for log-logistic baseline OS distribution}

\item{beta}{link parameter for derivative SLD model}

\item{gamma}{link parameter for TTG}

\item{psi_omega}{Variance for the SLD kinetrics parameters}

\item{eta_mu}{Mean for the random effects of the SLD parameters}

\item{eta_sigma}{Covariance for the random effects of the SLD parameters}

\item{beta_age}{Beta parameter for the age effect}

\item{beta_sex}{Beta parameters for sex effect}

\item{times}{Times in which SLD observations occoured at (time 0 = Baseline)}

\item{time_max}{Maximum time to calculate hazard at. In years. will error if patient
hasn't died before this time. Excessively large values will increase
the functions run time without adding value}

\item{pt_suffix}{A string to attach to the end of the pt name in order to ensure its unique
across multiple runs of the function (for different arms / studies)}

\item{psi_mu_placebo}{SLD Kinetics parameters for the placebo group}

\item{psi_mu_trt}{SLD Kinetics parameters for the treatment group}
}
\description{
Simulate OS and SLD data
}
