% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_module.R
\name{source_stan_part}
\alias{source_stan_part}
\title{source_stan_part exports a .stan file as a character}
\usage{
source_stan_part(filename)
}
\arguments{
\item{filename}{The name of the .stan file in the directory of the package.}
}
\description{
source_stan_part exports a .stan file as a character
}
