% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\docType{data}
\name{STAN_BLOCKS}
\alias{STAN_BLOCKS}
\title{List of Stan Blocks}
\format{
An object of class \code{list} of length 7.
}
\usage{
STAN_BLOCKS
}
\description{
A list with 1 element per standard Stan program blocks.
This object is  mostly used internally as a reference for
what blocks are expected to exist within a given Stan program.
}
\keyword{datasets}
