% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_generics.R
\name{standard-link-user}
\alias{standard-link-user}
\alias{linkNone}
\alias{linkTTG}
\alias{linkDSLD}
\alias{linkIdentity}
\alias{linkGrowth}
\alias{linkShrinkage}
\title{Standard Links}
\usage{
linkNone()

linkTTG(prior, model = PromiseLongitudinalModel(), ...)

linkDSLD(prior, model = PromiseLongitudinalModel(), ...)

linkIdentity(prior, model = PromiseLongitudinalModel(), ...)

linkGrowth(prior, model = PromiseLongitudinalModel(), ...)

linkShrinkage(prior, model = PromiseLongitudinalModel(), ...)
}
\arguments{
\item{prior}{(\code{Prior}) A \code{\link{Prior}} object.}

\item{model}{(\code{LongitudinalModel}) A \code{\link{LongitudinalModel}} object.}

\item{...}{Not used.}
}
\value{
A \code{Link} or \code{LinkComponent} object representing the requested association.
}
\description{
These functions are used to enable the use of the corresponding link function between
the survival and longitudinal models in a joint model. Note that the exact implementation
of the link function is model specific, see
\code{vignette("Statistical Specifications", package = "jmpost")} for more details.
}
\section{Functions}{
\itemize{
\item \code{linkNone()}: No link (fit the survival and longitudinal models independently)

\item \code{linkTTG()}: Time to growth link

\item \code{linkDSLD()}: Derivative of the SLD over time link

\item \code{linkIdentity()}: Current SLD value link

\item \code{linkGrowth()}: Growth Parameter link

\item \code{linkShrinkage()}: Shrinkage Parameter link

}}
\examples{
linkNone()
linkDSLD()
linkTTG(prior_normal(0, 1), model = LongitudinalGSF())
}
