% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkComponent.R
\name{standard-links}
\alias{standard-links}
\alias{link_ttg}
\alias{link_dsld}
\alias{link_identity}
\alias{link_none}
\title{Standard Links}
\usage{
link_ttg(prior = prior_normal(0, 2))

link_dsld(prior = prior_normal(0, 2))

link_identity(prior = prior_normal(0, 2))

link_none()
}
\arguments{
\item{prior}{(\code{Prior})\cr The prior to use for the corresponding link coeficient.}
}
\description{
These functions enable the inclusion of several common link functions in the survival model of
the joint model.

Note that the underlying implementation of these links is specific to each longitudinal model.
}
\section{Functions}{
\itemize{
\item \code{link_ttg()}: Time to growth link

\item \code{link_dsld()}: Derivative of the SLD over time link

\item \code{link_identity()}: Current SLD link

\item \code{link_none()}: No link (fit the survival and longitudinal models independently)

}}
