% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{stanmodule_arguments}
\alias{stanmodule_arguments}
\title{\code{StanModule} Function Arguments}
\arguments{
\item{stan}{(\code{StanModule})\cr code containing the Stan code specification.}

\item{parameters}{(\code{ParameterList})\cr the parameter specification.}

\item{...}{additional arguments for \code{\link[=StanModule]{StanModule()}}.}
}
\description{
The documentation lists all the conventional arguments for wrappers around
\code{\link[=StanModule]{StanModule()}}.
}
