% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{subjects_to_list}
\alias{subjects_to_list}
\title{Expand and Validate Subjects}
\usage{
subjects_to_list(subjects = NULL, data)
}
\arguments{
\item{subjects}{(\code{character}) vector of subjects that should exist in \code{data}}

\item{data}{(\code{DataJoint}) Survival and Longitudinal Data.}
}
\value{
A named list grouping the requested subject identifiers.
}
\description{
If \code{subjects} is \code{NULL} this will return a named list of all subjects in \code{data}.
Else it will return \code{subjects} as a named list ensuring that all subjects exist in \code{data}.
}
\keyword{internal}
