% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\name{subset,DataJoint-method}
\alias{subset,DataJoint-method}
\title{Subsetting \code{DataJoint} as a \code{data.frame}}
\usage{
\S4method{subset}{DataJoint}(x, patients)
}
\arguments{
\item{x}{(\code{DataJoint}) \cr A \link[=DataJoint-class]{DataJoint} object created by \code{\link[=DataJoint]{DataJoint()}}}

\item{patients}{(\code{character} or \code{list})\cr the patients that you wish to subset the \code{data.frame}
to contain. See details.}
}
\description{
Coerces the object into a \code{data.frame} containing just event times and status
filtering for specific patients. If \code{patients} is a list then an additional variable \code{group} will be added
onto the dataset specifying which group the row belongs to.
}
\examples{
\dontrun{
pts <- c("PT1", "PT3", "PT4")
subset(x, pts)

groups <- list(
    "g1" = c("PT1", "PT3", "PT4"),
    "g2" = c("PT2", "PT3")
)
subset(x, groups)
}
}
\seealso{
Other \code{DataJoint}: 
\code{\link{DataJoint}}

Other \code{subset}: 
\code{\link{subset}()}
}
\concept{\code{DataJoint}}
\concept{\code{subset}}
