% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\name{subset.DataJoint}
\alias{subset.DataJoint}
\title{Subsetting \code{DataJoint} as a \code{data.frame}}
\usage{
\method{subset}{DataJoint}(x, subjects, ...)
}
\arguments{
\item{x}{(\code{DataJoint}) object created by \code{\link[=DataJoint]{DataJoint()}}.}

\item{subjects}{(\code{character} or \code{list}) subjects that you wish to subset the \code{data.frame}
to contain. See details.}

\item{...}{Not used.}
}
\value{
A \code{DataJoint} object restricted to the requested subjects.
}
\description{
Coerces the object into a \code{data.frame} containing just event times and status
filtering for specific subjects If \code{subjects} is a list then an additional variable \code{group} will be added
onto the dataset specifying which group the row belongs to.
}
\examples{
\dontrun{
subjects <- c("SUB1", "SUB2", "SUB3", "SUB4")
subset(x, subjects)

groups <- list(
    "g1" = c("SUB1", "SUB3", "SUB4"),
    "g2" = c("SUB2", "SUB3")
  )
subset(x, groups)
}
}
\seealso{
Other DataJoint:
\code{\link{DataJoint-class}},
\code{\link[=as_stan_list.DataSubject]{as_stan_list.DataSubject()}}
}
\concept{DataJoint}
