% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subset}
\alias{subset}
\title{Subsetting Vectors, Matrices and Data Frames}
\usage{
subset(x, ...)
}
\arguments{
\item{x}{object to be subsetted.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
NOTE: This man page is for the \code{subset} S4 generic function defined within
jmpost. See \code{\link[base:subset]{base::subset()}} for the default method.
}
\seealso{
Other \code{subset}: 
\code{\link{subset,DataJoint-method}}
}
\concept{\code{subset}}
