% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\name{subset_and_add_grouping}
\alias{subset_and_add_grouping}
\title{\code{subset_and_add_grouping}}
\usage{
subset_and_add_grouping(dat, groupings)
}
\arguments{
\item{dat}{(\code{data.frame}) must have a column called \code{subject} which corresponds to the
values passed to \code{groupings}.}

\item{groupings}{(\code{character} or \code{list}) subjects that you wish to subset the dataset
to contain. If \code{groupings} is a list then an additional variable \code{group} will be added
onto the dataset specifying which group the row belongs to.}
}
\value{
A subsetted \code{data.frame} with a grouping column.
}
\description{
\code{subset_and_add_grouping}
}
\details{
Example of usage

\if{html}{\out{<div class="sourceCode">}}\preformatted{subjects <- c("SUB1", "SUB2", "SUB3", "SUB4")
subset_and_add_grouping(dat, subjects)

groups <- list(
    "g1" = c("SUB1", "SUB3", "SUB4"),
    "g2" = c("SUB2", "SUB3")
  )
subset_and_add_grouping(dat, groups)
}\if{html}{\out{</div>}}
}
\keyword{internal}
