% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalSamples.R
\name{summarise_by_group}
\alias{summarise_by_group}
\title{Summarise Quantities By Group}
\usage{
summarise_by_group(subject_index, time_index, quantities)
}
\arguments{
\item{subject_index}{(\code{numeric})\cr Which subject indices to extract from \code{quantities}.
See details.}

\item{time_index}{(\code{numeric})\cr Which time point indices to extract from \code{quantities}.
See details.}

\item{quantities}{(\code{\link[posterior:draws_matrix]{posterior::draws_matrix}})\cr A matrix of sample draws.
See details.}
}
\value{
A data frame containing 1 row per \code{time_index} (in order) with the following columns:
\itemize{
\item \code{median} - The median value of the samples in \code{quantities}
\item \code{lower} - The lower \verb{95\%} CI value of the samples in \code{quantities}
\item \code{upper} - The upper \verb{95\%} CI value of the samples in \code{quantities}
}
}
\description{
This function takes a \code{\link[posterior:draws_matrix]{posterior::draws_matrix()}} (matrix of cmdstanr sample draws) and calculates
summary statistics (median / lower ci / upper ci) for selected columns.
A key feature is that it allows for columns to be aggregated together (see details).
}
\details{
It is assumed that \code{quantities} consists of the cartesian product
of subject indices and time indices. That is, if the matrix contains 4 subjects and 3 time
points then it should have 12 columns.
It is also assumed that each column of \code{quantities} are named as:

\if{html}{\out{<div class="sourceCode">}}\preformatted{"quantity[x,y]"
}\if{html}{\out{</div>}}

Where
\itemize{
\item \code{x} is the subject index
\item \code{y} is the time point index
}

The resulting \code{data.frame} that is created will have 1 row per value of \code{time_index} where
each row represents the summary statistics for that time point.

Note that if multiple values are provided for \code{subject_index} then the pointwise average
will be calculated for each time point by taking the mean across the specified subjects
at that time point.
}
\keyword{internal}
