% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalQuantities.R
\name{summary.LongitudinalQuantities}
\alias{summary.LongitudinalQuantities}
\title{summary}
\usage{
\method{summary}{LongitudinalQuantities}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{(\code{LongitudinalQuantities}) longitudinal quantities.}

\item{conf.level}{(\code{numeric}) confidence level of the interval.}

\item{...}{not used.}
}
\value{
A \code{data.frame} containing posterior summary statistics.
}
\description{
This method returns a \code{data.frame} of the longitudinal quantities.
}
\seealso{
Other LongitudinalQuantities:
\code{\link{LongitudinalQuantities-class}},
\code{\link[=as.data.frame.LongitudinalQuantities]{as.data.frame.LongitudinalQuantities()}},
\code{\link[=autoplot.LongitudinalQuantities]{autoplot.LongitudinalQuantities()}}

Other summary:
\code{\link[=summary.SurvivalQuantities]{summary.SurvivalQuantities()}}
}
\concept{LongitudinalQuantities}
\concept{summary}
