% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalRandomEffects.R
\name{summary.RandomEffectQuantities}
\alias{summary.RandomEffectQuantities}
\title{summary}
\usage{
\method{summary}{RandomEffectQuantities}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{(\code{\link{RandomEffectQuantities}}) \cr generated quantities.}

\item{conf.level}{(\code{numeric}) \cr confidence level of the interval.}

\item{...}{not used.}
}
\value{
A \code{data.frame} with the following variables:
\itemize{
\item \code{subject} (\code{character}) \cr the subject identifier.
\item \code{parameter} (\code{character}) \cr the parameter identifier.
\item \code{median} (\code{numeric}) \cr the median value of the quantity.
\item \code{lower} (\code{numeric}) \cr the lower CI value of the quantity.
\item \code{upper} (\code{numeric}) \cr the upper CI value of the quantity.
}
}
\description{
This method returns a summary statistic \code{data.frame} of the random effect parameters
}
\seealso{
Other RandomEffectQuantities: 
\code{\link{LongitudinalRandomEffects}()},
\code{\link{RandomEffectQuantities-class}},
\code{\link{as.data.frame.RandomEffectQuantities}()},
\code{\link{as_print_string.RandomEffectQuantities}()}
}
\concept{RandomEffectQuantities}
\keyword{internal}
