% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary}
\alias{summary}
\title{Object Summaries}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
NOTE: This man page is for the \code{summary} S4 generic function defined within
jmpost. See \code{\link[base:summary]{base::summary()}} for the default method.
}
\seealso{
Other summary: 
\code{\link{summary,SurvivalQuantities-method}}
}
\concept{summary}
