% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-exports.R
\name{Surv}
\alias{Surv}
\title{Survival Formula Specification}
\value{
An object of class \code{Surv}.
}
\description{
See \code{\link[survival:Surv]{survival::Surv()}} for details.
}
\examples{
Surv(c(5, 8, 12), c(1, 0, 1))
}
\keyword{internal}
