% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{survival}
\alias{survival}
\title{\code{survival}}
\usage{
survival(object, ...)
}
\arguments{
\item{object}{samples to extract the survival function values from.}

\item{...}{additional options.}
}
\description{
Obtain the survival function samples from \code{\link{JointModelSamples}}.
}
