% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalExponential.R
\docType{class}
\name{SurvivalExponential-class}
\alias{SurvivalExponential-class}
\alias{.SurvivalExponential}
\alias{SurvivalExponential}
\title{\code{SurvivalExponential}}
\usage{
SurvivalExponential(lambda = prior_gamma(2, 5), beta = prior_normal(0, 2))
}
\arguments{
\item{lambda}{(\code{Prior}) for the exponential rate \code{lambda}.}

\item{beta}{(\code{Prior}) for covariates coefficients \code{beta}.}
}
\value{
A \code{SurvivalExponential} object.
}
\description{
This class extends the general \code{\link{SurvivalModel}} class for using the
exponential survival model.
}
\examples{
SurvivalExponential()
}
