% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalModel.R
\docType{class}
\name{SurvivalModel-class}
\alias{SurvivalModel-class}
\alias{.SurvivalModel}
\alias{SurvivalModel}
\title{\code{SurvivalModel}}
\usage{
SurvivalModel(
  stan = StanModule(),
  parameters = ParameterList(),
  name = "<Unnamed>",
  ...
)
}
\arguments{
\item{stan}{(\code{StanModule}) code containing the model-specific Stan code specification.}

\item{parameters}{(\code{ParameterList}) the parameter declaration and prior specification.}

\item{name}{(\code{character}) display name for the model object.}

\item{...}{additional arguments for \code{\link[=StanModel]{StanModel()}}.}
}
\value{
A \code{SurvivalModel} object.
}
\description{
This class extends the general \code{\link{StanModel}} class to comprise the survival
model specification.
}
\examples{
SurvivalModel()
}
