% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{SurvivalQuantities-Shared}
\alias{SurvivalQuantities-Shared}
\title{Re-used documentation for \code{SurvivalQuantities}}
\arguments{
\item{object}{(\code{\link{SurvivalQuantities}}) \cr survival quantities.}

\item{...}{not used.}
}
\description{
Re-used documentation for \code{SurvivalQuantities}
}
\keyword{internal}
