% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalSamples.R
\docType{class}
\name{SurvivalSamples-class}
\alias{SurvivalSamples-class}
\alias{.SurvivalSamples}
\alias{SurvivalSamples}
\alias{[,SurvivalSamples,ANY,ANY,ANY-method}
\title{\code{SurvivalSamples}}
\usage{
\S4method{[}{SurvivalSamples,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{(\code{SurvivalSamples})\cr the samples object to subset.}

\item{i}{(\code{vector})\cr the index vector.}

\item{j}{not used.}

\item{...}{not used.}

\item{drop}{not used.}
}
\value{
The subsetted \code{SurvivalSamples} object.
}
\description{
This class is an extension of \code{list} so that we
can define specific survival postprocessing methods for it.
}
