% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalSamples.R
\name{SurvivalSamples}
\alias{SurvivalSamples}
\title{SurvivalSamples Constructor Function}
\usage{
SurvivalSamples(object)
}
\arguments{
\item{object}{(\code{JointModelSamples}) \cr A \linkS4class{JointModelSamples} object}
}
\description{
Creates a \link[=SurvivalSample-class]{SurvivalSample} object from a
\link[=JointModelSamples-class]{JointModelSamples} objet
}
\seealso{
Other SurvivalSamples: 
\code{\link{SurvivalSamples-class}},
\code{\link{autoplot,SurvivalSamples-method}},
\code{\link{predict,SurvivalSamples-method}}
}
\concept{SurvivalSamples}
