% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalWeibullPH.R
\docType{class}
\name{SurvivalWeibullPH-class}
\alias{SurvivalWeibullPH-class}
\alias{.SurvivalWeibullPH}
\alias{SurvivalWeibullPH}
\title{\code{SurvivalWeibullPH}}
\usage{
SurvivalWeibullPH(
  lambda = prior_gamma(2, 0.5),
  gamma = prior_gamma(2, 0.5),
  beta = prior_normal(0, 2)
)
}
\arguments{
\item{lambda}{(\code{Prior}) for the scale \code{lambda}.}

\item{gamma}{(\code{Prior}) for the shape \code{gamma}.}

\item{beta}{(\code{Prior}) for covariates coefficients \code{beta}.}
}
\value{
A \code{SurvivalWeibullPH} object.
}
\description{
This class extends the general \code{\link{SurvivalModel}} class for using the
Weibull proportional hazards survival model.
}
\examples{
SurvivalWeibullPH()
}
