% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttg.R
\name{ttg}
\alias{ttg}
\title{Time to growth function}
\usage{
ttg(psi_ks, psi_kg, psi_phi)
}
\arguments{
\item{psi_ks}{The srinkage rate parameter}

\item{psi_kg}{The growth rate parameter}

\item{psi_phi}{The percentage of cells responded to the treatment}
}
\description{
Time to growth function
}
