% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jm_data.R
\name{vars}
\alias{vars}
\title{JMdata variable mapping}
\usage{
vars(
  longitudinal = "AVAL",
  os_user_id = "USUBJID",
  overall_survival_death = "DEATH",
  long_user_id = "USUBJID",
  time_survival = "AYR",
  os_study_id = "STUDYID",
  os_arm = "ARM",
  treatment = "TRT01P"
)
}
\arguments{
\item{longitudinal}{Longitudinal observations}

\item{os_user_id}{Patients ID for overall survival data}

\item{overall_survival_death}{Event indicator for survival data}

\item{long_user_id}{Patients ID for longitudinal data}

\item{time_survival}{Time of event for survival data}

\item{os_study_id}{Study indicator}

\item{os_arm}{Arm indicator}

\item{treatment}{Treatment indicator}
}
\description{
JMdata variable mapping
}
