% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimJointData.R
\name{add_pfs}
\alias{add_pfs}
\title{Sets PFS event at Tumour Progression}
\usage{
add_pfs(
  object,
  relative_threshold = 1.2,
  absolute_threshold = 5,
  from_time = 0,
  observed_after = FALSE
)
}
\arguments{
\item{object}{A \link{SimJointData} object}

\item{relative_threshold}{(\code{number})\cr a multiplicative threshold for the change in SLD compared to the \code{min(SLD)}.
Default is 1.2 meaning a 20\% increase.}

\item{absolute_threshold}{(\code{number})\cr an absolute threshold for the change in SLD compared to the minimum.
Default is 5.}

\item{from_time}{(\code{number})\cr Ignore observations before this time for determining SLD minimum.}

\item{observed_after}{(\code{logical})\cr If \code{FALSE} set longitudinal observations after the progression time to
\code{observed = FALSE}}
}
\description{
Sets PFS event at Tumour Progression
}
\details{
Both thresholds must be met for a progression to be declared.
}
\examples{
add_pfs(sim_data)
}
