% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Link.R, R/LinkNone.R,
%   R/defaults.R
\name{addLink}
\alias{addLink}
\alias{addLink,LongitudinalModel,Link-method}
\alias{addLink,LongitudinalModel,LinkNone-method}
\alias{addLink,LongitudinalModel,NULL-method}
\alias{addLink,NULL,Link-method}
\alias{addLink,NULL,NULL-method}
\title{\code{addLink}}
\usage{
addLink(x, y, ...)

\S4method{addLink}{LongitudinalModel,Link}(x, y, ...)

\S4method{addLink}{LongitudinalModel,LinkNone}(x, y, ...)

\S4method{addLink}{LongitudinalModel,`NULL`}(x, y, ...)

\S4method{addLink}{`NULL`,Link}(x, y, ...)

\S4method{addLink}{`NULL`,`NULL`}(x, y, ...)
}
\arguments{
\item{x}{the longitudinal model.}

\item{y}{the link to be added.}

\item{...}{additional arguments.}
}
\description{
Add a link to a longitudinal model.
}
