% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModel.R
\name{as.character,JointModel-method}
\alias{as.character,JointModel-method}
\title{As character}
\usage{
\S4method{as.character}{JointModel}(x)
}
\arguments{
\item{x}{A \code{JointModel} object}
}
\description{
As character
}
