% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModel.R
\name{as.character.JointModel}
\alias{as.character.JointModel}
\title{\code{JointModel} -> \code{character}}
\usage{
\method{as.character}{JointModel}(x, ...)
}
\arguments{
\item{x}{(\code{\link{JointModel}}) \cr Joint model specification.}

\item{...}{Not Used.}
}
\description{
Renders a \code{\link{JointModel}} object to a stan program
}
\seealso{
Other JointModel: 
\code{\link{JointModel-class}}
}
\concept{JointModel}
