% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{as.character.Parameter}
\alias{as.character.Parameter}
\title{\code{Parameter} -> \code{character}}
\usage{
\method{as.character}{Parameter}(x, ...)
}
\arguments{
\item{x}{(\code{Paramater}) \cr A model parameter}

\item{...}{Not Used.}
}
\description{
Converts a parameter object into its corresponding Stan code representation
}
\seealso{
Other Parameter: 
\code{\link{Parameter-Getter-Methods}},
\code{\link{Parameter-class}}
}
\concept{Parameter}
