% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{as.character.Prior}
\alias{as.character.Prior}
\title{\code{Prior} -> \code{character}}
\usage{
\method{as.character}{Prior}(x, ...)
}
\arguments{
\item{x}{(\code{\link{Prior}})\cr A Prior Distribution}

\item{...}{Not Used.}
}
\description{
Returns the character representation of the distribution by inserting the
distribution parameters into the \code{x@repr} string
}
\seealso{
Other Prior: 
\code{\link{Prior-Getter-Methods}},
\code{\link{Prior-class}},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_lognormal}()},
\code{\link{prior_none}()},
\code{\link{prior_normal}()},
\code{\link{prior_std_normal}()}
}
\concept{Prior}
