% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\name{as.character.StanModule}
\alias{as.character.StanModule}
\title{\code{StanModule} -> \code{character}}
\usage{
\method{as.character}{StanModule}(x, ...)
}
\arguments{
\item{x}{(\code{\link{StanModule}})\cr A stan program}

\item{...}{Not Used.}
}
\description{
Converts a \code{\link{StanModule}} object into a valid Stan program file where each
line of the returned \code{character} vector represents a line of the program
}
\seealso{
Other StanModule: 
\code{\link{StanModule-class}},
\code{\link{as.list.StanModule}()}
}
\concept{StanModule}
