% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModelSamples.R
\name{as.CmdStanMCMC}
\alias{as.CmdStanMCMC}
\alias{as.CmdStanMCMC.JointModelSamples}
\title{Coerce to CmdStanMCMC}
\usage{
as.CmdStanMCMC(object, ...)

\method{as.CmdStanMCMC}{JointModelSamples}(object, ...)
}
\arguments{
\item{object}{to be converted}

\item{...}{additional options}
}
\description{
Coerces an object to a \code{\link[cmdstanr:CmdStanMCMC]{cmdstanr::CmdStanMCMC}} object
}
