% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModelSamples.R, R/external-exports.R
\name{as.CmdStanMCMC.JointModelSamples}
\alias{as.CmdStanMCMC.JointModelSamples}
\alias{as.CmdStanMCMC}
\title{Coerce to \code{CmdStanMCMC}}
\usage{
\method{as.CmdStanMCMC}{JointModelSamples}(object, ...)
}
\arguments{
\item{object}{to be converted}

\item{...}{additional options}
}
\description{
Coerces an object to a \code{\link[cmdstanr:CmdStanMCMC]{cmdstanr::CmdStanMCMC}} object
}
