% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSubject.R
\name{as.data.frame.DataSubject}
\alias{as.data.frame.DataSubject}
\title{\code{DataSubject} -> \code{data.frame}}
\usage{
\method{as.data.frame}{DataSubject}(x, ...)
}
\arguments{
\item{x}{(\code{\link{DataSubject}}) \cr subject-level data.}

\item{...}{Not Used.}
}
\description{
Converts a \code{\link{DataSubject}} object into a \code{data.frame}.
The subject variable is cast to factor.
}
\seealso{
Other DataSubject: 
\code{\link{DataSubject-class}},
\code{\link{as_print_string.DataSubject}()},
\code{\link{as_stan_list.DataSubject}()},
\code{\link{extractVariableNames.DataSubject}()}
}
\concept{DataSubject}
