% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\name{as.data.frame.DataSurvival}
\alias{as.data.frame.DataSurvival}
\title{\code{DataSurvival} -> \code{data.frame}}
\usage{
\method{as.data.frame}{DataSurvival}(x, ...)
}
\arguments{
\item{x}{(\code{\link{DataSurvival}}) \cr Survival Data.}

\item{...}{Not Used.}
}
\description{
Converts a \code{\link{DataSurvival}} object into a \code{data.frame}.
The subject variable is cast to factor.
}
\seealso{
Other DataSurvival: 
\code{\link{DataSurvival-class}},
\code{\link{as.list.DataSurvival}()},
\code{\link{extractVariableNames.DataSurvival}()}
}
\concept{DataSurvival}
