% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\name{as.data.frame.Quantities}
\alias{as.data.frame.Quantities}
\title{\code{Quantities} -> \code{data.frame}}
\usage{
\method{as.data.frame}{Quantities}(x, time_grid, type, groups)
}
\arguments{
\item{x}{(\code{\link{Quantities}}) \cr Any generated model quantity}

\item{...}{Not used}
}
\description{
\code{Quantities} -> \code{data.frame}
}
\seealso{
Other Quantities: 
\code{\link{Quantities-class}},
\code{\link{summary.Quantities}()}
}
\concept{Quantities}
\keyword{internal}
