% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalRandomEffects.R
\name{as.data.frame.RandomEffectQuantities}
\alias{as.data.frame.RandomEffectQuantities}
\title{\code{RandomEffectQuantities} -> \code{data.frame}}
\usage{
\method{as.data.frame}{RandomEffectQuantities}(x, ...)
}
\arguments{
\item{x}{(\code{\link{RandomEffectQuantities}}) \cr generated quantities.}

\item{...}{not used.}
}
\description{
Returns a \code{data.frame} of the subject-level random effect parameter samples.
}
\seealso{
Other RandomEffectQuantities: 
\code{\link{LongitudinalRandomEffects}()},
\code{\link{RandomEffectQuantities-class}},
\code{\link{as_print_string.RandomEffectQuantities}()},
\code{\link{summary.RandomEffectQuantities}()}
}
\concept{RandomEffectQuantities}
\keyword{internal}
