% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{as.data.frame.SurvivalQuantities}
\alias{as.data.frame.SurvivalQuantities}
\title{\code{SurvivalQuantities} -> \code{data.frame}}
\usage{
\method{as.data.frame}{SurvivalQuantities}(x, ...)
}
\arguments{
\item{x}{(\code{\link{SurvivalQuantities}}) \cr survival quantities.}

\item{...}{not used.}
}
\description{
\code{SurvivalQuantities} -> \code{data.frame}
}
\seealso{
Other SurvivalQuantities: 
\code{\link{SurvivalQuantities-class}},
\code{\link{as_print_string.SurvivalQuantities}()},
\code{\link{autoplot.SurvivalQuantities}()},
\code{\link{brierScore.SurvivalQuantities}()},
\code{\link{summary.SurvivalQuantities}()}
}
\concept{SurvivalQuantities}
