% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link.R
\name{as.list.Link}
\alias{as.list.Link}
\title{\code{Link} -> \code{list}}
\usage{
\method{as.list}{Link}(object, ...)
}
\arguments{
\item{object}{(\code{\link{Link}})\cr a link object.}

\item{...}{Not Used.}
}
\description{
Returns a named list where each element of the list corresponds
to a Stan modelling block e.g. \code{data}, \code{model}, etc.
}
\seealso{
Other Link: 
\code{\link{Link-class}},
\code{\link{as.StanModule.Link}()},
\code{\link{length.Link}()}
}
\concept{Link}
