% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkComponent.R
\name{as.list.LinkComponent}
\alias{as.list.LinkComponent}
\title{\code{LinkComponent} -> \code{list}}
\usage{
\method{as.list}{LinkComponent}(x, ...)
}
\arguments{
\item{x}{(\code{\link{LinkComponent}})\cr a link component.}

\item{...}{Not Used.}
}
\description{
Returns a named list where each element of the list corresponds
to a Stan modelling block e.g. \code{data}, \code{model}, etc.
}
\seealso{
Other LinkComponent: 
\code{\link{LinkComponent-class}},
\code{\link{as.StanModule.LinkComponent}()},
\code{\link{getParameters}()},
\code{\link{initialValues}()}
}
\concept{LinkComponent}
