% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterList.R
\name{as.list.ParameterList}
\alias{as.list.ParameterList}
\title{\code{ParameterList} -> \code{list}}
\usage{
\method{as.list}{ParameterList}(x, ...)
}
\arguments{
\item{x}{(\code{ParameterList}) \cr A List of \code{\link{Parameter}} Objects.}

\item{...}{Not Used.}
}
\description{
Returns a named list where each element of the list corresponds
to a Stan modelling block e.g. \code{data}, \code{model}, etc.
}
\seealso{
Other ParameterList: 
\code{\link{ParameterList-Getter-Methods}},
\code{\link{ParameterList-class}},
\code{\link{as.StanModule.ParameterList}()},
\code{\link{as_print_string.ParameterList}()},
\code{\link{as_stan_list.ParameterList}()}
}
\concept{ParameterList}
