% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModel.R
\name{as.list.StanModel}
\alias{as.list.StanModel}
\title{\code{StanModel} -> \code{list}}
\usage{
\method{as.list}{StanModel}(x, ...)
}
\arguments{
\item{x}{(\code{StanModel}) A Stan Model}

\item{...}{Not Used.}
}
\description{
Returns a named list where each element of the list corresponds
to a Stan modelling block e.g. \code{data}, \code{model}, etc.
}
\seealso{
Other StanModel:
\code{\link{StanModel-class}}
}
\concept{StanModel}
