% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\name{as.list.StanModule}
\alias{as.list.StanModule}
\title{\code{StanModule} -> \code{list}}
\usage{
\method{as.list}{StanModule}(x, stan_blocks = STAN_BLOCKS, ...)
}
\arguments{
\item{x}{(\code{\link{StanModule}})\cr A Stan Module}

\item{stan_blocks}{(\code{list})\cr reference list of stan blocks.}

\item{...}{Not Used.}
}
\description{
Returns a named list where each element of the list corresponds
to a Stan modelling block e.g. \code{data}, \code{model}, etc.
}
\seealso{
Other StanModule: 
\code{\link{StanModule-class}},
\code{\link{as.character.StanModule}()}
}
\concept{StanModule}
