% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModel.R
\name{as.StanModule.JointModel}
\alias{as.StanModule.JointModel}
\title{\code{JointModel} -> \code{StanModule}}
\usage{
\method{as.StanModule}{JointModel}(object, ...)
}
\arguments{
\item{object}{(\code{\link{JointModel}}) \cr Joint model specification.}

\item{...}{Not Used.}
}
\description{
Converts a \code{\link{JointModel}} object to a \code{\link{StanModule}} object
}
\seealso{
Other JointModel: 
\code{\link{JointModel-class}},
\code{\link{as.character.JointModel}()}

Other as.StanModule: 
\code{\link{as.StanModule}()},
\code{\link{as.StanModule.Link}()},
\code{\link{as.StanModule.LinkComponent}()},
\code{\link{as.StanModule.Parameter}()},
\code{\link{as.StanModule.ParameterList}()},
\code{\link{as.StanModule.Prior}()}
}
\concept{JointModel}
\concept{as.StanModule}
