% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModelSamples.R
\name{as.StanModule.JointModelSamples}
\alias{as.StanModule.JointModelSamples}
\title{\code{JointModelSamples} -> \code{StanModule}}
\usage{
\method{as.StanModule}{JointModelSamples}(object, generator, type, ...)
}
\arguments{
\item{object}{object to obtain generated quantities from}

\item{generator}{(\code{QuantityGenerator})\cr object that specifies which subjects and time points
to calculate the quantities at}

\item{type}{(\code{character})\cr type of quantities to be generated, must be either "survival" or
"longitudinal".}

\item{...}{additional options.}
}
\description{
Converts a \code{JointModelSamples} object into a \code{StanModule} object ensuring
that the resulting \code{StanModule} object is able to generate post sampling
quantities.
}
