% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkComponent.R
\name{as.StanModule.LinkComponent}
\alias{as.StanModule.LinkComponent}
\title{\code{LinkComponent} -> \code{StanModule}}
\usage{
\method{as.StanModule}{LinkComponent}(object, ...)
}
\arguments{
\item{object}{(\code{\link{LinkComponent}})\cr a link component.}

\item{...}{Not Used.}
}
\description{
Converts a \code{\link{LinkComponent}} object to a \code{\link{StanModule}} object
}
\seealso{
Other LinkComponent: 
\code{\link{LinkComponent-class}},
\code{\link{as.list.LinkComponent}()},
\code{\link{getParameters}()},
\code{\link{initialValues}()}

Other as.StanModule: 
\code{\link{as.StanModule}()},
\code{\link{as.StanModule.JointModel}()},
\code{\link{as.StanModule.Link}()},
\code{\link{as.StanModule.Parameter}()},
\code{\link{as.StanModule.ParameterList}()},
\code{\link{as.StanModule.Prior}()}
}
\concept{LinkComponent}
\concept{as.StanModule}
