% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{as.StanModule.Parameter}
\alias{as.StanModule.Parameter}
\title{\code{Parameter} -> \code{character}}
\usage{
\method{as.StanModule}{Parameter}(object, ...)
}
\arguments{
\item{...}{Not Used.}

\item{x}{(\code{Paramater}) \cr A model parameter}
}
\description{
Converts a parameter object into its corresponding Stan code representation
}
\seealso{
Other Parameter: 
\code{\link{Parameter-Getter-Methods}},
\code{\link{Parameter-class}}
}
\concept{Parameter}
