% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{as.StanModule.Parameter}
\alias{as.StanModule.Parameter}
\title{\code{Parameter} -> \code{StanModule}}
\usage{
\method{as.StanModule}{Parameter}(object, ...)
}
\arguments{
\item{object}{(\code{\link{Parameter}})\cr a prior Distribution}

\item{...}{Not Used.}
}
\description{
Converts a \code{\link{Parameter}} object to a \code{\link{StanModule}} object
}
\seealso{
Other Parameter: 
\code{\link{Parameter-Getter-Methods}},
\code{\link{Parameter-class}},
\code{\link{as.character.Parameter}()},
\code{\link{as_stan_list.Parameter}()}

Other as.StanModule: 
\code{\link{as.StanModule}()},
\code{\link{as.StanModule.ParameterList}()},
\code{\link{as.StanModule.Prior}()}
}
\concept{Parameter}
\concept{as.StanModule}
