% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{as.StanModule.ParameterDeclaration}
\alias{as.StanModule.ParameterDeclaration}
\title{\code{Parameter} Declaration -> \code{StanModule}}
\usage{
\method{as.StanModule}{ParameterDeclaration}(object, ...)
}
\arguments{
\item{object}{(\code{Parameter}) a prior Distribution}

\item{...}{not used.}
}
\value{
A \code{\link{StanModule}} object.
}
\description{
Creates only the Stan declaration for a \code{\link{Parameter}}. Sampled parameters are
declared in the \code{parameters} block; constant parameters are declared in the
\verb{transformed parameters} block.
}
