% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterList.R
\name{as.StanModule.ParameterList}
\alias{as.StanModule.ParameterList}
\title{\code{ParameterList} -> \code{StanModule}}
\usage{
\method{as.StanModule}{ParameterList}(object)
}
\arguments{
\item{object}{(\code{ParameterList}) \cr A List of \code{\link{Parameter}} Objects.}
}
\description{
Returns a \code{\link{StanModule}} object of the model parameters.
}
\seealso{
Other ParameterList: 
\code{\link{ParameterList-Getter-Methods}},
\code{\link{ParameterList-class}},
\code{\link{as.character.ParameterList}()},
\code{\link{as.list.ParameterList}()}
}
\concept{ParameterList}
