% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{as.StanModule.Prior}
\alias{as.StanModule.Prior}
\title{\code{Prior} -> \code{StanModule}}
\usage{
\method{as.StanModule}{Prior}(object, name, ...)
}
\arguments{
\item{object}{(\code{\link{Prior}})\cr A Prior Distribution}

\item{name}{TODO}

\item{...}{Not Used.}
}
\description{
asdaw
}
\seealso{
Other Prior: 
\code{\link{Prior-Getter-Methods}},
\code{\link{Prior-class}},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_lognormal}()},
\code{\link{prior_none}()},
\code{\link{prior_normal}()},
\code{\link{prior_std_normal}()}
}
\concept{Prior}
