% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.StanModule}
\alias{as.StanModule}
\title{\code{as.StanModule}}
\usage{
as.StanModule(object, ...)
}
\arguments{
\item{object}{what to convert.}

\item{...}{additional options.}
}
\description{
Converts an object into a \code{\link{StanModule}}.
}
\seealso{
Other as.StanModule:
\code{\link[=as.StanModule.JointModel]{as.StanModule.JointModel()}},
\code{\link[=as.StanModule.Link]{as.StanModule.Link()}},
\code{\link[=as.StanModule.LinkComponent]{as.StanModule.LinkComponent()}},
\code{\link[=as.StanModule.Parameter]{as.StanModule.Parameter()}},
\code{\link[=as.StanModule.ParameterList]{as.StanModule.ParameterList()}},
\code{\link[=as.StanModule.Prior]{as.StanModule.Prior()}}
}
\concept{as.StanModule}
\keyword{internal}
