% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ParameterList.R, R/Link.R
\name{as.StanModule}
\alias{as.StanModule}
\alias{as.StanModule,ParameterList-method}
\alias{as.StanModule,Link-method}
\title{\code{as.StanModule}}
\usage{
as.StanModule(object)

\S4method{as.StanModule}{ParameterList}(object)

\S4method{as.StanModule}{Link}(object)
}
\arguments{
\item{object}{what to convert.}
}
\description{
Convert a \code{\link{Link}} or \code{\link{ParameterList}} into a \code{\link{StanModule}}.
}
\keyword{internal}
