% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Link.R
\name{as.StanModule}
\alias{as.StanModule}
\alias{as.StanModule.Link}
\title{\code{as.StanModule}}
\usage{
as.StanModule(object, ...)

\method{as.StanModule}{Link}(object)
}
\arguments{
\item{object}{what to convert.}

\item{...}{additional options.}
}
\description{
Converts an object into a \code{\link{StanModule}}.
}
\seealso{
Other as.StanModule: 
\code{\link{as.StanModule.Parameter}()},
\code{\link{as.StanModule.ParameterList}()},
\code{\link{as.StanModule.Prior}()}
}
\concept{as.StanModule}
\keyword{internal}
