% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.StanModule}
\alias{as.StanModule}
\title{\code{as.StanModule}}
\usage{
as.StanModule(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\description{
Description - TODO
}
\keyword{internal}
