% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_formula}
\alias{as_formula}
\title{As Formula}
\usage{
as_formula(x, ...)
}
\arguments{
\item{x}{(\code{ANY}) \cr object to convert to a formula.}

\item{...}{Not used.}
}
\description{
Utility wrapper function to convert an object to a formula.
}
