% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLongitudinal.R
\name{as_print_string.DataLongitudinal}
\alias{as_print_string.DataLongitudinal}
\title{\code{DataLongitudinal} -> Printable \code{Character}}
\usage{
\method{as_print_string}{DataLongitudinal}(object, indent = 1, ...)
}
\arguments{
\item{object}{(\code{\link{DataLongitudinal}}) \cr Longitudinal Data.}

\item{...}{Not Used.}
}
\description{
Converts \code{\link{DataLongitudinal}} object into a printable string.
}
\seealso{
Other DataLongitudinal: 
\code{\link{DataLongitudinal-class}},
\code{\link{as.data.frame.DataLongitudinal}()},
\code{\link{as_stan_list.DataSubject}()},
\code{\link{extractVariableNames.DataLongitudinal}()}
}
\concept{DataLongitudinal}
\keyword{internal}
