% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\name{as_print_string.Quantities}
\alias{as_print_string.Quantities}
\title{\code{Quantities} -> Printable \code{Character}}
\usage{
\method{as_print_string}{Quantities}(object, indent = 1, ...)
}
\arguments{
\item{object}{(\code{\link{Quantities}}) \cr generated quantities.}

\item{...}{not used.}
}
\description{
Converts \code{\link{Quantities}} object into a printable string.
}
\seealso{
Other Quantities: 
\code{\link{Quantities-class}},
\code{\link{as.data.frame.Quantities}()},
\code{\link{summary.Quantities}()}
}
\concept{Quantities}
\keyword{internal}
