% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalRandomEffects.R
\name{as_print_string.RandomEffectQuantities}
\alias{as_print_string.RandomEffectQuantities}
\title{\code{RandomEffectQuantities} -> Printable \code{Character}}
\usage{
\method{as_print_string}{RandomEffectQuantities}(object, indent = 1, ...)
}
\arguments{
\item{object}{(\code{\link{RandomEffectQuantities}}) \cr generated quantities.}

\item{indent}{(\code{numeric}) \cr the number of spaces to indent the string by.}

\item{...}{not used.}
}
\description{
Converts \code{\link{RandomEffectQuantities}} object into a printable string.
}
\seealso{
Other RandomEffectQuantities: 
\code{\link{LongitudinalRandomEffects}()},
\code{\link{RandomEffectQuantities-class}},
\code{\link{as.data.frame.RandomEffectQuantities}()},
\code{\link{summary.RandomEffectQuantities}()}
}
\concept{RandomEffectQuantities}
\keyword{internal}
