% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_print_string}
\alias{as_print_string}
\title{\code{as_print_string}}
\usage{
as_print_string(object, ...)
}
\arguments{
\item{object}{to be converted to string.}

\item{...}{additional options.}
}
\description{
Returns the character representation of an object which is suitable
for printing to the console
}
\concept{as_print_string}
\keyword{internal}
