% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/SimLongitudinal.R,
%   R/SimLongitudinalGSF.R, R/SimLongitudinalRandomSlope.R,
%   R/SimLongitudinalSteinFojo.R, R/SimSurvival.R
\name{as_print_string}
\alias{as_print_string}
\alias{as_print_string.SimLongitudinal}
\alias{as_print_string.SimLongitudinalGSF}
\alias{as_print_string.SimLongitudinalRandomSlope}
\alias{as_print_string.SimLongitudinalSteinFojo}
\alias{as_print_string.SimSurvival}
\title{\code{as_print_string}}
\usage{
as_print_string(object, ...)

\method{as_print_string}{SimLongitudinal}(object)

\method{as_print_string}{SimLongitudinalGSF}(object)

\method{as_print_string}{SimLongitudinalRandomSlope}(object)

\method{as_print_string}{SimLongitudinalSteinFojo}(object)

\method{as_print_string}{SimSurvival}(object)
}
\arguments{
\item{object}{to be converted to string.}

\item{...}{additional options.}
}
\description{
Returns the character representation of an object which is suitable
for printing to the console
}
\concept{as_print_string}
\keyword{internal}
