% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\name{as_print_string.StanModule}
\alias{as_print_string.StanModule}
\title{\code{StanModule} -> Printable \code{Character}}
\usage{
\method{as_print_string}{StanModule}(object, indent = 1, ...)
}
\arguments{
\item{object}{(\code{\link{StanModule}})\cr A stan program}

\item{indent}{(\code{numeric})\cr how much white space to prefix the print string with.}
}
\description{
Converts \code{\link{StanModule}} object into a printable string.
}
\seealso{
Other StanModule: 
\code{\link{StanModule-class}},
\code{\link{as.character.StanModule}()},
\code{\link{as.list.StanModule}()}
}
\concept{StanModule}
\keyword{internal}
