% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\name{as_stan_file}
\alias{as_stan_file}
\title{Merging Code Blocks into Stan Code Character Vector}
\usage{
as_stan_file(
  functions = "",
  data = "",
  transformed_data = "",
  parameters = "",
  transformed_parameters = "",
  model = "",
  generated_quantities = ""
)
}
\arguments{
\item{functions}{(\code{character})\cr code block.}

\item{data}{(\code{character})\cr code block.}

\item{transformed_data}{(\code{character})\cr code block.}

\item{parameters}{(\code{character})\cr code block.}

\item{transformed_parameters}{(\code{character})\cr code block.}

\item{model}{(\code{character})\cr code block.}

\item{generated_quantities}{(\code{character})\cr code block.}
}
\value{
Character vector of the complete Stan code.
}
\description{
Merging Code Blocks into Stan Code Character Vector
}
\keyword{internal}
