% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{as_stan_list.Parameter}
\alias{as_stan_list.Parameter}
\title{\code{Parameter} -> \code{list}}
\usage{
\method{as_stan_list}{Parameter}(object, ...)
}
\arguments{
\item{object}{(\code{\link{Parameter}})\cr a prior Distribution}

\item{...}{Not Used.}
}
\description{
Converts a Parameter object to a list of parameter data values
for a Stan model.
}
\seealso{
Other as_stan_list: 
\code{\link{as_stan_list.DataSubject}()},
\code{\link{as_stan_list.ParameterList}()},
\code{\link{as_stan_list.Prior}()},
\code{\link{as_stan_list}()}

Other Parameter: 
\code{\link{Parameter-Getter-Methods}},
\code{\link{Parameter-class}},
\code{\link{as.StanModule.Parameter}()}
}
\concept{Parameter}
\concept{as_stan_list}
