% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterList.R
\name{as_stan_list.ParameterList}
\alias{as_stan_list.ParameterList}
\title{\code{ParameterList} -> \code{list}}
\usage{
\method{as_stan_list}{ParameterList}(object, ...)
}
\arguments{
\item{object}{(\code{ParameterList}) \cr A List of \code{\link{Parameter}} Objects.}

\item{...}{Not Used.}
}
\description{
Converts a ParameterList object to a list of parameter data values
for a Stan model.
}
\seealso{
Other as_stan_list: 
\code{\link{as_stan_list}()},
\code{\link{as_stan_list.DataSubject}()},
\code{\link{as_stan_list.Parameter}()},
\code{\link{as_stan_list.Prior}()}

Other ParameterList: 
\code{\link{ParameterList-Getter-Methods}},
\code{\link{ParameterList-class}},
\code{\link{as.StanModule.ParameterList}()},
\code{\link{as.list.ParameterList}()},
\code{\link{as_print_string.ParameterList}()}
}
\concept{ParameterList}
\concept{as_stan_list}
