% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R, R/QuantityGeneratorPopulation.R,
%   R/QuantityGeneratorPrediction.R, R/QuantityGeneratorSubject.R
\name{as_stan_list.QuantityGenerator}
\alias{as_stan_list.QuantityGenerator}
\alias{as_stan_list.QuantityGeneratorPopulation}
\alias{as_stan_list.QuantityGeneratorPrediction}
\alias{as_stan_list.QuantityGeneratorSubject}
\title{\code{QuantityGenerator} -> \code{list}}
\usage{
\method{as_stan_list}{QuantityGenerator}(object, data, ...)

\method{as_stan_list}{QuantityGeneratorPopulation}(object, data, ...)

\method{as_stan_list}{QuantityGeneratorPrediction}(object, data, model, ...)

\method{as_stan_list}{QuantityGeneratorSubject}(object, data, ...)
}
\arguments{
\item{object}{(\code{QuantityGenerator}) object to convert to a list.}

\item{data}{(\code{DataJoint}) Survival and Longitudinal Data.}

\item{...}{Not currently used.}
}
\value{
A named \code{list} suitable for use as Stan data.
}
\description{
Converts a \code{QuantityGenerator} object to a list containing the required input data for a stan
model.
}
\keyword{internal}
