% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSubject.R, R/generics.R,
%   R/DataLongitudinal.R, R/DataSurvival.R, R/DataJoint.R
\name{as_stan_list.DataSubject}
\alias{as_stan_list.DataSubject}
\alias{as.list.DataSubject}
\alias{as_stan_list}
\alias{as_stan_list.default}
\alias{as_stan_list.DataLongitudinal}
\alias{as.list.DataLongitudinal}
\alias{as_stan_list.DataSurvival}
\alias{as.list.DataSurvival}
\alias{as_stan_list.DataJoint}
\alias{as.list.DataJoint}
\title{Data Object -> \code{list}}
\usage{
\method{as_stan_list}{DataSubject}(object, ...)

\method{as.list}{DataSubject}(x, ...)

as_stan_list(object, ...)

\method{as_stan_list}{default}(object, ...)

\method{as_stan_list}{DataLongitudinal}(object, subject_var, ...)

\method{as.list}{DataLongitudinal}(x, ...)

\method{as_stan_list}{DataSurvival}(object, ...)

\method{as.list}{DataSurvival}(x, ...)

\method{as_stan_list}{DataJoint}(object, ...)

\method{as.list}{DataJoint}(x, ...)
}
\arguments{
\item{object}{(\code{DataSubject} or \code{DataLongitudinal} or \code{DataSurvival}) \cr
data object to convert to a \code{list}}

\item{...}{not used.}

\item{x}{(\code{DataSubject} or \code{DataLongitudinal} or \code{DataSurvival}) \cr
data object to convert to a \code{list}}

\item{subject_var}{(\code{character}) \cr the name of the variable
containing the subject identifier.}
}
\description{
Coerces a data object into a \code{list} of data components required
for fitting a \code{\link{JointModel}}. See the vignette (TODO) for more details.
}
\seealso{
Other DataSubject: 
\code{\link{DataSubject-class}},
\code{\link{as.data.frame.DataSubject}()},
\code{\link{extractVariableNames.DataSubject}()}

Other DataLongitudinal: 
\code{\link{DataLongitudinal-class}},
\code{\link{as.data.frame.DataLongitudinal}()},
\code{\link{extractVariableNames.DataLongitudinal}()}

Other DataSurvival: 
\code{\link{DataSurvival-class}},
\code{\link{as.data.frame.DataSurvival}()},
\code{\link{extractVariableNames.DataSurvival}()}

Other DataJoint: 
\code{\link{DataJoint-class}},
\code{\link{subset.DataJoint}()}
}
\concept{DataJoint}
\concept{DataLongitudinal}
\concept{DataSubject}
\concept{DataSurvival}
