% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalQuantities.R
\name{autoplot.LongitudinalQuantities}
\alias{autoplot.LongitudinalQuantities}
\title{Automatic Plotting for \code{LongitudinalQuantities}}
\usage{
\method{autoplot}{LongitudinalQuantities}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{(\code{LongitudinalQuantities}) longitudinal quantities.}

\item{conf.level}{(\code{numeric}) confidence level of the interval. If values of \code{FALSE},
\code{NULL} or \code{0} are provided then confidence regions will not be added to the plot.}

\item{...}{not used.}
}
\description{
Automatic Plotting for \code{LongitudinalQuantities}
}
\seealso{
Other LongitudinalQuantities:
\code{\link{LongitudinalQuantities-class}},
\code{\link[=as.data.frame.LongitudinalQuantities]{as.data.frame.LongitudinalQuantities()}},
\code{\link[=summary.LongitudinalQuantities]{summary.LongitudinalQuantities()}}

Other autoplot:
\code{\link{autoplot}},
\code{\link[=autoplot.SurvivalQuantities]{autoplot.SurvivalQuantities()}}
}
\concept{LongitudinalQuantities}
\concept{autoplot}
