% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalQuantities.R
\name{autoplot.LongitudinalQuantities}
\alias{autoplot.LongitudinalQuantities}
\title{Automatic Plotting for \code{LongitudinalQuantities}}
\usage{
\method{autoplot}{LongitudinalQuantities}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{(\code{\link{LongitudinalQuantities}}) \cr longitudinal quantities.}

\item{conf.level}{(\code{numeric}) \cr confidence level of the interval. If values of \code{FALSE},
\code{NULL} or \code{0} are provided then confidence regions will not be added to the plot.}

\item{...}{not used.}
}
\description{
Automatic Plotting for \code{LongitudinalQuantities}
}
\seealso{
Other LongitudinalQuantities: 
\code{\link{LongitudinalQuantities-class}},
\code{\link{as.data.frame.LongitudinalQuantities}()},
\code{\link{as_print_string.LongitudinalQuantities}()},
\code{\link{summary.LongitudinalQuantities}()}

Other autoplot: 
\code{\link{autoplot}()},
\code{\link{autoplot.SurvivalQuantities}()}
}
\concept{LongitudinalQuantities}
\concept{autoplot}
