% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LongitudinalSamples.R,
%   R/SurvivalSamples.R
\name{autoplot}
\alias{autoplot}
\alias{autoplot,LongitudinalSamples-method}
\alias{autoplot,SurvivalSamples-method}
\title{Plotting Methods for Different Classes}
\usage{
\S4method{autoplot}{LongitudinalSamples}(object, ...)

\S4method{autoplot}{SurvivalSamples}(object, add_km = TRUE, ...)
}
\arguments{
\item{object}{what to plot.}

\item{...}{other arguments passed to plotting methods.}

\item{add_km}{(\code{flag})\cr whether to add the Kaplan-Meier plot of the
survival data to the plots.}
}
\description{
These plot methods visualize various objects.
}
