% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LongitudinalSamples.R
\name{autoplot}
\alias{autoplot}
\alias{autoplot,LongitudinalSamples-method}
\title{Plotting Methods for Different Classes}
\usage{
\S4method{autoplot}{LongitudinalSamples}(object, ...)
}
\arguments{
\item{object}{(\code{LongitudinalSamples})\cr what to plot.}

\item{...}{Other arguments passed to plotting methods.}
}
\description{
These plot methods visualize various objects.
}
