% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-exports.R
\name{autoplot}
\alias{autoplot}
\title{Create a complete ggplot appropriate to a particular data type}
\description{
See \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} for details.
}
\concept{autoplot}
