% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{autoplot.SurvivalQuantities}
\alias{autoplot.SurvivalQuantities}
\title{Automatic Plotting for `SurvivalQuantities``}
\usage{
\method{autoplot}{SurvivalQuantities}(object, conf.level = 0.95, add_km = FALSE, add_wrap = TRUE, ...)
}
\arguments{
\item{object}{(\code{\link{SurvivalQuantities}}) \cr survival quantities.}

\item{conf.level}{(\code{numeric}) \cr confidence level of the interval. If values of \code{FALSE},
\code{NULL} or \code{0} are provided then confidence regions will not be added to the plot}

\item{add_km}{(\code{logical}) \cr if \code{TRUE} Kaplan-Meier curves will be added to the plot for
each group/patient.}

\item{add_wrap}{(\code{logical}) \cr if \code{TRUE} will apply a \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} to the plot
by each group/patient.}

\item{...}{not used.}
}
\description{
Automatic Plotting for `SurvivalQuantities``
}
\seealso{
Other SurvivalQuantities: 
\code{\link{SurvivalQuantities-class}},
\code{\link{as.data.frame.SurvivalQuantities}()},
\code{\link{as_print_string.SurvivalQuantities}()},
\code{\link{brierScore.SurvivalQuantities}()},
\code{\link{summary.SurvivalQuantities}()}

Other autoplot: 
\code{\link{autoplot}()},
\code{\link{autoplot.LongitudinalQuantities}()}
}
\concept{SurvivalQuantities}
\concept{autoplot}
