% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_score.R
\name{brier_score}
\alias{brier_score}
\alias{bs_get_squared_dist}
\alias{bs_get_weights}
\title{Brier Score}
\usage{
brier_score(
  t,
  times,
  events,
  pred_mat,
  maintain_cen_order = TRUE,
  event_offset = TRUE
)

bs_get_squared_dist(t, times, events, pred_mat)

bs_get_weights(
  t,
  times,
  events,
  event_offset = TRUE,
  maintain_cen_order = TRUE
)
}
\arguments{
\item{t}{(\code{numeric})\cr timepoints to calculate the desired quantity at.}

\item{times}{(\code{numeric})\cr observed times.}

\item{events}{(\code{numeric})\cr event indicator for \code{times}. Either 1 for an event or 0 for censor.}

\item{maintain_cen_order}{(\code{logical})\cr If \code{TRUE} then, in the case of ties,
censor times are always considered
to have occurred after the event times when calculating the "reverse Kaplan-Meier" for the
IPCW estimates. Setting this to \code{TRUE} mirrors the implementation of the \code{{prodlim}}
package.}

\item{event_offset}{(\code{logical})\cr If \code{TRUE} then \eqn{G(T_i)} is evaluated at \eqn{G(T_i-)}.
Setting this as \code{TRUE} mirrors the implementation of the \code{{pec}} package.}
}
\description{
Implements the Brier Score as detailed in Blanche et. al. (2015)
}
\details{
\itemize{
\item \code{bs_get_squared_dist()} - implements the squared distance part
of the formula.
\item \code{bs_get_weights()} - implements the IPCW weighting
}
}
\keyword{internal}
