% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{brierScore}
\alias{brierScore}
\title{\code{brierScore}}
\usage{
brierScore(object, ...)
}
\arguments{
\item{object}{to calculate Brier Score for.}

\item{...}{additional options.}
}
\description{
Returns the Brier Score for a given model
}
\seealso{
Other brierScore: 
\code{\link{brierScore.SurvivalQuantities}()}
}
\concept{brierScore}
