% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{brierScore.SurvivalQuantities}
\alias{brierScore.SurvivalQuantities}
\title{\code{brierScore}}
\usage{
\method{brierScore}{SurvivalQuantities}(object, maintain_cen_order = TRUE, event_offset = TRUE, ...)
}
\arguments{
\item{object}{to calculate Brier Score for.}

\item{...}{not used.}

\item{times}{TODO}
}
\description{
TODO
}
\seealso{
Other brierScore: 
\code{\link{brierScore}()}
}
\concept{brierScore}
