% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{brierScore.SurvivalQuantities}
\alias{brierScore.SurvivalQuantities}
\title{\code{brierScore}}
\usage{
\method{brierScore}{SurvivalQuantities}(object, maintain_cen_order = TRUE, event_offset = TRUE, ...)
}
\arguments{
\item{object}{(\code{\link{SurvivalQuantities}}) \cr survival quantities.}

\item{maintain_cen_order}{(\code{logical})\cr If \code{TRUE} then, in the case of ties,
censor times are always considered
to have occurred after the event times when calculating the "reverse Kaplan-Meier" for the
IPCW estimates. Setting this to \code{TRUE} mirrors the implementation of the \code{{prodlim}}
package.}

\item{event_offset}{(\code{logical})\cr If \code{TRUE} then \eqn{G(T_i)} is evaluated at \eqn{G(T_i-)}.
Setting this as \code{TRUE} mirrors the implementation of the \code{{pec}} package.}

\item{...}{not used.}
}
\description{
Derives the Brier Scores (using Inverse Probability of Censoring Weighting)
for the Survival estimates as detailed in \insertCite{blanche2015}{jmpost}.
}
\references{
\insertAllCited{}
}
\seealso{
Other brierScore: 
\code{\link{brierScore}()}

Other SurvivalQuantities: 
\code{\link{SurvivalQuantities-class}},
\code{\link{as.data.frame.SurvivalQuantities}()},
\code{\link{as_print_string.SurvivalQuantities}()},
\code{\link{autoplot.SurvivalQuantities}()},
\code{\link{summary.SurvivalQuantities}()}
}
\concept{SurvivalQuantities}
\concept{brierScore}
