% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/GridFixed.R, R/GridGrouped.R,
%   R/GridPopulation.R, R/GridPrediction.R
\name{coalesceGridTime}
\alias{coalesceGridTime}
\alias{coalesceGridTime.GridFixed}
\alias{coalesceGridTime.GridGrouped}
\alias{coalesceGridTime.GridPopulation}
\alias{coalesceGridTime.GridPrediction}
\title{Coalesce Time}
\usage{
coalesceGridTime(object, times, ...)

\method{coalesceGridTime}{GridFixed}(object, times, ...)

\method{coalesceGridTime}{GridGrouped}(object, times, ...)

\method{coalesceGridTime}{GridPopulation}(object, times, ...)

\method{coalesceGridTime}{GridPrediction}(object, times, ...)
}
\arguments{
\item{object}{(\code{\link{Grid}}) \cr object to coalesce time for.}

\item{times}{(\code{numeric}) \cr the times to coalesce to.}

\item{...}{Not used

Method used to replace NULL times on grid objects (if appropriate)}
}
\description{
Coalesce Time
}
\keyword{internal}
