% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StanModule.R, R/JointModel.R
\name{compileStanModel}
\alias{compileStanModel}
\alias{compileStanModel.StanModule}
\alias{compileStanModel.JointModel}
\title{\code{compileStanModel}}
\usage{
compileStanModel(object)

\method{compileStanModel}{StanModule}(object)

\method{compileStanModel}{JointModel}(object)
}
\arguments{
\item{object}{the module.}
}
\description{
Compile the Stan module.
}
