% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{compileStanModel}
\alias{compileStanModel}
\title{\code{compileStanModel}}
\usage{
compileStanModel(object, exe_file)
}
\arguments{
\item{object}{TODO}

\item{exe_file}{TODO}
}
\description{
Description - TODO
}
