% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{compileStanModel}
\alias{compileStanModel}
\title{\code{compileStanModel}}
\usage{
compileStanModel(object, exe_file)
}
\arguments{
\item{object}{the module.}

\item{exe_file}{(\code{string})\cr output file.}
}
\description{
Compile the Stan module.
}
