% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\name{covariates.DataSurvival}
\alias{covariates.DataSurvival}
\title{Extract Covariate Names for Survival Data}
\usage{
\method{covariates}{DataSurvival}(object, ...)
}
\arguments{
\item{object}{(\code{DataSurvival}) Survival data object to extract covariate names from.}
}
\description{
Extract Covariate Names for Survival Data
}
\seealso{
Other covariates:
\code{\link[=covariates]{covariates()}}
}
\concept{covariates}
