% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{covariates}
\alias{covariates}
\title{Extract Covariate Names}
\usage{
covariates(object, ...)
}
\arguments{
\item{object}{(\code{ANY}) the object to extract covariate names from.}

\item{...}{additional arguments added by methods.}
}
\description{
Extract Covariate Names
}
\seealso{
Other covariates:
\code{\link[=covariates.DataSurvival]{covariates.DataSurvival()}}
}
\concept{covariates}
