% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimJointData.R
\name{cut_data}
\alias{cut_data}
\title{Cut Study Data}
\usage{
cut_data(object, cut_time)
}
\arguments{
\item{object}{A \link{SimJointData} object}

\item{cut_time}{(\code{numeric})\cr A vector of cut off times, either length 1 for all patients or
\code{nrow(object@survival)} for a time per patient.}
}
\description{
Cut Study Data
}
\details{
All observations after this time are remove. Survival is censored at this time and any longitudinal
values are removed.
}
