% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\docType{class}
\name{DataJoint-class}
\alias{DataJoint-class}
\alias{.DataJoint}
\alias{DataJoint}
\title{Joint Data Object and Constructor Function}
\usage{
DataJoint(subject, survival = NULL, longitudinal = NULL)
}
\arguments{
\item{survival}{(\code{DataSurvival})\cr object created by \code{\link[=DataSurvival]{DataSurvival()}}.}

\item{longitudinal}{(\code{DataLongitudinal})\cr object created by \code{\link[=DataLongitudinal]{DataLongitudinal()}}.}
}
\description{
The \code{DataJoint} class handles combining data from a \code{\link{DataSurvival}} object and a
\code{\link{DataLongitudinal}} object.
}
\section{Slots}{

\describe{
\item{\code{survival}}{(\code{DataSurvival})\cr See Argument for details.}

\item{\code{longitudinal}}{(\code{DataLongitudinal})\cr See Argument for details.}
}}

\seealso{
Other DataObjects: 
\code{\link{DataLongitudinal-class}},
\code{\link{DataSurvival-class}}

Other DataJoint: 
\code{\link{as_stan_list.DataJoint}()},
\code{\link{subset.DataJoint}()}
}
\concept{DataJoint}
\concept{DataObjects}
