% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\docType{class}
\name{DataJoint-class}
\alias{DataJoint-class}
\alias{.DataJoint}
\alias{DataJoint}
\title{\code{DataJoint}}
\usage{
DataJoint(survival, longitudinal)
}
\arguments{
\item{survival}{(\code{DataSurvival})\cr object created by \code{\link[=DataSurvival]{DataSurvival()}}.}

\item{longitudinal}{(\code{DataLongitudinal})\cr object created by \code{\link[=DataLongitudinal]{DataLongitudinal()}}.}
}
\description{
The \code{DataJoint} class handles combining data from a \code{\link{DataSurvival}} object and a
\code{\link{DataLongitudinal}} object.
}
\details{
\itemize{
\item \code{as.list(x)}, \code{as(x, "list")}: Coerces x into a list of data components required
for fitting a \code{\link[=JointModel]{JointModel()}}. See the vignette (TODO) for more details.
}
}
\section{Slots}{

\describe{
\item{\code{survival}}{(\code{DataSurvival})\cr object created by \code{\link[=DataSurvival]{DataSurvival()}}.}

\item{\code{longitudinal}}{(\code{DataLongitudinal})\cr object created by \code{\link[=DataLongitudinal]{DataLongitudinal()}}.}
}}

