% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\name{DataJoint-Shared}
\alias{DataJoint-Shared}
\title{Re-used documentation for \code{DataJoint}}
\arguments{
\item{object}{(\code{\link{DataJoint}}) \cr Survival and Longitudinal Data.}

\item{x}{(\code{\link{DataJoint}}) \cr Survival and Longitudinal Data.}

\item{...}{Not Used.}
}
\description{
Re-used documentation for \code{DataJoint}
}
\keyword{internal}
