% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLongitudinal.R
\docType{class}
\name{DataLongitudinal-class}
\alias{DataLongitudinal-class}
\alias{.DataLongitudinal}
\alias{DataLongitudinal}
\title{Longitudinal Data Object and Constructor Function}
\usage{
DataLongitudinal(data, formula, threshold = NULL)
}
\arguments{
\item{data}{(\code{data.frame})\cr containing the observed longitudinal data.}

\item{formula}{(\code{formula})\cr of the form \code{outcome ~ time}, and cannot contain any additional covariates.}

\item{threshold}{(\code{numeric})\cr cut-off value to be used to declare an observation as censored
(below detection limit).}
}
\description{
The \code{\link{DataLongitudinal}} class handles the processing of the longitudinal data for fitting a Joint Model.
}
\section{Slots}{

\describe{
\item{\code{data}}{(\code{data.frame})\cr See Arguments for details; Note that
observations that contain missing values in the required variables are removed.}

\item{\code{formula}}{(\code{formula})\cr See Arguments for details}

\item{\code{threshold}}{(\code{numeric})\cr See Arguments for details}
}}

\seealso{
Other DataObjects: 
\code{\link{DataJoint-class}},
\code{\link{DataSurvival-class}}

Other DataLongitudinal: 
\code{\link{as.data.frame.DataLongitudinal}()},
\code{\link{as_stan_list.DataLongitudinal}()},
\code{\link{extractVariableNames.DataLongitudinal}()}
}
\concept{DataLongitudinal}
\concept{DataObjects}
