% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLongitudinal.R
\docType{class}
\name{DataLongitudinal-class}
\alias{DataLongitudinal-class}
\alias{.DataLongitudinal}
\alias{DataLongitudinal}
\alias{as.data.frame,DataLongitudinal-method}
\alias{as.list,DataLongitudinal-method}
\title{\code{DataLongitudinal}}
\usage{
DataLongitudinal(data, formula, subject, threshold = NULL, time_grid = NULL)

\S4method{as.data.frame}{DataLongitudinal}(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.list}{DataLongitudinal}(x)
}
\arguments{
\item{data}{(\code{data.frame})\cr containing the observed longitudinal data.}

\item{formula}{(\code{formula})\cr of the form \code{outcome ~ time}, and cannot contain any additional covariates.}

\item{subject}{(\code{string})\cr the name of the subject identifier variable.}

\item{threshold}{(\code{number})\cr cut-off value to be used to declare an observation as censored
(below detection limit).}

\item{time_grid}{(\code{numeric})\cr grid of time points to use for providing samples
of the longitudinal model fit functions. If \code{NULL}, will be taken as a sequence of
201 values from the minimum to the maximum observed times.}

\item{x}{(\code{DataLongitudinal})\cr to be coerced.}

\item{row.names}{not used.}

\item{optional}{not used.}
}
\description{
The \code{\link{DataLongitudinal}} class handles the processing of the longitudinal data for fitting a Joint Model.
}
\details{
\itemize{
\item \code{as.list(x)}, \code{as(x, "list")}: Coerces x into a list of data components required
for fitting a \code{\link{JointModel}}.
See the vignette (TODO) for more details
\item \code{as.data.frame(x)}
}
}
\section{Slots}{

\describe{
\item{\code{data}}{(\code{data.frame})\cr containing the observed longitudinal data. Note that
observations that contain missing values in the required variables are removed
in the slot.}

\item{\code{formula}}{(\code{formula})\cr of the form \code{outcome ~ time}, and cannot contain any additional covariates.}

\item{\code{subject}}{(\code{string})\cr the name of the subject identifier variable.}

\item{\code{threshold}}{(\code{number})\cr cut-off value to be used to declare an observation as censored
(below detection limit).}

\item{\code{time_grid}}{(\code{numeric})\cr grid of time points to use for providing samples
of the longitudinal model fit functions. If \code{NULL}, will be taken as a sequence of
201 values from the minimum to the maximum observed times.}
}}

\seealso{
\code{\link{DataJoint}}, \code{\link{DataSurvival}}.
}
