% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\docType{class}
\name{DataSurvival-class}
\alias{DataSurvival-class}
\alias{.DataSurvival}
\alias{DataSurvival}
\title{\code{DataSurvival}}
\usage{
DataSurvival(data, formula, subject, arm, study)
}
\arguments{
\item{data}{(\code{data.frame})\cr the observed time-to-event data.}

\item{formula}{(\code{formula})\cr of the form \code{Surv(time, event) ~ cov1 + cov2 + ...}.
See \code{\link[survival:Surv]{survival::Surv()}} for more details, though note that this package only supports right censoring.}

\item{subject}{(\code{string})\cr the name of the subject identifier variable.}

\item{arm}{(\code{string})\cr the name of the treatment arm variable.}

\item{study}{(\code{string})\cr the name of the study variable.}
}
\description{
The \code{\link{DataSurvival}} class handles the processing of the survival data for fitting a Joint Model.
}
\section{Slots}{

\describe{
\item{\code{data}}{(\code{data.frame})\cr the observed time-to-event data. Note that
observations that contain missing values in the required variables are removed
in the slot.}

\item{\code{formula}}{(\code{formula})\cr of the form \code{Surv(time, event) ~ cov1 + cov2 + ...}.
See \code{\link[survival:Surv]{survival::Surv()}} for more details, though note that this package only supports right censoring.}

\item{\code{subject}}{(\code{string})\cr the name of the subject identifier variable.}

\item{\code{arm}}{(\code{string})\cr the name of the treatment arm variable.}

\item{\code{study}}{(\code{string})\cr the name of the study variable.}
}}

\seealso{
\code{\link{DataJoint}}, \code{\link{DataLongitudinal}}
}
