% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\docType{class}
\name{DataSurvival-class}
\alias{DataSurvival-class}
\alias{.DataSurvival}
\alias{DataSurvival}
\title{Survival Data Object and Constructor Function}
\usage{
DataSurvival(data, formula)
}
\arguments{
\item{data}{(\code{data.frame})\cr the observed time-to-event data.}

\item{formula}{(\code{formula})\cr of the form \code{Surv(time, event) ~ cov1 + cov2 + ...}.
See \code{\link[survival:Surv]{survival::Surv()}} for more details, though note that this package only supports right censoring.}
}
\description{
The \code{\link{DataSurvival}} class handles the processing of the survival data for fitting a Joint Model.
}
\section{Slots}{

\describe{
\item{\code{data}}{(\code{data.frame})\cr See Arguments for details.}

\item{\code{formula}}{(\code{formula})\cr See Arguments for details.}
}}

\seealso{
Other DataObjects: 
\code{\link{DataJoint-class}},
\code{\link{DataLongitudinal-class}}

Other DataSurvival: 
\code{\link{as.data.frame.DataSurvival}()},
\code{\link{as_stan_list.DataSurvival}()},
\code{\link{extractVariableNames.DataSurvival}()}
}
\concept{DataObjects}
\concept{DataSurvival}
