% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{decompose_patients}
\alias{decompose_patients}
\title{Decompose Patients into Relevant Components}
\usage{
decompose_patients(patients, all_pts)
}
\arguments{
\item{patients}{(\code{character} or \code{list})\cr patient identifiers. If \code{NULL} will be set to \code{all_pts}.}

\item{all_pts}{(\code{character})\cr the set of allowable patient identifiers.
Will cause an error if any value of \code{patients} is not in this vector.}
}
\value{
A list containing three components:
\itemize{
\item \code{groups}: (\code{list})\cr each element of the list is a character vector
specifying which patients belong to a given "group" where the "group" is the element name
\item \code{unique_values}: (\code{character})\cr vector of the unique patients within \code{patients}
\item \code{indexes}: (\code{list})\cr each element is a named and is a numeric index vector
that maps the values of \code{grouped} to \code{unique_values}
}
}
\description{
This function takes in a character vector or list of patients and decomposes it into a
structured format.
}
\details{
The primary use of this function is to correctly setup indexing variables for
predicting survival quantities (see \code{\link[=SurvivalSamples-class]{predict(SurvivalSamples)}})
}
\examples{
\dontrun{
result <- decompose_patients(c("A", "B"), c("A", "B", "C", "D"))
result <- decompose_patients(
    list("g1" = c("A", "B"), "g2" = c("B", "C")),
    c("A", "B", "C", "D")
)
}
}
\seealso{
\code{\link[=expand_patients]{expand_patients()}}, \code{\link[=SurvivalSamples-class]{predict(SurvivalSamples)}}
}
\keyword{internal}
