% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\name{dim.Quantities}
\alias{dim.Quantities}
\title{Dimensions of \code{Quantitites}}
\usage{
\method{dim}{Quantities}(x)
}
\arguments{
\item{x}{(\code{\link{Quantities}}) \cr Any generated model quantity

Returns the number of rows and columns within any given
element of the \link{Quantities} object. As each element is of the same
dimensionality this just returns a single number for the rows and columns
of all elements of the \link{Quantities}}
}
\description{
Dimensions of \code{Quantitites}
}
\keyword{internal}
