% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModel.R
\name{ensure_initial_values}
\alias{ensure_initial_values}
\title{Ensure that initial values are correctly specified}
\usage{
ensure_initial_values(initial_values, data, parameters)
}
\arguments{
\item{initial_values}{(\code{list})\cr A list of lists containing the initial values
must be 1 list per desired chain. All elements should have identical names}

\item{data}{(\code{list})\cr specifies the size to expand each of our initial values to be.
That is elements of size 1 in \code{initial_values} will be expanded to be the same
size as the corresponding element in \code{data} by broadcasting the value.}

\item{parameters}{(\code{\link{ParameterList}})\cr the parameters object}
}
\description{
Ensure that initial values are correctly specified
}
\details{
This function is mostly a thin wrapper around \code{expand_initial_values} to
enable easier unit testing.
}
\keyword{internal}
