% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_initial_values}
\alias{expand_initial_values}
\title{Replicate Single Values in a List}
\usage{
expand_initial_values(initial_values, sizes)
}
\arguments{
\item{initial_values}{(\code{list})\cr initial values with names.}

\item{sizes}{(\code{list})\cr each size corresponds to an element in \code{initial_values},
matched by the names.}
}
\value{
A named list of values, with any single values in the \code{initial_values} list
replicated according to the corresponding values in the \code{sizes} list.
The resulting list has the same names as the original lists.
}
\description{
Replicate Single Values in a List
}
\keyword{internal}
