% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/DataLongitudinal.R,
%   R/DataSurvival.R
\name{extractVariableNames}
\alias{extractVariableNames}
\alias{extractVariableNames,DataLongitudinal-method}
\alias{extractVariableNames,DataSurvival-method}
\title{\code{extractVariableNames}}
\usage{
extractVariableNames(object, ...)

\S4method{extractVariableNames}{DataLongitudinal}(object)

\S4method{extractVariableNames}{DataSurvival}(object, ...)
}
\arguments{
\item{object}{the module.}
}
\value{
A \code{list} of variable names mapping to key variables with the
source-data stored inside the data class.
}
\description{
\code{extractVariableNames}
}
\keyword{internal}
