% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extractVariableNames}
\alias{extractVariableNames}
\title{Extract Mapping to Standardised Variable Names}
\usage{
extractVariableNames(object)
}
\arguments{
\item{object}{the data object.}
}
\description{
Extract a \code{list} that maps the variable names in a user-defined
\code{data.frame} to standardised values.
}
\seealso{
Other extractVariableNames: 
\code{\link{extractVariableNames.DataLongitudinal}()},
\code{\link{extractVariableNames.DataSubject}()},
\code{\link{extractVariableNames.DataSurvival}()}
}
\concept{extractVariableNames}
\keyword{internal}
