% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModelSamples.R
\name{generateQuantities}
\alias{generateQuantities}
\alias{generateQuantities,JointModelSamples-method}
\title{\code{generateQuantities}}
\usage{
generateQuantities(object, ...)

\S4method{generateQuantities}{JointModelSamples}(object, time_grid_lm = NULL, time_grid_sm = NULL, ...)
}
\arguments{
\item{object}{object to obtain generated quantities from}

\item{...}{additional options.}

\item{time_grid_lm}{(\code{numeric})\cr grid of time points to use for providing samples
of the longitudinal model fit functions. If \code{NULL}, will be taken as a sequence of
201 values from 0 to the maximum observed event time.}

\item{time_grid_os}{(\code{numeric})\cr grid of time points to use for providing samples
of the survival model fit functions. If \code{NULL}, will be taken as a sequence of
201 values from 0 to the maximum observed event time.}
}
\description{
Obtain the generated quantities from a Stan Model.
}
