% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModelSamples.R
\name{generateQuantities}
\alias{generateQuantities}
\alias{generateQuantities,JointModelSamples-method}
\title{\code{generateQuantities}}
\usage{
generateQuantities(object, ...)

\S4method{generateQuantities}{JointModelSamples}(object, patients, time_grid_lm, time_grid_sm, ...)
}
\arguments{
\item{object}{object to obtain generated quantities from}

\item{...}{additional options.}

\item{patients}{(\code{character})\cr explicit vector of patient IDs for whom the
generated quantities should be extracted}

\item{time_grid_lm}{(\code{numeric})\cr grid of time points to use for providing samples
of the longitudinal model fit functions. If \code{NULL}, will be taken as a sequence of
201 values from 0 to the maximum observed event time.}

\item{time_grid_os}{(\code{numeric})\cr grid of time points to use for providing samples
of the survival model fit functions. If \code{NULL}, will be taken as a sequence of
201 values from 0 to the maximum observed event time.}
}
\description{
Obtain the generated quantities from a Stan Model.
}
