% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModelSamples.R
\name{generateQuantities}
\alias{generateQuantities}
\alias{generateQuantities.JointModelSamples}
\title{\code{generateQuantities}}
\usage{
generateQuantities(object, ...)

\method{generateQuantities}{JointModelSamples}(object, generator, type, ...)
}
\arguments{
\item{object}{object to obtain generated quantities from}

\item{...}{additional options.}

\item{type}{(\code{character})\cr type of quantities to be generated, must be either "survial" or
"longitudinal".}

\item{patients}{(\code{character})\cr explicit vector of patient IDs for whom the
generated quantities should be extracted}

\item{times}{TODO}
}
\description{
Obtain the generated quantities from a Stan Model.
}
