% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModelSamples.R
\name{generateQuantities}
\alias{generateQuantities}
\alias{generateQuantities.JointModelSamples}
\title{\code{generateQuantities}}
\usage{
generateQuantities(object, ...)

\method{generateQuantities}{JointModelSamples}(object, generator, type, ...)
}
\arguments{
\item{object}{object to obtain generated quantities from}

\item{...}{additional options.}

\item{generator}{(\code{QuantityGenerator})\cr object that specifies which subjects and time points
to calculate the quantities at}

\item{type}{(\code{character})\cr type of quantities to be generated, must be either "survival" or
"longitudinal".}
}
\description{
Obtain the generated quantities from a Stan Model.
}
