% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\name{get_arm_study_data}
\alias{get_arm_study_data}
\title{\code{get_arm_study_data}}
\usage{
get_arm_study_data(pt, arm, study)
}
\arguments{
\item{pt}{a factor vector of subject identifiers}

\item{arm}{a factor vector of treatment arms. Must be same length as \code{pt}}

\item{study}{a factor vector of study identifiers. Must be same length as \code{pt}

The function is purely a sub-component of \code{as.list(DataSurvival)}. Its
purpose is to extract key logic into its own function to enable unit tests}
}
\value{
TODO
}
\description{
\code{get_arm_study_data}
}
\keyword{internal}
