% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StanModel.R
\name{getParameters}
\alias{getParameters}
\alias{getParameters,StanModel-method}
\title{\code{getParameters}}
\usage{
getParameters(object, ...)

\S4method{getParameters}{StanModel}(object)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\description{
Description - TODO
}
\keyword{internal}
