% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StanModel.R
\name{getParameters}
\alias{getParameters}
\alias{getParameters,StanModel-method}
\title{\code{getParameters}}
\usage{
getParameters(object)

\S4method{getParameters}{StanModel}(object)
}
\arguments{
\item{object}{where to obtain the parameters from.}
}
\description{
Obtain the parameters from a \code{\link{StanModel}}.
}
\keyword{internal}
