% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StanModel.R, R/LinkComponent.R,
%   R/Link.R, R/defaults.R
\name{getParameters}
\alias{getParameters}
\alias{getParameters.StanModel}
\alias{getParameters.LinkComponent}
\alias{getParameters.Link}
\alias{getParameters.default}
\title{\code{getParameters}}
\usage{
getParameters(object, ...)

\method{getParameters}{StanModel}(object)

\method{getParameters}{LinkComponent}(object, ...)

\method{getParameters}{Link}(object, ...)

\method{getParameters}{default}(object)
}
\arguments{
\item{object}{where to obtain the parameters from.}

\item{...}{additional options.}
}
\description{
Extract any modelling parameters as a \code{\link{ParameterList}} object
from a model.
}
\seealso{
Other LinkComponent: 
\code{\link{LinkComponent-class}},
\code{\link{as.StanModule.LinkComponent}()},
\code{\link{as.list.LinkComponent}()},
\code{\link{initialValues}()}
}
\concept{LinkComponent}
