% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LongitudinalClaretBruno.R,
%   R/LongitudinalGSF.R, R/LongitudinalRandomSlope.R, R/LongitudinalSteinFojo.R
\name{getPredictionNames}
\alias{getPredictionNames}
\alias{getPredictionNames.default}
\alias{getPredictionNames.LongitudinalClaretBruno}
\alias{getPredictionNames.LongitudinalGSF}
\alias{getPredictionNames.LongitudinalRandomSlope}
\alias{getPredictionNames.LongitudinalSteinFojo}
\title{Get Prediction Names}
\usage{
getPredictionNames(object, ...)

\method{getPredictionNames}{default}(object, ...)

\method{getPredictionNames}{LongitudinalClaretBruno}(object, ...)

\method{getPredictionNames}{LongitudinalGSF}(object, ...)

\method{getPredictionNames}{LongitudinalRandomSlope}(object, ...)

\method{getPredictionNames}{LongitudinalSteinFojo}(object, ...)
}
\arguments{
\item{object}{(\code{LongitudinalModel}) \cr A longitudinal model object}

\item{...}{Not used.}
}
\description{
Utility function that returns the names of the required parameters for predicting
survival quantities with \code{\link{GridPrediction}}.
}
