% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R, R/GridFixed.R, R/GridGrouped.R,
%   R/GridObserved.R
\name{Grid-Functions}
\alias{Grid-Functions}
\alias{GridFixed}
\alias{GridGrouped}
\alias{GridObserved}
\title{Quantity Grid Specification}
\usage{
GridFixed(subjects = NULL, times = NULL)

GridGrouped(groups, times = NULL)

GridObserved(subjects = NULL)
}
\arguments{
\item{subjects}{(\code{character} or \code{NULL})\cr vector of subjects to extract quantities for.
If \code{NULL} will default to all subjects within the dataset.}

\item{times}{(\code{numeric} or \code{NULL})\cr vector of time points to extract quantities at.
If \code{NULL} will default to 201 evenly spaced timepoints between 0 and either the max
observation time (for \code{\link{LongitudinalQuantities}}) or max event time (for \code{\link{SurvivalQuantities}}).}

\item{groups}{(\code{list})\cr named list of subjects to extract quantities for. See Group Specification.}
}
\description{
These functions are used to specify which subjects and timepoints should be generated
when calculating quantities via \code{\link{SurvivalQuantities}} and \code{\link{LongitudinalQuantities}}.
}
\section{Group Specification}{

For \code{GridGrouped()}, \code{groups} must be a named list of character vectors. Each element of the list
must be a character vector of the subjects that will form the group where the element name
is the corresponding name of the group. For example if the goal was to create two groups
named \code{Group-1} and \code{Group-2} which are composed of the subjects \code{pt-1}, \code{pt-2} and
\code{pt-3}, \code{pt-4} respectively then this would be specified as:

\if{html}{\out{<div class="sourceCode">}}\preformatted{GridGrouped(
    groups = list(
        "Group-1" = c("pt-1", "pt-2"),
        "Group-2" = c("pt-3", "pt-4")
    )
)
}\if{html}{\out{</div>}}
}

\seealso{
\code{\link{SurvivalQuantities}}, \code{\link{LongitudinalQuantities}}
}
