% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R, R/GridEven.R, R/GridFixed.R,
%   R/GridGrouped.R, R/GridManual.R, R/GridObserved.R, R/GridPopulation.R,
%   R/GridPrediction.R
\name{Grid-Functions}
\alias{Grid-Functions}
\alias{GridEven}
\alias{GridFixed}
\alias{GridGrouped}
\alias{GridManual}
\alias{GridObserved}
\alias{GridPopulation}
\alias{GridPrediction}
\title{Quantity Grid Specification}
\usage{
GridEven(subjects = NULL, length.out = 30)

GridFixed(subjects = NULL, times = NULL)

GridGrouped(groups, times = NULL)

GridManual(spec)

GridObserved(subjects = NULL)

GridPopulation(times = NULL)

GridPrediction(times = NULL, newdata, params = list())
}
\arguments{
\item{subjects}{(\code{character} or \code{NULL})\cr vector of subjects to extract quantities for.
If \code{NULL} will default to all subjects within the dataset.}

\item{length.out}{(\code{numeric})\cr number of evenly spaced timepoints to generate quantities at.}

\item{times}{(\code{numeric} or \code{NULL})\cr vector of time points to extract quantities at.
If \code{NULL} will default to 201 evenly spaced timepoints between 0 and either the max
observation time (for \code{\link{LongitudinalQuantities}}) or max event time (for \code{\link{SurvivalQuantities}}).}

\item{groups}{(\code{list})\cr named list of subjects to extract quantities for. See Group Specification.}

\item{spec}{(\code{list})\cr named list of subjects to extract quantities for. The names of each
element should be the required subjects with the element itself being a numeric vector of timepoints
to generate the quantity at.}

\item{newdata}{(\code{data.frame}) \cr new data to generate quantities for. Must contain the same columns
and factor levels of the original data used in the \code{\link{DataSurvival}} object.}

\item{params}{(\code{list})\cr named list of parameters to fix the longitudinal model parameters at when
predicting survival quantities. See \code{\link[=getPredictionNames]{getPredictionNames()}} for the required parameters.}
}
\description{
These functions are used to specify which subjects and timepoints should be generated
when calculating quantities via \code{\link{SurvivalQuantities}} and \code{\link{LongitudinalQuantities}}.
}
\details{
\itemize{
\item \code{GridFixed()} is used to specify a fixed set of timepoints to generate quantities at for
all the specified subjects.
\item \code{GridGrouped()} is similar to \code{GridFixed()} but allows for groupwise averaging
(see Group Specification).
\item \code{GridObserved()} generates quantities at the observed longitudinal timepoints for each
subject.
\item \code{GridManual()} allows for individual timepoint specification for each subject.
\item \code{GridEven()} generates quantities for each subject at N evenly spaced timepoints
between each subjects first and last longitudinal observations.
\item \code{GridPopulation()} generates longitudinal model quantities based on the population parameters at the
specified time points. Generates 1 set of quantities for each distinct combination of \code{arm}
and \code{study} within the \code{\link{DataSubject}} object provided to the \code{\link{JointModel}}.
\item \code{GridPrediction()} generates survival quantities based on any user-defined values at the
specified time points. This is useful for generating quantities for a new dataset
on specific longitudinal model parameters. See \code{\link[=getPredictionNames]{getPredictionNames()}} to determine which
longitudinal model parameters need to be defined for a given longitudinal model.
}
}
\section{Group Specification}{

For \code{GridGrouped()}, \code{groups} must be a named list of character vectors. Each element of the list
must be a character vector of the subjects that will form the group where the element name
is the corresponding name of the group. For example if the goal was to create two groups
named \code{Group-1} and \code{Group-2} which are composed of the subjects \code{sub-1}, \code{sub-2} and
\code{sub-3}, \code{sub-4} respectively then this would be specified as:

\if{html}{\out{<div class="sourceCode">}}\preformatted{GridGrouped(
    groups = list(
        "Group-1" = c("sub-1", "sub-2"),
        "Group-2" = c("sub-3", "sub-4")
    )
)
}\if{html}{\out{</div>}}
}

\seealso{
\code{\link{SurvivalQuantities}}, \code{\link{LongitudinalQuantities}}
}
