% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gsf_sld}
\alias{gsf_sld}
\alias{gsf_ttg}
\alias{gsf_dsld}
\title{Generalized Stein-Fojo Functionals}
\usage{
gsf_sld(time, b, s, g, phi)

gsf_ttg(time, b, s, g, phi)

gsf_dsld(time, b, s, g, phi)
}
\arguments{
\item{time}{(\code{numeric})\cr time grid.}

\item{b}{(\code{number})\cr baseline.}

\item{s}{(\code{number})\cr shrinkage.}

\item{g}{(\code{number})\cr growth.}

\item{phi}{(\code{number})\cr shrinkage proportion.}
}
\value{
The function results.
}
\description{
Generalized Stein-Fojo Functionals
}
\examples{
gsf_sld(1:10, 20, 0.3, 0.6, 0.2)
gsf_ttg(1:10, 20, 0.3, 0.6, 0.2)
gsf_dsld(1:10, 20, 0.3, 0.6, 0.2)
}
\keyword{internal}
