% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HazardLink.R
\docType{class}
\name{HazardLink-class}
\alias{HazardLink-class}
\alias{HazardLink}
\title{Hazard Link Class}
\description{
TODO - Description
}
\section{Slots}{

\describe{
\item{\code{stan}}{TODO}

\item{\code{contribution}}{TODO}

\item{\code{parameter}}{TODO}
}}

