% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSurvival.R
\name{hazardWindows.SimSurvival}
\alias{hazardWindows.SimSurvival}
\title{Construct Time Intervals}
\usage{
\method{hazardWindows}{SimSurvival}(object)
}
\arguments{
\item{x}{(\code{numeric})\cr grid of time points.}
}
\value{
A \code{tibble} with \code{lower}, \code{upper}, \code{time}, \code{eval} and \code{width}.
}
\description{
Construct Time Intervals
}
\keyword{internal}
