% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{initialValues}
\alias{initialValues}
\title{\code{initialValues}}
\usage{
initialValues(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\value{
a \code{list} of initial values to be passed to the stan sampler
}
\description{
\code{initialValues}
}
\keyword{internal}
