% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModel.R, R/Link.R
\name{initialValues}
\alias{initialValues}
\alias{initialValues.JointModel}
\alias{initialValues.Link}
\title{\code{initialValues}}
\usage{
initialValues(object)

\method{initialValues}{JointModel}(object)

\method{initialValues}{Link}(object)
}
\arguments{
\item{object}{where to get the initial values from.}
}
\description{
Obtain the \code{list} of initial values to be passed to the Stan sampler.
}
\keyword{internal}
