% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\name{is_file}
\alias{is_file}
\title{Is String a Valid File?}
\usage{
is_file(filename = NULL)
}
\arguments{
\item{filename}{(\code{string})\cr file name.}
}
\description{
A utility function to check if a string is a valid file or not.
Used to help address short comings of \code{\link[=file.exists]{file.exists()}} that will return \code{TRUE}
for a directory as well as a file.
}
\keyword{internal}
