% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\name{is_file}
\alias{is_file}
\title{Is string a valid file}
\usage{
is_file(filename = NULL)
}
\arguments{
\item{filename}{A character string}
}
\description{
A utility function to check if a string is a valid file or not.
Used to help address short comings of file.exists that will return TRUE
for a directory as well as a file
}
