% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmpost-package.R
\docType{package}
\name{jmpost-package}
\alias{jmpost}
\alias{jmpost-package}
\title{jmpost: TODO}
\description{
TODO
}
\author{
\strong{Maintainer}: Craig Gower-Page \email{craig.gower-page@roche.com}

Authors:
\itemize{
  \item Georgios Kazantzidis \email{georgios.kazantzidis@roche.com}
  \item Xiaoyan Fang \email{xiaoyan.fang@roche.com}
  \item Daniel Sabanes Bove \email{daniel.sabanes_bove@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
