% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmpost-package.R
\docType{package}
\name{jmpost-package}
\alias{jmpost}
\alias{jmpost-package}
\title{jmpost: Joint Models for Predicting Overall Survival Trajectories}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements joint models combining a non-linear mixed effects model for longitudinal measurements with a time-to-event model for the survival outcome. The acronym 'jmpost' stands for Joint Models for Predicting Overall Survival Trajectories, which refers to the application. In phase 1 studies patients are followed until progression only. Thus, reliable overall survival data and hence estimates are not available. However, we can use additional information from previous clinical trials or real-world data - we can correlate the tumor response data, that are longitudinal measurements, with the overall survival of the patients or their hazard ratios. Thereby we can predict the overall survival from our phase 1 study data and therefore make better decisions.
}
\author{
\strong{Maintainer}: Craig Gower-Page \email{craig.gower-page@roche.com}

Authors:
\itemize{
  \item Georgios Kazantzidis \email{georgios.kazantzidis@roche.com}
  \item Daniel Sabanes Bove \email{daniel.sabanes_bove@roche.com}
  \item Xiaoyan Fang \email{xiaoyan.fang@roche.com}
  \item Francois Mercier \email{francois.mercier@roche.com}
  \item Isaac Gravestock \email{isaac.gravestock@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
