% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{jmpost-settings}
\alias{jmpost-settings}
\title{jmpost settings}
\description{
Define settings that modify the behaviour of the \code{jmpost} package

Each of the following are the name of options that can be set via:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(<option_name> = <value>)
}\if{html}{\out{</div>}}
\subsection{\code{jmpost.prior_shrinkage}}{

Default = \code{0.5}

By default all initial values are drawn as random sample from the respective prior
distribution with a shrinkage factor towards the mean. That is:

\if{html}{\out{<div class="sourceCode">}}\preformatted{initial_value = prior_sample * prior_shrinkage + (1 - prior_shrinkage) * prior_mean
}\if{html}{\out{</div>}}

This setting controls the shrinkage factor. A value of 0 means no shrinkage (i.e.
pure random draw) whilst a value of 1 means the initial value is just the mean.
}
}
\examples{
\dontrun{
options(jmpost.prior_shrinkage = 0.5)
}
}
