% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModel.R
\docType{class}
\name{JointModel-class}
\alias{JointModel-class}
\alias{.JointModel}
\alias{JointModel}
\title{Joint Model Object and Constructor Function}
\usage{
JointModel(longitudinal = NULL, survival = NULL, link = Link())
}
\arguments{
\item{longitudinal}{(\code{LongitudinalModel} or \code{NULL}) the longitudinal model.}

\item{survival}{(\code{SurvivalModel} or \code{NULL}) the survival model.}

\item{link}{(\code{Link}) the link.}
}
\description{
Joint Model Object and Constructor Function
}
\section{Slots}{

\describe{
\item{\code{longitudinal}}{(\code{\link{LongitudinalModel}} or \code{NULL})\cr the longitudinal model.}

\item{\code{survival}}{(\code{\link{SurvivalModel}} or \code{NULL})\cr the survival model.}

\item{\code{link}}{(\code{Link})\cr the link.}

\item{\code{parameters}}{(\code{ParameterList})\cr the parameter declaration and prior specification.}
}}

\seealso{
Other JointModel:
\code{\link[=as.StanModule.JointModel]{as.StanModule.JointModel()}},
\code{\link[=as.character.JointModel]{as.character.JointModel()}}
}
\concept{JointModel}
