% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModel.R
\docType{class}
\name{JointModel-class}
\alias{JointModel-class}
\alias{.JointModel}
\alias{JointModel}
\title{\code{JointModel}}
\usage{
JointModel(longitudinal = NULL, survival = NULL, link = NULL)
}
\arguments{
\item{longitudinal}{(\code{LongitudinalModel} or \code{NULL})\cr the longitudinal model.}

\item{survival}{(\code{SurvivalModel} or \code{NULL})\cr the survival model.}

\item{link}{(\code{Link})\cr the link.}
}
\description{
\code{JointModel}
}
\section{Slots}{

\describe{
\item{\code{stan}}{(\code{StanModule})\cr code containing the joint model specification.}

\item{\code{parameters}}{(\code{ParameterList})\cr the parameter specification.}
}}

