% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModelSamples.R
\docType{class}
\name{JointModelSamples-class}
\alias{JointModelSamples-class}
\alias{.JointModelSamples}
\title{Samples from a \code{JointModel}}
\description{
Samples from a \code{JointModel}
}
\section{Slots}{

\describe{
\item{\code{model}}{(\code{JointModel})\cr the original model.}

\item{\code{data}}{(\code{list})\cr data input.}

\item{\code{init}}{(\code{list})\cr initial values.}

\item{\code{results}}{(\code{CmdStanMCMC})\cr the results from \code{\link[=sampleStanModel]{sampleStanModel()}}.}
}}

