% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModelSamples.R
\docType{class}
\name{JointModelSamples-class}
\alias{JointModelSamples-class}
\alias{.JointModelSamples}
\alias{JointModelSamples}
\title{\code{JointModelSamples}}
\description{
Contains samples from a \code{\link{JointModel}}.
}
\section{Slots}{

\describe{
\item{\code{model}}{(\code{\link{JointModel}})\cr the model that the samples were drawn from.}

\item{\code{data}}{(\code{\link{DataJoint}})\cr the data that the model was fitted on.}

\item{\code{results}}{(\code{\link{CmdStanMCMC}})\cr the STAN samples.}
}}

