% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link.R
\docType{class}
\name{Link-class}
\alias{Link-class}
\alias{.Link}
\alias{Link}
\title{\code{Link}}
\usage{
Link(stan = StanModule(), parameters = ParameterList(), ...)
}
\arguments{
\item{stan}{(\code{StanModule})\cr code containing the Stan code specification.}

\item{parameters}{(\code{ParameterList})\cr the parameter specification.}

\item{...}{additional arguments passed to the constructor.}
}
\description{
\code{Link}
}
\section{Slots}{

\describe{
\item{\code{stan}}{(\code{StanModule})\cr code containing the link specification.}

\item{\code{parameters}}{(\code{ParameterList})\cr the parameter specification.}
}}

