% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkGSF.R
\docType{class}
\name{link_gsf_abstract-class}
\alias{link_gsf_abstract-class}
\alias{.link_gsf_abstract}
\alias{link_gsf_abstract}
\title{\code{link_gsf_abstract}}
\usage{
link_gsf_abstract(stan, parameter, parameter_name, contribution_fname)
}
\arguments{
\item{stan}{(\code{StanModule})\cr code containing the Stan code specification.}

\item{parameter_name}{(\code{character})\cr the name of the parameter.}

\item{contribution_fname}{(\code{character})\cr the function name of the contribution.}
}
\description{
Class to specify the contents for \code{\link{LinkGSF}}.
}
\section{Slots}{

\describe{
\item{\code{stan}}{(\code{StanModule})\cr the Stan code.}

\item{\code{parameter}}{(\code{ParameterList})\cr the parameter specification.}

\item{\code{parameter_name}}{(\code{character})\cr the name of the parameter.}

\item{\code{contribution_fname}}{(\code{character})\cr the function name of the contribution.}
}}

\note{
Only the \code{functions} part of \code{stan} will be used.
}
