% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkComponent.R
\docType{class}
\name{LinkComponent-class}
\alias{LinkComponent-class}
\alias{.LinkComponent}
\alias{LinkComponent}
\title{\code{Link}}
\usage{
LinkComponent(stan, parameters = ParameterList(), key = "", ...)
}
\description{
\code{Link}
}
\section{Slots}{

\describe{
\item{\code{stan}}{(\code{StanModule})\cr code containing the link specification.}

\item{\code{parameters}}{(\code{ParameterList})\cr the parameter specification.}

\item{\code{name}}{(\code{character})\cr display name for the link object.}
}}

