% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkGSF.R
\docType{class}
\name{LinkGSF-class}
\alias{LinkGSF-class}
\alias{.LinkGSF}
\alias{LinkGSF}
\title{\code{LinkGSF}}
\usage{
LinkGSF(components = list(link_gsf_dsld(), link_gsf_ttg()))
}
\arguments{
\item{components}{(\code{list})\cr which link components should be included.}
}
\description{
This class extends the general \code{\link{Link}} class for the \code{\link{LongitudinalGSF}}
model.
}
