% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Local_Sample}
\alias{Local_Sample}
\alias{local_rnorm}
\alias{local_rnorm_vector}
\alias{local_rhorseshoe}
\alias{local_rcauchy}
\alias{local_rgamma}
\alias{local_rlnorm}
\alias{local_rbeta}
\alias{local_runif}
\alias{local_rt}
\alias{local_rlogis}
\alias{local_rloglogis}
\alias{local_rinvgamma}
\title{Stub functions for sampling from distributions}
\usage{
local_rnorm(...)

local_rnorm_vector(n, mus, sigmas)

local_rhorseshoe(n, df, df_global, df_slab, scale_global, scale_slab)

local_rcauchy(...)

local_rgamma(...)

local_rlnorm(...)

local_rbeta(...)

local_runif(...)

local_rt(n, nu, mu, sigma)

local_rlogis(...)

local_rloglogis(n, alpha, beta)

local_rinvgamma(n, alpha, beta)
}
\arguments{
\item{...}{Pass any additional arguments to the underlying distribution.}

\item{nu}{(\code{number}) Parameter for underlying distribution.}

\item{mu}{(\code{number}) Parameter for underlying distribution.}

\item{sigma}{(\code{number}) Parameter for underlying distribution.}

\item{alpha}{(\code{number}) Parameter for underlying distribution.}

\item{beta}{(\code{number}) Parameter for underlying distribution.}
}
\value{
A numeric vector of random draws.
}
\description{
These functions only exist so that they can be mocked during unit
tests in order to provide deterministic values. In most cases
these are just straight forward pass throughs for the underlying
distributions.
}
\details{
\subsection{Log-Logistic}{

There is no log-logistic sampling function within base R so it was implemented
in terms of sampling from the CDF distribution. Using the Stan parameterisation
the CDF is defined as:
\deqn{
u = F(x) = \frac{1}{1 + (x/ \alpha)^{-\beta}}
}
The inverse of this function is:
\deqn{
x = ((u / (1 - u))^(1 / beta)) * alpha
}

Thus we can sample u from a \eqn{Uni(0, 1)} distribution and then derive x from this.
}

\subsection{Inverse-Gamma}{

The inverse Gamma distribution is defined as 1/Gamma thus we calculate this simply
by sampling sampling from the Gamma distribution and then taking the reciprocal.
}

\subsection{Student-t}{

R's sampling functions only produce the standard Student-t distribution so in order
to match Stan's implementation we multiply by the scale parameter and add the location
parameter. See this \href{https://stats.stackexchange.com/a/623611}{Stack Overflow} post
for details
}
}
\keyword{internal}
