% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Local_Sample}
\alias{Local_Sample}
\alias{local_rnorm}
\alias{local_rcauchy}
\alias{local_rgamma}
\alias{local_rlnorm}
\alias{local_rbeta}
\alias{local_runif}
\alias{local_rt}
\alias{local_rlogis}
\alias{local_rloglogis}
\alias{local_rinvgamma}
\title{Stub functions for sampling from distributions}
\usage{
local_rnorm(...)

local_rcauchy(...)

local_rgamma(...)

local_rlnorm(...)

local_rbeta(...)

local_runif(...)

local_rt(n, nu, mu, sigma)

local_rlogis(...)

local_rloglogis(n, alpha, beta)

local_rinvgamma(n, alpha, beta)
}
\arguments{
\item{...}{Pass any additional arguments to the underlying distribution.}

\item{nu}{(\code{number})\cr Parameter for underlying distribution.}

\item{mu}{(\code{number})\cr Parameter for underlying distribution.}

\item{sigma}{(\code{number})\cr Parameter for underlying distribution.}

\item{alpha}{(\code{number})\cr Parameter for underlying distribution.}

\item{beta}{(\code{number})\cr Parameter for underlying distribution.}
}
\description{
These functions only exist so that they can be mocked during unit
tests in order to provide deterministic values. In most cases
these are just straight forward pass throughs for the underlying
distributions.
}
\keyword{internal}
