% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OsModel.R
\name{LogLogisticModule}
\alias{LogLogisticModule}
\title{LogLogisticModule}
\usage{
LogLogisticModule(
  functions = "os_functions.stan",
  data = "os_data.stan",
  parameters = "os_parameters.stan",
  transformed_parameters = "os_transformed_parameters.stan",
  generated_quantities = "os_generated_quantities.stan",
  priors = os_prior(),
  inits = list(),
  templated = TRUE
)
}
\description{
Log logistic module helper function.
}
