% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModelQuantities.R
\name{longitudinal}
\alias{longitudinal}
\alias{longitudinal,JointModelQuantities-method}
\title{\code{longitudinal}}
\usage{
longitudinal(object, ...)

\S4method{longitudinal}{JointModelQuantities}(object, patients = NULL, ...)
}
\arguments{
\item{object}{samples to extract the longitudinal fits from.}

\item{...}{additional options.}

\item{patients}{(\code{character} or \code{NULL})\cr optional subset of patients for
which the longitudinal fit samples should be extracted, the default \code{NULL}
meaning all patients.}
}
\description{
Obtain the longitudinal fit samples from \code{\link{JointModelSamples}}.
}
