% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{longitudinal}
\alias{longitudinal}
\title{\code{longitudinal}}
\usage{
longitudinal(object, ...)
}
\arguments{
\item{object}{samples to extract the longitudinal fits from.}

\item{...}{not used.}
}
\value{
The longitudinal fit samples.
}
\description{
\code{longitudinal}
}
