% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalModel.R,
%   R/LongitudinalRandomSlopeCov.R, R/LongitudinalSteinFojoCov.R
\name{longitudinal_model_stan_data}
\alias{longitudinal_model_stan_data}
\alias{longitudinal_model_stan_data.default}
\alias{longitudinal_model_stan_data.LongitudinalRandomSlopeCov}
\alias{longitudinal_model_stan_data.LongitudinalSteinFojoCov}
\title{Create longitudinal-model-specific Stan data}
\usage{
longitudinal_model_stan_data(model, subject)

\method{longitudinal_model_stan_data}{default}(model, subject)

\method{longitudinal_model_stan_data}{LongitudinalRandomSlopeCov}(model, subject)

\method{longitudinal_model_stan_data}{LongitudinalSteinFojoCov}(model, subject)
}
\arguments{
\item{model}{A \code{\link{LongitudinalModel}} object.}

\item{subject}{A \code{\link{DataSubject}} object.}
}
\value{
A named list of model-specific Stan data components.
}
\description{
Create longitudinal-model-specific Stan data
}
\keyword{internal}
