% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalRandomSlopeCov.R
\name{longitudinal_model_stan_data}
\alias{longitudinal_model_stan_data}
\alias{.random_slope_cov_stan_data}
\alias{longitudinal_model_stan_data.default}
\title{Create longitudinal-model-specific Stan data}
\usage{
longitudinal_model_stan_data(model, subject)

.random_slope_cov_stan_data(model, subject)

longitudinal_model_stan_data.default(model, subject)
}
\arguments{
\item{model}{A \code{\link{LongitudinalModel}} object.}

\item{subject}{A \code{\link{DataSubject}} object.}
}
\value{
A named list of model-specific Stan data components.
}
\description{
Create longitudinal-model-specific Stan data
}
\keyword{internal}
