% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalQuantities.R
\name{longitudinal_plot}
\alias{longitudinal_plot}
\title{Longitudinal Plot}
\usage{
longitudinal_plot(data, data_obs = NULL, add_ci = FALSE)
}
\arguments{
\item{data}{(\code{data.frame})\cr summary statistics for longitudinal
value estimates. See details.}

\item{data_obs}{(\code{data.frame})\cr real observed values to be
overlaid for reference.  See details.}

\item{add_ci}{(\code{logical})\cr Should confidence intervals be added? Default = \code{TRUE}.}
}
\description{
Internal plotting function to create longitudinal plots
This function predominately exists to extract core logic into its own function
to enable easier unit testing.
}
\details{
\subsection{\code{data}}{

Should contain the following columns:
\itemize{
\item \code{time} - Time point
\item \code{group} - The group in which the observation belongs to
\item \code{median} - The median value for the summary statistic
\item \code{upper} - The upper 95\% CI for the summary statistic
\item \code{lower} - The lower 95\% CI for the summary statistic
}
}

\subsection{\code{data_obs}}{

Should contain the following columns:
\itemize{
\item \code{time} - The time at which the observed value occurred
\item \code{Yob} - The real observed value
\item \code{group} - Which group the event belongs to, should correspond to values in \code{data$group}
}
}
}
\keyword{internal}
