% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalClaretBrunoCov.R
\docType{class}
\name{LongitudinalClaretBrunoCov-class}
\alias{LongitudinalClaretBrunoCov-class}
\alias{.LongitudinalClaretBrunoCov}
\title{Claret-Bruno longitudinal model with subject-level covariates}
\description{
Separate covariate predictors model the log-normal distribution parameters
for baseline tumour size, growth, resistance, and growth inhibition.
}
\section{Available Links}{

\itemize{
\item \code{\link[=linkDSLD]{linkDSLD()}}
\item \code{\link[=linkTTG]{linkTTG()}}
\item \code{\link[=linkIdentity]{linkIdentity()}}
\item \code{\link[=linkGrowth]{linkGrowth()}}
}
}

