% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalGSF.R
\docType{class}
\name{LongitudinalGSF-class}
\alias{LongitudinalGSF-class}
\alias{.LongitudinalGSF}
\alias{LongitudinalGSF}
\title{\code{LongitudinalGSF}}
\usage{
LongitudinalGSF(
  mu_bsld = prior_normal(log(60), 1),
  mu_ks = prior_normal(log(0.5), 1),
  mu_kg = prior_normal(log(0.3), 1),
  mu_phi = prior_normal(qlogis(0.5), 1),
  omega_bsld = prior_lognormal(log(0.2), 1),
  omega_ks = prior_lognormal(log(0.2), 1),
  omega_kg = prior_lognormal(log(0.2), 1),
  omega_phi = prior_lognormal(log(0.2), 1),
  sigma = prior_lognormal(log(0.1), 1),
  scaled_variance = FALSE,
  centred = FALSE
)
}
\arguments{
\item{mu_bsld}{(\code{Prior}) for the mean baseline value \code{mu_bsld}.}

\item{mu_ks}{(\code{Prior}) for the mean shrinkage rate \code{mu_ks}.}

\item{mu_kg}{(\code{Prior}) for the mean growth rate \code{mu_kg}.}

\item{mu_phi}{(\code{Prior}) for the mean proportion of cells affected by the treatment \code{mu_phi}.}

\item{omega_bsld}{(\code{Prior}) for the baseline value standard deviation \code{omega_bsld}.}

\item{omega_ks}{(\code{Prior}) for the shrinkage rate standard deviation \code{omega_ks}.}

\item{omega_kg}{(\code{Prior}) for the growth rate standard deviation \code{omega_kg}.}

\item{omega_phi}{(\code{Prior}) for the standard deviation of the proportion of cells
affected by the treatment \code{omega_phi}.}

\item{sigma}{(\code{Prior}) for the variance of the longitudinal values \code{sigma}.}

\item{scaled_variance}{(\code{logical}) whether the variance should be scaled by the expected value, corresponding to a multiplicative model.
As a default, the variance is not scaled by the expected value, corresponding to an additive model.
(See the "Statistical Specifications" vignette for more details.)}

\item{centred}{(\code{logical}) whether to use the centred parameterization.}
}
\value{
A \code{LongitudinalGSF} object.
}
\description{
This class extends the general \code{\link{LongitudinalModel}} class for using the
Generalized Stein-Fojo (GSF) model for the longitudinal outcome.
}
\section{Available Links}{

\itemize{
\item \code{\link[=linkDSLD]{linkDSLD()}}
\item \code{\link[=linkTTG]{linkTTG()}}
\item \code{\link[=linkIdentity]{linkIdentity()}}
\item \code{\link[=linkGrowth]{linkGrowth()}}
}
}

\examples{
LongitudinalGSF()
}
