% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalGSF.R
\docType{class}
\name{LongitudinalGSF-class}
\alias{LongitudinalGSF-class}
\alias{.LongitudinalGSF}
\alias{LongitudinalGSF}
\title{\code{LongitudinalGSF}}
\usage{
LongitudinalGSF(
  mu_bsld = prior_normal(log(60), 1),
  mu_ks = prior_normal(log(0.5), 1),
  mu_kg = prior_normal(log(0.3), 1),
  omega_bsld = prior_lognormal(log(0.2), 1),
  omega_ks = prior_lognormal(log(0.2), 1),
  omega_kg = prior_lognormal(log(0.2), 1),
  a_phi = prior_lognormal(log(5), 1),
  b_phi = prior_lognormal(log(5), 1),
  sigma = prior_lognormal(log(0.1), 1),
  centred = FALSE
)
}
\arguments{
\item{mu_bsld}{(\code{Prior})\cr for the mean baseline value \code{mu_bsld}.}

\item{mu_ks}{(\code{Prior})\cr for the mean shrinkage rate \code{mu_ks}.}

\item{mu_kg}{(\code{Prior})\cr for the mean growth rate \code{mu_kg}.}

\item{omega_bsld}{(\code{Prior})\cr for the baseline value standard deviation \code{omega_bsld}.}

\item{omega_ks}{(\code{Prior})\cr for the shrinkage rate standard deviation \code{omega_ks}.}

\item{omega_kg}{(\code{Prior})\cr for the growth rate standard deviation \code{omega_kg}.}

\item{a_phi}{(\code{Prior})\cr for the alpha parameter for the fraction of cells that respond to treatment.}

\item{b_phi}{(\code{Prior})\cr for the beta parameter for the fraction of cells that respond to treatment.}

\item{sigma}{(\code{Prior})\cr for the variance of the longitudinal values \code{sigma}.}

\item{centred}{(\code{logical})\cr whether to use the centred parameterization.}
}
\description{
This class extends the general \code{\link{LongitudinalModel}} class for using the
Generalized Stein-Fojo (GSF) model for the longitudinal outcome.
}
\section{Available Links}{

\itemize{
\item \code{\link[=linkDSLD]{linkDSLD()}}
\item \code{\link[=linkTTG]{linkTTG()}}
\item \code{\link[=linkIdentity]{linkIdentity()}}
}
}

\concept{LongitudinalGSF_links}
