% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalGSF.R
\docType{class}
\name{LongitudinalGSF-class}
\alias{LongitudinalGSF-class}
\alias{.LongitudinalGSF}
\alias{LongitudinalGSF}
\title{\code{LongitudinalGSF}}
\usage{
LongitudinalGSF(
  mu_bsld = prior_lognormal(log(55), 5, init = 55),
  mu_ks = prior_lognormal(0, 0.5, init = 0),
  mu_kg = prior_lognormal(-0.36, 1, init = -0.36),
  mu_phi = prior_beta(2, 8, init = 0.2),
  omega_bsld = prior_lognormal(0, 1, init = 0),
  omega_ks = prior_lognormal(0, 1, init = 0),
  omega_kg = prior_lognormal(0, 1, init = 0),
  omega_phi = prior_lognormal(0, 1, init = 0),
  sigma = prior_lognormal(-1.6, 0.8, init = -1.6)
)
}
\arguments{
\item{mu_bsld}{(\code{Prior})\cr for the mean baseline value \code{mu_bsld}.}

\item{mu_ks}{(\code{Prior})\cr for the mean shrinkage rate \code{mu_ks}.}

\item{mu_kg}{(\code{Prior})\cr for the mean growth rate \code{mu_kg}.}

\item{mu_phi}{(\code{Prior})\cr for the mean shrinkage proportion \code{mu_phi}.}

\item{omega_bsld}{(\code{Prior})\cr for the baseline value standard deviation \code{omega_bsld}.}

\item{omega_ks}{(\code{Prior})\cr for the shrinkage rate standard deviation \code{omega_ks}.}

\item{omega_kg}{(\code{Prior})\cr for the growth rate standard deviation \code{omega_kg}.}

\item{omega_phi}{(\code{Prior})\cr for the shrinkage proportion standard deviation \code{omega_phi}.}

\item{sigma}{(\code{Prior})\cr for the variance of the longitudinal values \code{sigma}.}
}
\description{
This class extends the general \code{\link{LongitudinalModel}} class for using the
Generalized Stein-Fojo (GSF) model for the longitudinal outcome.
}
