% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalGSFCov.R
\docType{class}
\name{LongitudinalGSFCov-class}
\alias{LongitudinalGSFCov-class}
\alias{.LongitudinalGSFCov}
\title{Generalized Stein-Fojo longitudinal model with subject-level covariates}
\description{
Separate covariate predictors model the log-normal distribution parameters
for baseline, shrinkage, growth, and the logit-normal treatment fraction.
}
\section{Available Links}{

\itemize{
\item \code{\link[=linkDSLD]{linkDSLD()}}
\item \code{\link[=linkTTG]{linkTTG()}}
\item \code{\link[=linkIdentity]{linkIdentity()}}
\item \code{\link[=linkGrowth]{linkGrowth()}}
\item \code{\link[=linkShrinkage]{linkShrinkage()}}
}
}

