% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalModel.R
\docType{class}
\name{LongitudinalModel-class}
\alias{LongitudinalModel-class}
\alias{.LongitudinalModel}
\alias{LongitudinalModel}
\title{\code{LongitudinalModel}}
\usage{
LongitudinalModel(
  stan = StanModule(),
  parameters = ParameterList(),
  name = "<Unnamed>",
  scaled_variance = NA,
  ...
)
}
\arguments{
\item{stan}{(\code{StanModule}) code containing the model-specific Stan code specification.}

\item{parameters}{(\code{ParameterList}) the parameter declaration and prior specification.}

\item{name}{(\code{character}) display name for the model object.}

\item{scaled_variance}{(\code{flag}) whether the variance should be scaled
by the expected value, corresponding to a multiplicative model. If not,
then an additive error model is used.}

\item{...}{additional arguments for \code{\link[=StanModel]{StanModel()}}.}
}
\description{
This class extends the general \code{\link{StanModel}} class to comprise the longitudinal
model specification.
}
\section{Slots}{

\describe{
\item{\code{scaled_variance}}{logical scalar, whether the variance should be scaled
by the expected value, corresponding to a multiplicative model. If not,
then an additive error model is used.}
}}
