% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalModel.R
\docType{class}
\name{LongitudinalModel-class}
\alias{LongitudinalModel-class}
\alias{.LongitudinalModel}
\alias{LongitudinalModel}
\title{\code{LongitudinalModel}}
\usage{
LongitudinalModel(stan = StanModule(), parameters = ParameterList(), ...)
}
\arguments{
\item{stan}{(\code{StanModule})\cr code containing the Stan code specification.}

\item{parameters}{(\code{ParameterList})\cr the parameter specification.}

\item{...}{additional arguments for \code{\link[=StanModule]{StanModule()}}.}
}
\description{
This class extends the general \code{\link{StanModel}} class to comprise the longitudinal
model specification.
}
