% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalModel.R
\docType{class}
\name{LongitudinalModel-class}
\alias{LongitudinalModel-class}
\alias{.LongitudinalModel}
\alias{LongitudinalModel}
\title{\code{LongitudinalModel}}
\usage{
LongitudinalModel(
  stan = StanModule(),
  parameters = ParameterList(),
  name = "<Unnamed>",
  ...
)
}
\arguments{
\item{stan}{(\code{StanModule})\cr code containing the Stan code specification.}

\item{parameters}{(\code{ParameterList})\cr the parameter specification.}

\item{name}{(\code{character})\cr display name for the model object.}

\item{...}{additional arguments for \code{\link[=StanModel]{StanModel()}}.}
}
\description{
This class extends the general \code{\link{StanModel}} class to comprise the longitudinal
model specification.
}
