% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalQuantities.R
\docType{class}
\name{LongitudinalQuantities-class}
\alias{LongitudinalQuantities-class}
\alias{.LongitudinalQuantities}
\alias{LongitudinalQuantities}
\title{\code{LongitudinalQuantities} Object & Constructor Function}
\usage{
LongitudinalQuantities(object, groups = NULL, time_grid = NULL)
}
\arguments{
\item{object}{(\code{\link{JointModelSamples}}) \cr samples as drawn from a Joint Model.}

\item{groups}{(\code{character}, \code{NULL})\cr which patients to calculate the desired
quantities for.}

\item{time_grid}{(\code{numeric} or \code{NULL})\cr a vector of time points to calculate the desired
quantity at. If \code{NULL} will be set to \code{seq(0, max_longitudinal_time, length = 201)}.}
}
\description{
Constructor function to generate a \code{LongitudinalQuantities} object.
}
\details{
Note that unlike \code{\link{SurvivalQuantities}}, \code{\link{LongitudinalQuantities}} does not support
group aggregation.
}
\section{Slots}{

\describe{
\item{\code{quantities}}{(\code{Quantities})\cr The sampled quantities. Should contain 1 element per
element of \code{group}}

\item{\code{groups}}{(\code{list})\cr See argument section for details}

\item{\code{time_grid}}{(\code{numeric})\cr See argument section for details}

\item{\code{data}}{(\code{\link{DataJoint}})\cr The data that the Joint Model was fitted to to produce
the samples/quantities}
}}

\seealso{
Other LongitudinalQuantities: 
\code{\link{as.data.frame.LongitudinalQuantities}()},
\code{\link{as_print_string.LongitudinalQuantities}()},
\code{\link{autoplot.LongitudinalQuantities}()},
\code{\link{summary.LongitudinalQuantities}()}
}
\concept{LongitudinalQuantities}
