% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalQuantities.R
\docType{class}
\name{LongitudinalQuantities-class}
\alias{LongitudinalQuantities-class}
\alias{.LongitudinalQuantities}
\alias{LongitudinalQuantities}
\title{\code{LongitudinalQuantities} Object & Constructor Function}
\usage{
LongitudinalQuantities(object, grid)
}
\arguments{
\item{object}{(\code{\link{JointModelSamples}}) \cr samples as drawn from a Joint Model.}

\item{grid}{(\code{Grid}) \cr object that specifies which subjects and time points to calculate the
quantities for. See \link{Grid-Functions}.}
}
\description{
Constructor function to generate a \code{LongitudinalQuantities} object.
}
\details{
Note that unlike \code{\link{SurvivalQuantities}}, \code{\link{LongitudinalQuantities}} does not support
group aggregation.
}
\section{Slots}{

\describe{
\item{\code{quantities}}{(\code{Quantities})\cr The sampled quantities. Should contain 1 element per
element of \code{group}}

\item{\code{data}}{(\code{DataJoint})\cr Survival and Longitudinal Data.}
}}

\seealso{
Other LongitudinalQuantities: 
\code{\link{as.data.frame.LongitudinalQuantities}()},
\code{\link{autoplot.LongitudinalQuantities}()},
\code{\link{summary.LongitudinalQuantities}()}
}
\concept{LongitudinalQuantities}
