% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalRandomEffects.R
\name{LongitudinalRandomEffects}
\alias{LongitudinalRandomEffects}
\title{Extract Random Effects Samples from a Longitudinal Model}
\usage{
LongitudinalRandomEffects(object)
}
\arguments{
\item{object}{(\code{\link{JointModelSamples}}) \cr samples as drawn from a Joint Model.}
}
\description{
Helper function to extract subject-level random effects samples from the longitudinal
sub-model of a joint model samples object.
}
\seealso{
Other RandomEffectQuantities: 
\code{\link{RandomEffectQuantities-class}},
\code{\link{as.data.frame.RandomEffectQuantities}()},
\code{\link{as_print_string.RandomEffectQuantities}()},
\code{\link{summary.RandomEffectQuantities}()}

Other JointModelSamples: 
\code{\link{as_print_string.JointModelSamples}()}
}
\concept{JointModelSamples}
\concept{RandomEffectQuantities}
