% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalRandomSlope.R
\docType{class}
\name{LongitudinalRandomSlope-class}
\alias{LongitudinalRandomSlope-class}
\alias{.LongitudinalRandomSlope}
\alias{LongitudinalRandomSlope}
\title{\code{LongitudinalRandomSlope}}
\usage{
LongitudinalRandomSlope(
  intercept = prior_normal(30, 10),
  slope_mu = prior_normal(1, 3),
  slope_sigma = prior_lognormal(0, 1.5),
  sigma = prior_lognormal(0, 1.5),
  scaled_variance = FALSE
)
}
\arguments{
\item{intercept}{(\code{Prior}) for the \code{intercept}.}

\item{slope_mu}{(\code{Prior}) for the population slope \code{slope_mu} (one per arm).}

\item{slope_sigma}{(\code{Prior}) for the random slope standard deviation \code{slope_sigma} (one per arm).}

\item{sigma}{(\code{Prior}) for the variance of the longitudinal values \code{sigma}.}

\item{scaled_variance}{(\code{logical}) whether the variance should be scaled by the expected value, corresponding to a multiplicative model.
As a default, the variance is not scaled by the expected value, corresponding to an additive model.
(See the "Statistical Specifications" vignette for more details.)}
}
\description{
This class extends the general \code{\link{LongitudinalModel}} class for using the
random slope linear model for the longitudinal outcome.
}
\section{Available Links}{

\itemize{
\item \code{\link[=linkDSLD]{linkDSLD()}}
\item \code{\link[=linkIdentity]{linkIdentity()}}
}
}

