% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalRandomSlope.R
\docType{class}
\name{LongitudinalRandomSlope-class}
\alias{LongitudinalRandomSlope-class}
\alias{.LongitudinalRandomSlope}
\alias{LongitudinalRandomSlope}
\title{\code{LongitudinalRandomSlope}}
\usage{
LongitudinalRandomSlope(
  intercept = prior_normal(30, 10, init = 30),
  slope_mu = prior_normal(0, 15, init = 0.001),
  slope_sigma = prior_cauchy(0, 4, init = 0.4),
  sigma = prior_cauchy(0, 4, init = 0.4),
  random_slope = prior_none(init = 0)
)
}
\arguments{
\item{intercept}{(\code{Prior})\cr for the \code{intercept}.}

\item{slope_mu}{(\code{Prior})\cr for the population slope \code{slope_mu}.}

\item{slope_sigma}{(\code{Prior})\cr for the random slope standard deviation \code{slope_sigma}.}

\item{sigma}{(\code{Prior})\cr for the variance of the longitudinal values \code{sigma}.}

\item{random_slope}{(\code{Prior})\cr must be \code{prior_none()}, just used to specify initial values.}
}
\description{
This class extends the general \code{\link{LongitudinalModel}} class for using the
random slope linear model for the longitudinal outcome.
}
