% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalRandomSlopeCov.R
\name{LongitudinalRandomSlopeCov}
\alias{LongitudinalRandomSlopeCov}
\title{Construct a random-slope model with subject-level covariates}
\usage{
LongitudinalRandomSlopeCov(
  mu_formula = ~study,
  slope_mu_formula = ~arm,
  slope_sigma_formula = ~arm,
  mu_parametrization = "linear",
  slope_mu_parametrization = "linear",
  slope_sigma_parametrization = "log-linear",
  mu_intercept_prior = prior_normal(30, 10),
  mu_coefficients_prior = prior_normal(0, 3),
  slope_mu_intercept_prior = prior_normal(1, 3),
  slope_mu_coefficients_prior = prior_normal(0, 3),
  slope_sigma_intercept_prior = prior_normal(0, 1.5),
  slope_sigma_coefficients_prior = prior_normal(0, 1.5),
  sigma = prior_lognormal(0, 1.5),
  scaled_variance = FALSE
)
}
\arguments{
\item{mu_formula, slope_mu_formula, slope_sigma_formula}{One-sided formulas
evaluated in the subject-level data.}

\item{mu_parametrization, slope_mu_parametrization, slope_sigma_parametrization}{Predictor parametrizations.}

\item{mu_intercept_prior, mu_coefficients_prior}{Priors for the subject-intercept predictor.}

\item{slope_mu_intercept_prior, slope_mu_coefficients_prior}{Priors for the mean-slope predictor.}

\item{slope_sigma_intercept_prior, slope_sigma_coefficients_prior}{Priors for the slope-SD predictor.}

\item{sigma}{Prior for the observation-error standard deviation.}

\item{scaled_variance}{Whether to use multiplicative observation error.}
}
\value{
A \code{LongitudinalRandomSlopeCov} object.
}
\description{
The intercept column generated by each formula is removed. The separately
specified intercept parameter is the reference-level value except for the
code{exponential} parametrization, whose reference-level value is one.
\code{linear} uses \verb{intercept + X beta}, \code{proportional} uses
\verb{intercept * (1 + X beta)}, \code{exponential} uses
\verb{intercept ^ (X beta)}, and \code{log-linear} uses
\verb{exp(intercept + X beta)}.
}
\examples{
LongitudinalRandomSlopeCov()
}
