% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalSteinFojoCov.R
\docType{class}
\name{LongitudinalSteinFojoCov-class}
\alias{LongitudinalSteinFojoCov-class}
\alias{.LongitudinalSteinFojoCov}
\title{Stein-Fojo longitudinal model with subject-level covariates}
\description{
Models the log-normal distribution parameters for baseline, shrinkage, and
growth using separate subject-level covariate predictors.
}
\section{Available Links}{

\itemize{
\item \code{\link[=linkDSLD]{linkDSLD()}}
\item \code{\link[=linkTTG]{linkTTG()}}
\item \code{\link[=linkIdentity]{linkIdentity()}}
\item \code{\link[=linkGrowth]{linkGrowth()}}
\item \code{\link[=linkShrinkage]{linkShrinkage()}}
}
}

