% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longmodel.R
\docType{class}
\name{LongModel-class}
\alias{LongModel-class}
\alias{LongModel}
\alias{getLinkTTG,LongModel-method}
\alias{getLinkDSLD,LongModel-method}
\alias{getLink,LongModel-method}
\title{Title - TODO}
\usage{
\S4method{getLinkTTG}{LongModel}(object)

\S4method{getLinkDSLD}{LongModel}(object)

\S4method{getLink}{LongModel}(object, selection = c("ttg", "dsld"))
}
\description{
Description - TODO
}
\section{Slots}{

\describe{
\item{\code{stan}}{TODO}
}}

