% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_score.R
\name{match_order}
\alias{match_order}
\title{Match Order}
\usage{
match_order(x)
}
\arguments{
\item{x}{(\code{numeric})\cr a vector for which we want to generate an index so other
vectors can be put into the same sort order

Assuming we have a vector that is sorted then this function
will return the index vector to convert that sorted vector
into the same sort order as the input vector \code{x}.
For example let \verb{x = 8, 7, 9 , 7}. If sorted we would get
\verb{x_sort = 7, 7, 8, 9}. So in order to convert \code{x_sort}
back into \code{x} we'd need an index vector of \verb{3, 1, 4, 2}.
This function is used to determine that corresponding
index vector for an arbitrarily sorted vector \code{x}.

There is no specific handling of ties. It is assuming that in the case
of ties for \code{x} that the vector you are re-indexing also has tied values
thus the specific tied element selection does not matter.}
}
\description{
Match Order
}
\keyword{internal}
