% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/stan_module.R, R/hazard_link.R,
%   R/joint.R
\name{merge}
\alias{merge}
\alias{merge,StanModule,StanModule-method}
\alias{merge,HazardLink,HazardLink-method}
\alias{merge,LongModel,OsModel-method}
\title{Merge}
\usage{
merge(x, y, ...)

\S4method{merge}{StanModule,StanModule}(x, y)

\S4method{merge}{HazardLink,HazardLink}(x, y)

\S4method{merge}{LongModel,OsModel}(x, y)
}
\arguments{
\item{x}{An Object}

\item{y}{An Object to be merged with \code{x}}
}
\description{
Generic function to collapse two similar objects into a single combined object
}
