% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{merge}
\alias{merge}
\title{\code{merge}}
\usage{
merge(x, y, ...)
}
\arguments{
\item{x}{first module.}

\item{y}{second module.}

\item{...}{additional arguments.}
}
\description{
Merge two Stan modules.
}
