% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSurvival.R
\name{mirror_design_matrix}
\alias{mirror_design_matrix}
\title{Build design matrix for prediction data}
\usage{
mirror_design_matrix(olddata, newdata)
}
\arguments{
\item{olddata}{(\code{\link{DataSurvival}}) \cr The original data to be used as a template for the new data}

\item{newdata}{(\code{\link{data.frame}}) \cr The new data to be used to generate the design matrix}
}
\description{
This function takes a \code{DataSurvival} object and a \code{data.frame} object and generates
a design matrix for the \code{data.frame} that has the identical structure to the
design matrix of the \code{DataSurvival} object.

This is used for predicting new data using a model that was trained on a different
original data source
}
\keyword{internal}
