% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OsModel.R
\docType{class}
\name{OsModel-class}
\alias{OsModel-class}
\alias{OsModel}
\title{Title - TODO}
\description{
Description - TODO
}
\section{Slots}{

\describe{
\item{\code{stan}}{TODO}

\item{\code{templated}}{Logical. Defines whether the OsModel object needs to be completed}
}}

