% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\docType{class}
\name{Parameter-class}
\alias{Parameter-class}
\alias{.Parameter}
\alias{Parameter}
\title{\code{Parameter}}
\usage{
Parameter(prior, name, size = 1)
}
\arguments{
\item{prior}{(\code{Prior})\cr for the parameter.}

\item{name}{(\code{string})\cr of the parameter.}

\item{size}{(\verb{"}numeric\code{or}string\verb{"})\cr dimension of the parameter.}
}
\description{
Stores the name, the prior distribution and the size of a parameter.
}
\section{Slots}{

\describe{
\item{\code{name}}{(\code{string})\cr of the parameter.}

\item{\code{prior}}{(\code{Prior})\cr for the parameter.}

\item{\code{size}}{(\code{numeric} or \code{string})\cr dimension of the parameter.}
}}

