% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{Parameter-Getter-Methods}
\alias{Parameter-Getter-Methods}
\alias{names.Parameter}
\alias{initialValues.Parameter}
\alias{size.Parameter}
\title{Parameter Getter Functions}
\usage{
\method{names}{Parameter}(x)

\method{initialValues}{Parameter}(object, ...)

\method{size}{Parameter}(object)
}
\arguments{
\item{x}{(\code{Paramater}) \cr A model parameter}

\item{object}{(\code{Paramater}) \cr A model parameter}

\item{...}{Not used.}
}
\description{
Getter functions for the slots of a \code{\link{Parameter}} object
}
\section{Functions}{
\itemize{
\item \code{names(Parameter)}: The parameter's name

\item \code{initialValues(Parameter)}: The parameter's initial values

\item \code{size(Parameter)}: The parameter's dimensionality

}}
\seealso{
Other Parameter: 
\code{\link{Parameter-class}},
\code{\link{as.StanModule.Parameter}()},
\code{\link{as.character.Parameter}()},
\code{\link{as_stan_list.Parameter}()}
}
\concept{Parameter}
