% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterList.R
\docType{class}
\name{ParameterList-class}
\alias{ParameterList-class}
\alias{.ParameterList}
\alias{ParameterList}
\title{\code{ParameterList}}
\usage{
ParameterList(...)
}
\arguments{
\item{...}{(\code{Parameter})\cr which parameter specifications to include.}
}
\description{
This class extends the general \code{\link{list}} type for containing \code{\link{Parameter}}
specifications.
}
