% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterList.R
\name{ParameterList-Getter-Methods}
\alias{ParameterList-Getter-Methods}
\alias{names.ParameterList}
\alias{initialValues.ParameterList}
\alias{size.ParameterList}
\title{Parameter-List Getter Functions}
\usage{
\method{names}{ParameterList}(x)

\method{initialValues}{ParameterList}(object)

\method{size}{ParameterList}(object)
}
\arguments{
\item{x}{(\code{ParameterList}) \cr A List of \code{\link{Parameter}} Objects.}

\item{object}{(\code{ParameterList}) \cr A List of \code{\link{Parameter}} Objects.}
}
\description{
Getter functions for the slots of a \code{\link{ParameterList}} object
}
\section{Functions}{
\itemize{
\item \code{names(ParameterList)}: The parameter-list's parameter names

\item \code{initialValues(ParameterList)}: The parameter-list's parameter initial values

\item \code{size(ParameterList)}: The parameter-list's parameter dimensionality

}}
\seealso{
Other ParameterList: 
\code{\link{ParameterList-class}},
\code{\link{as.StanModule.ParameterList}()},
\code{\link{as.character.ParameterList}()},
\code{\link{as.list.ParameterList}()}
}
\concept{ParameterList}
