% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/OsModel.R
\name{parametrize}
\alias{parametrize}
\alias{parametrize,OsModel,HazardLink-method}
\title{Parametrize TemplatedStanOs object with the selected Hazardlink}
\usage{
parametrize(osmod, link)

\S4method{parametrize}{OsModel,HazardLink}(osmod, link)
}
\arguments{
\item{osmod}{TemplatedStanOs object}

\item{link}{HazardLink object}
}
\description{
Parametrize TemplatedStanOs object with the selected Hazardlink
}
