% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationHR.R
\name{PopulationHR}
\alias{PopulationHR}
\title{Calculate Population Hazard Ratios}
\usage{
PopulationHR(
  object,
  HR_terms = attr(terms(object@data@survival@formula), "term.labels"),
  arm = TRUE,
  baseline = "bs(time, df=10)"
)
}
\arguments{
\item{object}{(\code{\link{JointModelSamples}}) \cr samples as drawn from a Joint Model.}

\item{HR_terms}{(\code{character}) \cr to include in the hazard ratio calculation.
By default this include the terms used in the survival model.
Set to \code{NULL} not include any terms}

\item{arm}{(\code{logical}) \cr calculate hazard ratio for \code{arm}.}

\item{baseline}{(\code{string}) \cr terms to model baseline hazard using variable \code{"time"}.
Default is a B-spline from \link{splines}: \code{"bs(time, df=10)"}}
}
\value{
A matrix
}
\description{
Calculates hazard ratios marginalised over subject specific random effects
}
