% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\docType{class}
\name{Prior-class}
\alias{Prior-class}
\alias{.Prior}
\alias{Prior}
\title{Prior Object and Constructor Function}
\usage{
Prior(parameters, repr_model, repr_data, init, validation)
}
\arguments{
\item{parameters}{(\code{list})\cr the prior distribution parameters.}

\item{repr_model}{(\code{string})\cr the Stan code representation for the model block.}

\item{repr_data}{(\code{string})\cr the Stan code representation for the data block.}

\item{init}{(\code{numeric})\cr the initial value.}

\item{validation}{(\code{list})\cr the prior distribution parameter validation functions. Must have
the same names as the \code{paramaters} slot.}
}
\description{
Specifies the prior distribution in a Stan Model
}
\section{Slots}{

\describe{
\item{\code{parameters}}{(\code{list})\cr See arguments.}

\item{\code{repr_model}}{(\code{string})\cr See arguments.}

\item{\code{repr_data}}{(\code{string})\cr See arguments.}

\item{\code{init}}{(\code{numeric})\cr See arguments.}

\item{\code{validation}}{(\code{list})\cr See arguments.}
}}

\seealso{
Other Prior-internal: 
\code{\link{Prior-Getter-Methods}},
\code{\link{as.StanModule.Prior}()},
\code{\link{as_stan_list.Prior}()}
}
\concept{Prior-internal}
