% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\docType{class}
\name{Prior-class}
\alias{Prior-class}
\alias{.Prior}
\alias{Prior}
\title{\code{Prior}}
\description{
\code{Prior}
}
\section{Slots}{

\describe{
\item{\code{parameters}}{(\code{list})\cr the prior distribution parameters.}

\item{\code{repr}}{(\code{string})\cr the Stan code regular expression encoding the distribution.}

\item{\code{init}}{(\code{numeric})\cr the initial value.}

\item{\code{validation}}{(\code{list})\cr the prior distribution parameter validation functions. Must have
the same names as the \code{paramaters} slot.}
}}

