% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Prior-Getter-Methods}
\alias{Prior-Getter-Methods}
\alias{initialValues.Prior}
\title{Prior Getter Functions}
\usage{
\method{initialValues}{Prior}(object)
}
\arguments{
\item{object}{(\code{\link{Prior}})\cr A Prior Distribution}
}
\description{
Getter functions for the slots of a \code{\link{Prior}} object
}
\section{Functions}{
\itemize{
\item \code{initialValues(Prior)}: The prior's initial value

}}
\seealso{
Other Prior: 
\code{\link{Prior-class}},
\code{\link{as.StanModule.Prior}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_lognormal}()},
\code{\link{prior_none}()},
\code{\link{prior_normal}()},
\code{\link{prior_std_normal}()}
}
\concept{Prior}
