% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Prior-Getter-Methods}
\alias{Prior-Getter-Methods}
\alias{initialValues.Prior}
\title{Prior Getter Functions}
\usage{
\method{initialValues}{Prior}(object)
}
\arguments{
\item{object}{(\code{\link{Prior}})\cr a prior Distribution}
}
\description{
Getter functions for the slots of a \code{\link{Prior}} object
}
\section{Functions}{
\itemize{
\item \code{initialValues(Prior)}: The prior's initial value

}}
\seealso{
Other Prior-internal: 
\code{\link{Prior-class}},
\code{\link{as.StanModule.Prior}()},
\code{\link{as.character.Prior}()},
\code{\link{as_stan_list.Prior}()}
}
\concept{Prior-internal}
