% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_beta}
\alias{prior_beta}
\title{Beta Prior Distribution}
\usage{
prior_beta(a, b, init = a/(a + b))
}
\arguments{
\item{a}{(\code{number})\cr first parameter.}

\item{b}{(\code{number})\cr second parameter}

\item{init}{(\code{number})\cr initial value.}
}
\description{
Beta Prior Distribution
}
