% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_beta}
\alias{prior_beta}
\title{Beta Prior Distribution}
\usage{
prior_beta(a, b, init = a/(a + b))
}
\arguments{
\item{a}{(\code{number})\cr first parameter.}

\item{b}{(\code{number})\cr second parameter}

\item{init}{(\code{number})\cr initial value.}
}
\description{
Beta Prior Distribution
}
\seealso{
Other Prior: 
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_invgamma}()},
\code{\link{prior_logistic}()},
\code{\link{prior_loglogistic}()},
\code{\link{prior_lognormal}()},
\code{\link{prior_none}()},
\code{\link{prior_normal}()},
\code{\link{prior_std_normal}()},
\code{\link{prior_student_t}()},
\code{\link{prior_uniform}()}
}
\concept{Prior}
