% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_cauchy}
\alias{prior_cauchy}
\title{Cauchy Prior Distribution}
\usage{
prior_cauchy(mu, sigma, init = mu)
}
\arguments{
\item{mu}{(\code{number})\cr mean.}

\item{sigma}{(\code{number})\cr scale.}

\item{init}{(\code{number})\cr initial value.}
}
\description{
Cauchy Prior Distribution
}
