% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_cauchy}
\alias{prior_cauchy}
\title{Cauchy Prior Distribution}
\usage{
prior_cauchy(mu, sigma, init = mu)
}
\arguments{
\item{mu}{(\code{number})\cr mean.}

\item{sigma}{(\code{number})\cr scale.}

\item{init}{(\code{number})\cr initial value.}
}
\description{
Cauchy Prior Distribution
}
\seealso{
Other Prior: 
\code{\link{prior_beta}()},
\code{\link{prior_gamma}()},
\code{\link{prior_lognormal}()},
\code{\link{prior_none}()},
\code{\link{prior_normal}()},
\code{\link{prior_std_normal}()}
}
\concept{Prior}
