% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_gamma}
\alias{prior_gamma}
\title{Gamma Prior Distribution}
\usage{
prior_gamma(alpha, beta, init = alpha/beta)
}
\arguments{
\item{alpha}{(\code{number})\cr shape.}

\item{beta}{(\code{number})\cr inverse scale.}

\item{init}{(\code{number})\cr initial value.}
}
\description{
Gamma Prior Distribution
}
