% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_horseshoe}
\alias{prior_horseshoe}
\title{Regularized Horseshoe Prior for a Vector Distribution}
\usage{
prior_horseshoe(
  df = 1,
  df_global = 1,
  df_slab = 4,
  scale_global = 1,
  scale_slab = 2
)
}
\arguments{
\item{df}{(\code{number}) degrees of freedom of the half-Student-t prior for local shrinkage parameters.}

\item{df_global}{(\code{number}) degrees of freedom of the half-Student-t prior for the global shrinkage parameter.}

\item{df_slab}{(\code{number}) degrees of freedom of the Student-t slab.}

\item{scale_global}{(\code{number}) scale of the half-Student-t prior for the global shrinkage parameter.}

\item{scale_slab}{(\code{number}) scale of the Student-t slab.}
}
\description{
Regularized Horseshoe Prior for a Vector Distribution
}
\seealso{
Other Prior:
\code{\link[=prior_beta]{prior_beta()}},
\code{\link[=prior_cauchy]{prior_cauchy()}},
\code{\link[=prior_gamma]{prior_gamma()}},
\code{\link[=prior_init_only]{prior_init_only()}},
\code{\link[=prior_invgamma]{prior_invgamma()}},
\code{\link[=prior_logistic]{prior_logistic()}},
\code{\link[=prior_loglogistic]{prior_loglogistic()}},
\code{\link[=prior_lognormal]{prior_lognormal()}},
\code{\link[=prior_normal]{prior_normal()}},
\code{\link[=prior_normal_vector]{prior_normal_vector()}},
\code{\link[=prior_std_normal]{prior_std_normal()}},
\code{\link[=prior_student_t]{prior_student_t()}},
\code{\link[=prior_uniform]{prior_uniform()}}
}
\concept{Prior}
