% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_init_only}
\alias{prior_init_only}
\title{Initial Values Specification}
\usage{
prior_init_only(dist)
}
\arguments{
\item{dist}{(\code{Prior})\cr a prior Distribution}
}
\description{
This function is used to specify only the initial values for a parameter.
This is primarily used for hierarchical parameters whose distributions
are fixed within the model and cannot be altered by the user.
}
\seealso{
Other Prior: 
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_invgamma}()},
\code{\link{prior_logistic}()},
\code{\link{prior_loglogistic}()},
\code{\link{prior_lognormal}()},
\code{\link{prior_normal}()},
\code{\link{prior_std_normal}()},
\code{\link{prior_student_t}()},
\code{\link{prior_uniform}()}
}
\concept{Prior}
