% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_lognormal}
\alias{prior_lognormal}
\title{Log-Normal Prior Distribution}
\usage{
prior_lognormal(mu, sigma, init = exp(mu + (sigma^2)/2))
}
\arguments{
\item{mu}{(\code{number})\cr mean of the logarithm.}

\item{sigma}{(\code{number})\cr standard deviation of the logarithm.}

\item{init}{(\code{number})\cr initial value.}
}
\description{
Log-Normal Prior Distribution
}
