% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_lognormal}
\alias{prior_lognormal}
\title{Log-Normal Prior Distribution}
\usage{
prior_lognormal(mu, sigma)
}
\arguments{
\item{mu}{(\code{number})\cr mean of the logarithm.}

\item{sigma}{(\code{number})\cr standard deviation of the logarithm.}
}
\description{
Log-Normal Prior Distribution
}
\seealso{
Other Prior: 
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_init_only}()},
\code{\link{prior_invgamma}()},
\code{\link{prior_logistic}()},
\code{\link{prior_loglogistic}()},
\code{\link{prior_normal}()},
\code{\link{prior_std_normal}()},
\code{\link{prior_student_t}()},
\code{\link{prior_uniform}()}
}
\concept{Prior}
