% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_none}
\alias{prior_none}
\title{Only Initial Values Specification}
\usage{
prior_none(init = 1e-05)
}
\arguments{
\item{init}{(\code{number})\cr initial value.}
}
\description{
Only Initial Values Specification
}
\seealso{
Other Prior: 
\code{\link{Prior-Getter-Methods}},
\code{\link{Prior-class}},
\code{\link{as.character.Prior}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_lognormal}()},
\code{\link{prior_normal}()},
\code{\link{prior_std_normal}()}
}
\concept{Prior}
