% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_none}
\alias{prior_none}
\title{Only Initial Values Specification}
\usage{
prior_none()
}
\description{
Only Initial Values Specification
}
\seealso{
Other Prior: 
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_invgamma}()},
\code{\link{prior_logistic}()},
\code{\link{prior_loglogistic}()},
\code{\link{prior_lognormal}()},
\code{\link{prior_normal}()},
\code{\link{prior_std_normal}()},
\code{\link{prior_student_t}()},
\code{\link{prior_uniform}()}
}
\concept{Prior}
