% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_normal}
\alias{prior_normal}
\title{Normal Prior Distribution}
\usage{
prior_normal(mu, sigma, init = mu)
}
\arguments{
\item{mu}{(\code{number})\cr mean.}

\item{sigma}{(\code{number})\cr standard deviation.}

\item{init}{(\code{number})\cr initial value.}
}
\description{
Normal Prior Distribution
}
