% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_std_normal}
\alias{prior_std_normal}
\title{Standard Normal Prior Distribution}
\usage{
prior_std_normal()
}
\description{
Standard Normal Prior Distribution
}
\seealso{
Other Prior: 
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_invgamma}()},
\code{\link{prior_logistic}()},
\code{\link{prior_loglogistic}()},
\code{\link{prior_lognormal}()},
\code{\link{prior_none}()},
\code{\link{prior_normal}()},
\code{\link{prior_student_t}()},
\code{\link{prior_uniform}()}
}
\concept{Prior}
