% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_std_normal}
\alias{prior_std_normal}
\title{Standard Normal Prior Distribution}
\usage{
prior_std_normal(init = 0)
}
\arguments{
\item{init}{(\code{number})\cr initial value.}
}
\description{
Standard Normal Prior Distribution
}
