% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_student_t}
\alias{prior_student_t}
\title{Student-t Prior Distribution}
\usage{
prior_student_t(nu, mu, sigma)
}
\arguments{
\item{nu}{(\code{number})\cr Degrees of freedom parameter.}

\item{mu}{(\code{number})\cr Location parameter.}

\item{sigma}{(\code{number})\cr Scale parameter.}
}
\description{
Student-t Prior Distribution
}
\seealso{
Other Prior: 
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_invgamma}()},
\code{\link{prior_logistic}()},
\code{\link{prior_loglogistic}()},
\code{\link{prior_lognormal}()},
\code{\link{prior_none}()},
\code{\link{prior_normal}()},
\code{\link{prior_std_normal}()},
\code{\link{prior_uniform}()}
}
\concept{Prior}
