% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior_uniform}
\alias{prior_uniform}
\title{Uniform Prior Distribution}
\usage{
prior_uniform(alpha, beta, init = 0.5 * (alpha + beta))
}
\arguments{
\item{alpha}{(\code{number})\cr minimum value parameter.}

\item{beta}{(\code{number})\cr maximum value parameter.}

\item{init}{(\code{number})\cr initial value.}
}
\description{
Uniform Prior Distribution
}
\seealso{
Other Prior: 
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_invgamma}()},
\code{\link{prior_logistic}()},
\code{\link{prior_loglogistic}()},
\code{\link{prior_lognormal}()},
\code{\link{prior_none}()},
\code{\link{prior_normal}()},
\code{\link{prior_std_normal}()},
\code{\link{prior_student_t}()}
}
\concept{Prior}
