% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/stan_module.R, R/hazard_link.R,
%   R/longmodel.R, R/OsModel.R
\name{priors}
\alias{priors}
\alias{priors,StanModule-method}
\alias{priors,HazardLink-method}
\alias{priors,LongModel-method}
\alias{priors,OsModel-method}
\title{priors() get method for \code{LongModel}, \code{HazardLink} , \code{StanModule} or \code{OsModel}}
\usage{
priors(object)

\S4method{priors}{StanModule}(object)

\S4method{priors}{HazardLink}(object)

\S4method{priors}{LongModel}(object)

\S4method{priors}{OsModel}(object)
}
\arguments{
\item{object}{A \code{OsModel} object generated from function \code{\link[=OsModel]{OsModel()}}}

\item{value}{the character strings of the prior information for replacement}
}
\description{
priors() get method for \code{LongModel}, \code{HazardLink} , \code{StanModule} or \code{OsModel}

getter method for StanModule:

get method for Hazard link

get method for LongModel:

getter method for OsModel
getter method for OsModel:
}
\examples{
priors(object)
}
