% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Promise.R
\docType{class}
\name{PromiseLinkComponent-class}
\alias{PromiseLinkComponent-class}
\alias{.PromiseLinkComponent}
\alias{PromiseLinkComponent}
\title{Promise of a \code{LinkComponent}}
\usage{
PromiseLinkComponent(fun, prior, key)
}
\description{
An object that promises to resolve to a \code{\link{LinkComponent}} object.
Inheriting from \code{\link{Promise}} and \code{\link{LinkComponent}}.
}
\details{
The \code{fun} slot should be a function of signature \verb{function(prior, model)} and should return
a \code{\link{LinkComponent}} object. An error will be thrown if the returned \code{\link{LinkComponent}} object
does not have the same \code{key} slot value as the original \code{PromiseLinkComponent}.
}
\section{Slots}{

\describe{
\item{\code{fun}}{(\code{function}) \cr a function that returns a \code{LinkComponent}. See details.}
}}

