% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Promise.R
\docType{class}
\name{PromiseLongitudinalModel-class}
\alias{PromiseLongitudinalModel-class}
\alias{.PromiseLongitudinalModel}
\alias{PromiseLongitudinalModel}
\title{Promise of a \code{LongitudinalModel}}
\usage{
PromiseLongitudinalModel()
}
\description{
An object that promises to resolve to a \code{\link{LongutidunalModel}} object.
}
