% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pt_2_factor}
\alias{pt_2_factor}
\title{\code{pt_2_factor}}
\usage{
pt_2_factor(pt)
}
\arguments{
\item{pt}{(\code{character} or \code{factor})\cr subject identifiers.}
}
\value{
A \code{factor} with the levels in alphabetical order.
}
\description{
Converts subject identifiers to factors.
If \code{pt} is already a factor it will re-level it to ensure the levels are in alphabetical order.
This is to ensure that \code{\link{DataLongitudinal}} and \code{\link{DataSurvival}} use identical index numbers
for the subjects to ensure data alignment in the joint model.
}
\keyword{internal}
