% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\docType{class}
\name{Quantities-class}
\alias{Quantities-class}
\alias{.Quantities}
\alias{Quantities}
\title{Generated Quantities Container}
\usage{
Quantities(x)
}
\arguments{
\item{x}{(\code{list})\cr a \code{list} of \code{matrix} objects. Each matrix must have the exact
same dimensions.}
}
\description{
A simple wrapper around a \code{list} to ensure each element complies
to specific formatting rules
}
\seealso{
Other Quantities: 
\code{\link{as.data.frame.Quantities}()},
\code{\link{summary.Quantities}()}
}
\concept{Quantities}
\keyword{internal}
