% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\docType{class}
\name{Quantities-class}
\alias{Quantities-class}
\alias{.Quantities}
\alias{Quantities}
\title{Generated Quantities Container}
\usage{
Quantities(quantities, times, groups)
}
\arguments{
\item{quantities}{(\code{matrix})\cr of generated quantities.}

\item{times}{(\code{numeric})\cr labels specifying which time point the quantity was generated at.}

\item{groups}{(\code{character})\cr labels for which group the quantity belongs to.}
}
\description{
A simple wrapper around a \code{matrix} to store required metadata
}
\details{
Each row of the matrix represents a sample and each column represents a distinct quantity.
As such the number of columns in the matrix should equal the length of \code{times} and \code{groups}
which provide metadata for who the quantity belongs to and at what time point it was generated at.
}
\seealso{
Other Quantities: 
\code{\link{as.data.frame.Quantities}()},
\code{\link{as_print_string.Quantities}()},
\code{\link{summary.Quantities}()}
}
\concept{Quantities}
\keyword{internal}
