% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\name{Quantities-Shared}
\alias{Quantities-Shared}
\title{Re-used documentation for \code{Quantities}}
\arguments{
\item{x}{(\code{\link{Quantities}}) \cr generated quantities.}

\item{object}{(\code{\link{Quantities}}) \cr generated quantities.}

\item{time_grid}{(\code{numeric})\cr sets the \code{time} variable.
Must be equal in length to \code{ncol(x)}.}

\item{type}{(\code{character})\cr sets the \code{type} variable.}

\item{groups}{(\code{list})\cr named \code{list}.
The element names are used to set the \code{group} variable.}

\item{conf.level}{(\code{numeric}) \cr confidence level of the interval.}

\item{...}{not used.}
}
\description{
Re-used documentation for \code{Quantities}
}
\keyword{internal}
