% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_module.R
\name{read_stan}
\alias{read_stan}
\title{read_stan_part returns stan code as a character.}
\usage{
read_stan(string)
}
\arguments{
\item{file}{Character, either the absolute path of a stan file, or the name of the stan
file in the package directory or the stan code as a string.}
}
\description{
read_stan_part returns stan code as a character.
}
