% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\name{read_stan}
\alias{read_stan}
\title{Stan Code as Character}
\usage{
read_stan(string)
}
\arguments{
\item{string}{(\code{string})\cr either the absolute path of a stan file, or the name of the stan
file in the package directory or the stan code as a string.}
}
\description{
Stan Code as Character
}
