% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{render_stan_const_declaration}
\alias{render_stan_const_declaration}
\title{Render Stan Constant Parameter Declaration}
\usage{
<<<<<<< HEAD
render_stan_const_declaration(name, size, limits, is_vector = FALSE)
=======
render_stan_const_declaration(name, size, limits)
>>>>>>> main
}
\arguments{
\item{name}{(\code{string}) parameter name.}

\item{size}{(\code{numeric_OR_character}) parameter size.}

\item{limits}{(\code{numeric}) lower and upper parameter limits.}
<<<<<<< HEAD

\item{is_vector}{(\code{flag}) whether the supplied constant is a vector.}
=======
>>>>>>> main
}
\value{
A length-one character vector.
}
\description{
Creates a Stan transformed-parameter declaration that fixes a scalar or vector
parameter at a data-supplied constant value.
}
\keyword{internal}
