% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{render_stan_limits}
\alias{render_stan_limits}
\title{Creates Stan Syntax for Truncated distributions}
\usage{
render_stan_limits(object)
}
\arguments{
\item{object}{a \code{\link{Prior}} or numeric lower/upper limits for a
truncated distribution}
}
\value{
(\code{character}) the Stan syntax for truncated distributions
}
\description{
This function creates the Stan syntax for truncated distributions
}
\keyword{internal}
