% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{render_stan_limits}
\alias{render_stan_limits}
\alias{render_stan_limits.Prior}
\alias{render_stan_limits.numeric}
\title{Creates Stan Syntax for Truncated distributions}
\usage{
render_stan_limits(object)

\method{render_stan_limits}{Prior}(object)

\method{render_stan_limits}{numeric}(object)
}
\arguments{
\item{object}{(\code{Prior | numeric}) prior or lower and upper limits for a truncated distribution.}
}
\value{
(\code{character}) the Stan syntax for truncated distributions

A character vector containing the Stan truncation syntax.
}
\description{
This function creates the Stan syntax for truncated distributions
}
\section{Methods (by class)}{
\itemize{
\item \code{render_stan_limits(Prior)}: method for \link{Prior} objects.

\item \code{render_stan_limits(numeric)}: method for numeric vectors.

}}
\keyword{internal}
