% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{replace_with_lookup}
\alias{replace_with_lookup}
\title{replace_with_lookup}
\usage{
replace_with_lookup(sizes, data)
}
\arguments{
\item{sizes}{A list of sizes, which may include character elements that correspond to
names in the data list.}

\item{data}{A list of data containing numeric values.}
}
\value{
A list of sizes with character elements replaced by their corresponding
numeric values.
}
\description{
This function takes a list of sizes and a list of data and returns a modified list
of sizes where each character element is replaced by the corresponding numeric value
from the data list. Each element of size must be a length one numeric after the lookup
}
\keyword{internal}
