% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/LongitudinalRandomSlopeCov.R,
%   R/LongitudinalSteinFojoCov.R
\name{required_simulation_covariates}
\alias{required_simulation_covariates}
\alias{required_simulation_covariates.default}
\alias{required_simulation_covariates.LongitudinalRandomSlopeCov}
\alias{required_simulation_covariates.LongitudinalSteinFojoCov}
\title{Required Longitudinal Simulation Covariates}
\usage{
required_simulation_covariates(object, ...)

\method{required_simulation_covariates}{default}(object, ...)

\method{required_simulation_covariates}{LongitudinalRandomSlopeCov}(object, ...)

\method{required_simulation_covariates}{LongitudinalSteinFojoCov}(object, ...)
}
\arguments{
\item{object}{(\code{LongitudinalModel}) A longitudinal model object.}

\item{...}{Not used.}
}
\value{
A character vector of covariate names.
}
\description{
Return the subject-level covariates required to simulate a longitudinal
model from posterior draws. Unlike \code{\link[=required_longitudinal_covs]{required_longitudinal_covs()}}, this
includes covariates used only by variability predictors.
}
