% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link.R
\name{resolvePromise.Link}
\alias{resolvePromise.Link}
\title{Resolve any promises}
\usage{
\method{resolvePromise}{Link}(object, model, ...)
}
\arguments{
\item{object}{(\code{Link}) a link object.}

\item{model}{(\code{LongitudinalModel}) the model object.}

\item{...}{Not Used.}
}
\description{
Loops over all components and ensures that any \code{\link{PromiseLinkComponent}} objects
are resolved to \code{\link{LinkComponent}} objects.
}
