% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{resolvePromise}
\alias{resolvePromise}
\alias{resolvePromise.default}
\title{Resolve a Promise}
\usage{
resolvePromise(object, ...)

\method{resolvePromise}{default}(object, ...)
}
\arguments{
\item{object}{(\code{ANY})\cr an object to resolve.}

\item{...}{(\code{ANY})\cr additional arguments.

If \code{object} is not a promise will just return itself else will resolve the promise
and return the promised object.}
}
\description{
Resolve a Promise
}
