% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{samples_median_ci}
\alias{samples_median_ci}
\title{Obtain Median and Credible Intervals from MCMC samples}
\usage{
samples_median_ci(samples, level = 0.95)
}
\arguments{
\item{samples}{(\code{matrix})\cr with samples in rows and parameters in columns.}

\item{level}{(\code{number})\cr credibility level to use for the credible intervals.}
}
\value{
A \code{data.frame} with columns \code{median}, \code{lower} and \code{upper}.
}
\description{
Obtain Median and Credible Intervals from MCMC samples
}
\examples{
set.seed(123)
samples <- cbind(rnorm(100, 0, 1), rexp(100, 0.5), rpois(100, 5))
samples_median_ci(samples)
}
