% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sampleStanModel}
\alias{sampleStanModel}
\title{\code{sampleStanModel}}
\usage{
sampleStanModel(object, ..., exe_file)
}
\arguments{
\item{object}{the module.}

\item{...}{additional arguments.}

\item{exe_file}{(\code{string})\cr output file.}
}
\description{
Sample from a Stan Module.
}
