% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModel.R
\name{sampleStanModel}
\alias{sampleStanModel}
\alias{sampleStanModel,JointModel-method}
\title{\code{sampleStanModel}}
\usage{
sampleStanModel(object, ..., exe_file)

\S4method{sampleStanModel}{JointModel}(object, data, ..., exe_file = NULL)
}
\arguments{
\item{object}{the module.}

\item{...}{additional arguments.}

\item{exe_file}{(\code{string})\cr output file.}
}
\description{
Sample from a Stan Module.
}
