% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sampleStanModel}
\alias{sampleStanModel}
\title{\code{sampleStanModel}}
\usage{
sampleStanModel(object, ..., exe_file)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}

\item{exe_file}{TODO}
}
\description{
Description - TODO
}
