% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModel.R
\name{sampleStanModel}
\alias{sampleStanModel}
\alias{sampleStanModel,JointModel-method}
\title{\code{sampleStanModel}}
\usage{
sampleStanModel(object, ...)

\S4method{sampleStanModel}{JointModel}(object, data, ...)
}
\arguments{
\item{object}{the module.}

\item{...}{additional arguments.}

\item{data}{(\code{DataJoint} or \code{list})\cr input data.}
}
\description{
Sample from a Stan Module.
}
