% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModel.R
\name{sampleStanModel}
\alias{sampleStanModel}
\alias{sampleStanModel,JointModel-method}
\title{\code{sampleStanModel}}
\usage{
sampleStanModel(object, ..., exe_dir)

\S4method{sampleStanModel}{JointModel}(object, data, ..., exe_dir)
}
\arguments{
\item{object}{the module.}

\item{...}{additional arguments.}

\item{exe_dir}{(\code{string})\cr Directory for where compiled models should be stored. If not set
will default to \verb{options("jmpost.cache.dir")``, if this is also not set will default to }tempdir()`}

\item{data}{(\code{DataJoint} or \code{list})\cr input data.}
}
\description{
Sample from a Stan Module.
}
