% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/SimLongitudinalClaretBruno.R,
%   R/SimLongitudinalGSF.R, R/SimLongitudinalGSFCov.R,
%   R/SimLongitudinalRandomSlope.R, R/SimLongitudinalRandomSlopeCov.R,
%   R/SimLongitudinalSteinFojo.R, R/SimLongitudinalSteinFojoCov.R,
%   R/SimSurvival.R
\name{sampleSubjects}
\alias{sampleSubjects}
\alias{sampleSubjects.SimLongitudinalClaretBruno}
\alias{sampleSubjects.SimLongitudinalGSF}
\alias{sampleSubjects.SimLongitudinalGSFCov}
\alias{sampleSubjects.SimLongitudinalRandomSlope}
\alias{sampleSubjects.SimLongitudinalRandomSlopeCov}
\alias{sampleSubjects.SimLongitudinalSteinFojo}
\alias{sampleSubjects.SimLongitudinalSteinFojoCov}
\alias{sampleSubjects.SimSurvival}
\title{Generate Simulated Subjects}
\usage{
sampleSubjects(object, subjects_df)

\method{sampleSubjects}{SimLongitudinalClaretBruno}(object, subjects_df)

\method{sampleSubjects}{SimLongitudinalGSF}(object, subjects_df)

\method{sampleSubjects}{SimLongitudinalGSFCov}(object, subjects_df)

\method{sampleSubjects}{SimLongitudinalRandomSlope}(object, subjects_df)

\method{sampleSubjects}{SimLongitudinalRandomSlopeCov}(object, subjects_df)

\method{sampleSubjects}{SimLongitudinalSteinFojo}(object, subjects_df)

\method{sampleSubjects}{SimLongitudinalSteinFojoCov}(object, subjects_df)

\method{sampleSubjects}{SimSurvival}(object, subjects_df)
}
\arguments{
\item{object}{(\code{SimLongitudinal} or \code{SimSurvival}) object to generate subjects from.}

\item{subjects_df}{(\code{data.frame}) the subjects to generate observations for. See details.}
}
\value{
A \code{data.frame} containing simulated subject-level parameters.
}
\description{
Generate Simulated Subjects
}
\details{
The \code{subjects_df} argument should be a \code{data.frame} with 1 row per desired subject to create
with the following columns:
\itemize{
\item \code{study} (\code{factor}) the study identifier.
\item \code{arm} (\code{factor}) the treatment arm identifier.
\item \code{subject} (\code{character}) the subject identifier.
}

This method takes care of generating all the individual subject data required for the
\code{\link{sampleObservations}} method to generate the observations.
}
\examples{
subjects <- data.frame(study = factor("S"), arm = factor("A"), subject = "1")
sim <- SimLongitudinalRandomSlope(slope_mu = 0.01, slope_sigma = 0.5)
sampleSubjects(sim, subjects)
}
