% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModelSamples.R
\name{saveObject.JointModelSamples}
\alias{saveObject.JointModelSamples}
\title{Save a \code{JointModelSamples} object to a file.}
\usage{
\method{saveObject}{JointModelSamples}(object, file, ...)
}
\arguments{
\item{object}{(\code{\link{JointModelSamples}})\cr the object to save.}

\item{file}{(\code{character})\cr the file to save the object to.}

\item{...}{(\code{ANY})\cr additional arguments to \code{\link{saveRDS}}.}
}
\description{
This function is just a wrapper around \code{saveRDS} that saves the object to a file
ensuring that all of the Stan samples are correctly stored. Note that as
CmdStanR objects store their samples as a csv file the samples may be lost
if you call \code{saveRDS} directly on the object.
}
\seealso{
Other saveObject: 
\code{\link{saveObject}()}
}
\concept{saveObject}
