% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{saveObject}
\alias{saveObject}
\title{Save Object to File}
\usage{
saveObject(object, file, ...)
}
\arguments{
\item{object}{(\code{ANY}) \cr object to save.}

\item{file}{(\code{character}) \cr file to save object to.}

\item{...}{(\code{ANY}) \cr additional arguments.}
}
\description{
Save Object to File
}
\seealso{
Other saveObject: 
\code{\link{saveObject.JointModelSamples}()}
}
\concept{saveObject}
