% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Prior.R
\name{set_limits}
\alias{set_limits}
\alias{set_limits.Prior}
\title{Set Constraints}
\usage{
set_limits(object, lower = -Inf, upper = Inf)

\method{set_limits}{Prior}(object, lower = -Inf, upper = Inf)
}
\arguments{
\item{object}{(\code{Prior})\cr a prior distribution to apply constraints to}

\item{lower}{(\code{numeric})\cr upper constraint boundary}
}
\description{
Applies constraints to a prior distribution to ensure any sampled numbers
from the distribution fall within the constraints
}
