% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_sf.R
\name{sf_sld}
\alias{sf_sld}
\alias{sf_ttg}
\alias{sf_dsld}
\title{Stein-Fojo Functionals}
\usage{
sf_sld(time, b, s, g)

sf_ttg(time, b, s, g)

sf_dsld(time, b, s, g)
}
\arguments{
\item{time}{(\code{numeric})\cr time grid.}

\item{b}{(\code{number})\cr baseline.}

\item{s}{(\code{number})\cr shrinkage.}

\item{g}{(\code{number})\cr growth.}
}
\value{
The function results.
}
\description{
Stein-Fojo Functionals
}
\examples{
sf_sld(1:10, 20, 0.3, 0.6)
sf_ttg(1:10, 20, 0.3, 0.6)
sf_dsld(1:10, 20, 0.3, 0.6)
}
\keyword{internal}
