% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSubject.R, R/generics.R,
%   R/DataLongitudinal.R, R/DataSurvival.R, R/DataJoint.R, R/StanModule.R,
%   R/Prior.R, R/Parameter.R, R/ParameterList.R, R/StanModel.R, R/Link.R,
%   R/JointModel.R, R/LinkNone.R
\name{show,DataSubject-method}
\alias{show,DataSubject-method}
\alias{show-object}
\alias{show,DataLongitudinal-method}
\alias{show,DataSurvival-method}
\alias{show,DataJoint-method}
\alias{show,StanModule-method}
\alias{show,Prior-method}
\alias{show,Parameter-method}
\alias{show,ParameterList-method}
\alias{show,StanModel-method}
\alias{show,Link-method}
\alias{show,JointModel-method}
\alias{show,LinkNone-method}
\title{Show an Object}
\usage{
\S4method{show}{DataSubject}(object)

\S4method{show}{DataLongitudinal}(object)

\S4method{show}{DataSurvival}(object)

\S4method{show}{DataJoint}(object)

\S4method{show}{StanModule}(object)

\S4method{show}{Prior}(object)

\S4method{show}{Parameter}(object)

\S4method{show}{ParameterList}(object)

\S4method{show}{StanModel}(object)

\S4method{show}{Link}(object)

\S4method{show}{JointModel}(object)

\S4method{show}{LinkNone}(object)
}
\arguments{
\item{object}{Object to be printed}
}
\description{
Prints an object to the console.
}
