% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StanModule.R
\name{show}
\alias{show}
\alias{show,StanModule-method}
\title{Printing of Different Classes}
\usage{
\S4method{show}{StanModule}(object)
}
\arguments{
\item{object}{what to print.}
}
\description{
These methods print objects of different classes.
}
