% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModelSamples.R
\name{shrinkage.JointModelSamples}
\alias{shrinkage.JointModelSamples}
\title{Extract shrinkage factors from a \code{JointModelSamples} object}
\usage{
\method{shrinkage}{JointModelSamples}(object, ...)
}
\arguments{
\item{object}{(\code{JointModelSamples}) the object to extract shrinkage factors from.}
}
\value{
the shrinkage factors correctly named after the covariate names.
}
\description{
Extract shrinkage factors from a \code{JointModelSamples} object
}
\seealso{
Other shrinkage:
\code{\link[=shrinkage]{shrinkage()}}
}
\concept{shrinkage}
