% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{shrinkage}
\alias{shrinkage}
\title{Extract Covariate Shrinkage Factors from the Posterior Samples}
\usage{
shrinkage(object, ...)
}
\arguments{
\item{object}{(\code{ANY}) the object to extract shrinkage factors from.}

\item{...}{additional arguments added by methods.}
}
\description{
Extract Covariate Shrinkage Factors from the Posterior Samples
}
\seealso{
Other shrinkage:
\code{\link[=shrinkage.JointModelSamples]{shrinkage.JointModelSamples()}}
}
\concept{shrinkage}
