% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_rs.R
\name{sim_lm_random_slope}
\alias{sim_lm_random_slope}
\title{Construct a Simulation Function for Longitudinal Data from Random Slope Model}
\usage{
sim_lm_random_slope(
  intercept = 50,
  slope_mu = c(0.01, 0.03),
  slope_sigma = 0.5,
  sigma = 2,
  link_dsld = 0,
  link_identity = 0
)
}
\arguments{
\item{intercept}{(\code{number})\cr the mean baseline value for each study.}

\item{slope_mu}{(\code{numeric})\cr the population slope for each treatment arm.}

\item{slope_sigma}{(\code{number})\cr the random slope standard deviation.}

\item{sigma}{(\code{number})\cr the variance of the longitudinal values.}

\item{link_dsld}{(\code{number})\cr the link coefficient for the DSLD contribution.}

\item{link_identity}{(\code{number})\cr the link coefficient for the identity contribution.}
}
\value{
A function with argument \code{lm_base} that can be used to simulate
longitudinal data from the corresponding random slope model.
}
\description{
Construct a Simulation Function for Longitudinal Data from Random Slope Model
}
