% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_os_loglogistic}
\alias{sim_os_loglogistic}
\title{Construct a Log Hazard Function for the Log-Logistic Model}
\usage{
sim_os_loglogistic(lambda, p)
}
\arguments{
\item{lambda}{(\code{number})\cr the inverse median parameter.}

\item{p}{(\code{number})\cr the shape parameter.}
}
\value{
A function of \code{time} returning the log hazard.
}
\description{
Construct a Log Hazard Function for the Log-Logistic Model
}
