% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\docType{class}
\name{SimGroup-class}
\alias{SimGroup-class}
\alias{.SimGroup}
\alias{SimGroup}
\title{Define Simulation Group}
\usage{
SimGroup(n, arm, study)
}
\arguments{
\item{n}{(\code{numeric})\cr number of subjects in the group.}

\item{arm}{(\code{character})\cr treatment arm.}

\item{study}{(\code{character})\cr study name.}
}
\description{
Specifies a simulation group to be used by \code{\link[=simulate_joint_data]{simulate_joint_data()}}.
}
\section{Slots}{

\describe{
\item{\code{n}}{(\code{numeric})\cr See arguments.}

\item{\code{arm}}{(\code{character})\cr See arguments.}

\item{\code{study}}{(\code{character})\cr See arguments.}
}}

\examples{
SimGroup(n = 50, arm = "Arm-A", study = "Study-1")
}
