% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinal.R
\docType{class}
\name{SimLongitudinal-class}
\alias{SimLongitudinal-class}
\alias{.SimLongitudinal}
\alias{SimLongitudinal}
\title{Abstract Simulation Class for Longitudinal Data}
\usage{
SimLongitudinal(times = seq(0, 100, 50))
}
\arguments{
\item{times}{(\code{numeric}) the times to generate observations at.}
}
\description{
This class exists to be extended by other classes that simulate longitudinal data.
It is not intended to be used directly.
}
\seealso{
Other SimLongitudinal: 
\code{\link{SimLongitudinalClaretBruno-class}},
\code{\link{SimLongitudinalGSF-class}},
\code{\link{SimLongitudinalRandomSlope-class}},
\code{\link{SimLongitudinalSteinFojo-class}}
}
\concept{SimLongitudinal}
