% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalClaretBrunoCov.R
\docType{class}
\name{SimLongitudinalClaretBrunoCov-class}
\alias{SimLongitudinalClaretBrunoCov-class}
\alias{.SimLongitudinalClaretBrunoCov}
\alias{SimLongitudinalClaretBrunoCov}
\title{Simulate a Claret-Bruno longitudinal model with covariates}
\usage{
SimLongitudinalClaretBrunoCov(
  times = c(-100, -50, 0, 50, 100, 150, 250, 350, 450, 550)/365,
  mu_b_formula = ~study,
  omega_b_formula = ~study,
  mu_g_formula = ~arm,
  omega_g_formula = ~arm,
  mu_c_formula = ~arm,
  omega_c_formula = ~arm,
  mu_p_formula = ~arm,
  omega_p_formula = ~arm,
  mu_b_parametrization = "linear",
  omega_b_parametrization = "log-linear",
  mu_g_parametrization = "linear",
  omega_g_parametrization = "log-linear",
  mu_c_parametrization = "linear",
  omega_c_parametrization = "log-linear",
  mu_p_parametrization = "linear",
  omega_p_parametrization = "log-linear",
  mu_b_intercept = log(60),
  mu_b_coefficients = numeric(),
  omega_b_intercept = log(0.2),
  omega_b_coefficients = numeric(),
  mu_g_intercept = log(1),
  mu_g_coefficients = numeric(),
  omega_g_intercept = log(0.2),
  omega_g_coefficients = numeric(),
  mu_c_intercept = log(0.4),
  mu_c_coefficients = numeric(),
  omega_c_intercept = log(0.2),
  omega_c_coefficients = numeric(),
  mu_p_intercept = log(2),
  mu_p_coefficients = numeric(),
  omega_p_intercept = log(0.2),
  omega_p_coefficients = numeric(),
  sigma = 0.01,
  link_dsld = 0,
  link_ttg = 0,
  link_identity = 0,
  link_growth = 0,
  scaled_variance = FALSE
)
}
\arguments{
\item{times}{(\code{numeric}) Observation times.}

\item{mu_b_formula}{(\code{formula}) One-sided formula for subject-level covariate effects on the baseline tumour size mean.}

\item{omega_b_formula}{(\code{formula}) One-sided formula for subject-level covariate effects on the baseline tumour size standard deviation.}

\item{mu_g_formula}{(\code{formula}) One-sided formula for subject-level covariate effects on the growth mean.}

\item{omega_g_formula}{(\code{formula}) One-sided formula for subject-level covariate effects on the growth standard deviation.}

\item{mu_c_formula}{(\code{formula}) One-sided formula for subject-level covariate effects on the resistance mean.}

\item{omega_c_formula}{(\code{formula}) One-sided formula for subject-level covariate effects on the resistance standard deviation.}

\item{mu_p_formula}{(\code{formula}) One-sided formula for subject-level covariate effects on the growth inhibition mean.}

\item{omega_p_formula}{(\code{formula}) One-sided formula for subject-level covariate effects on the growth inhibition standard deviation.}

\item{mu_b_parametrization}{(\code{character}) Parametrization of the baseline tumour size mean predictor.}

\item{omega_b_parametrization}{(\code{character}) Parametrization of the baseline tumour size standard deviation predictor.
Only \code{"exponential"} and \code{"log-linear"} are supported.}

\item{mu_g_parametrization}{(\code{character}) Parametrization of the growth mean predictor.}

\item{omega_g_parametrization}{(\code{character}) Parametrization of the growth standard deviation predictor. Only \code{"exponential"} and \code{"log-linear"} are supported.}

\item{mu_c_parametrization}{(\code{character}) Parametrization of the resistance mean predictor.}

\item{omega_c_parametrization}{(\code{character}) Parametrization of the resistance standard deviation predictor. Only
\code{"exponential"} and \code{"log-linear"} are supported.}

\item{mu_p_parametrization}{(\code{character}) Parametrization of the growth inhibition mean predictor.}

\item{omega_p_parametrization}{(\code{character}) Parametrization of the growth inhibition standard deviation predictor.
Only \code{"exponential"} and \code{"log-linear"} are supported.}

\item{mu_b_intercept}{(\code{numeric}) Intercept for the baseline tumour size mean predictor.}

\item{mu_b_coefficients}{(\code{numeric}) Covariate coefficients for the baseline tumour size mean predictor.}

\item{omega_b_intercept}{(\code{numeric}) Intercept for the baseline tumour size standard deviation predictor.}

\item{omega_b_coefficients}{(\code{numeric}) Covariate coefficients for the baseline tumour size standard deviation predictor.}

\item{mu_g_intercept}{(\code{numeric}) Intercept for the growth mean predictor.}

\item{mu_g_coefficients}{(\code{numeric}) Covariate coefficients for the growth mean predictor.}

\item{omega_g_intercept}{(\code{numeric}) Intercept for the growth standard deviation predictor.}

\item{omega_g_coefficients}{(\code{numeric}) Covariate coefficients for the growth standard deviation predictor.}

\item{mu_c_intercept}{(\code{numeric}) Intercept for the resistance mean predictor.}

\item{mu_c_coefficients}{(\code{numeric}) Covariate coefficients for the resistance mean predictor.}

\item{omega_c_intercept}{(\code{numeric}) Intercept for the resistance standard deviation predictor.}

\item{omega_c_coefficients}{(\code{numeric}) Covariate coefficients for the resistance standard deviation predictor.}

\item{mu_p_intercept}{(\code{numeric}) Intercept for the growth inhibition mean predictor.}

\item{mu_p_coefficients}{(\code{numeric}) Covariate coefficients for the growth inhibition mean predictor.}

\item{omega_p_intercept}{(\code{numeric}) Intercept for the growth inhibition standard deviation predictor.}

\item{omega_p_coefficients}{(\code{numeric}) Covariate coefficients for the growth inhibition standard deviation predictor.}

\item{sigma}{(\code{numeric}) Observation-error standard deviation.}

\item{link_dsld}{(\code{numeric}) Coefficient for the DSLD association link.}

\item{link_ttg}{(\code{numeric}) Coefficient for the TTG association link.}

\item{link_identity}{(\code{numeric}) Coefficient for the identity association link.}

\item{link_growth}{(\code{numeric}) Coefficient for the growth association link.}

\item{scaled_variance}{(\code{logical}) Whether to use multiplicative observation error.}
}
\value{
A \code{SimLongitudinalClaretBrunoCov} object.
}
\description{
Simulate a Claret-Bruno longitudinal model with covariates
}
