% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalClaretBrunoCov.R
\docType{class}
\name{SimLongitudinalClaretBrunoCov-class}
\alias{SimLongitudinalClaretBrunoCov-class}
\alias{.SimLongitudinalClaretBrunoCov}
\alias{SimLongitudinalClaretBrunoCov}
\title{Simulate a Claret-Bruno longitudinal model with covariates}
\usage{
SimLongitudinalClaretBrunoCov(
  times = c(-100, -50, 0, 50, 100, 150, 250, 350, 450, 550)/365,
  mu_b_formula = ~study,
  omega_b_formula = ~study,
  mu_g_formula = ~arm,
  omega_g_formula = ~arm,
  mu_c_formula = ~arm,
  omega_c_formula = ~arm,
  mu_p_formula = ~arm,
  omega_p_formula = ~arm,
  mu_b_parametrization = "linear",
  omega_b_parametrization = "log-linear",
  mu_g_parametrization = "linear",
  omega_g_parametrization = "log-linear",
  mu_c_parametrization = "linear",
  omega_c_parametrization = "log-linear",
  mu_p_parametrization = "linear",
  omega_p_parametrization = "log-linear",
  mu_b_intercept = log(60),
  mu_b_coefficients = numeric(),
  omega_b_intercept = log(0.2),
  omega_b_coefficients = numeric(),
  mu_g_intercept = log(1),
  mu_g_coefficients = numeric(),
  omega_g_intercept = log(0.2),
  omega_g_coefficients = numeric(),
  mu_c_intercept = log(0.4),
  mu_c_coefficients = numeric(),
  omega_c_intercept = log(0.2),
  omega_c_coefficients = numeric(),
  mu_p_intercept = log(2),
  mu_p_coefficients = numeric(),
  omega_p_intercept = log(0.2),
  omega_p_coefficients = numeric(),
  sigma = 0.01,
  link_dsld = 0,
  link_ttg = 0,
  link_identity = 0,
  link_growth = 0,
  scaled_variance = FALSE
)
}
\arguments{
\item{times}{Observation times.}

\item{mu_b_formula, mu_g_formula, mu_c_formula, mu_p_formula}{One-sided covariate formulas for the mean parameters.}

\item{omega_b_formula, omega_g_formula, omega_c_formula, omega_p_formula}{One-sided covariate formulas for the variance parameters.}

\item{mu_b_parametrization, mu_g_parametrization, mu_c_parametrization, mu_p_parametrization}{Parametrization for the mean parameters.}

\item{omega_b_parametrization, omega_g_parametrization, omega_c_parametrization, omega_p_parametrization}{Parametrization for the variance parameters.}

\item{mu_b_intercept, mu_g_intercept, mu_c_intercept, mu_p_intercept}{Intercepts for the mean parameters.}

\item{mu_b_coefficients, mu_g_coefficients, mu_c_coefficients, mu_p_coefficients}{Coefficients for the mean parameters.}

\item{omega_b_intercept, omega_g_intercept, omega_c_intercept, omega_p_intercept}{Intercepts for the variance parameters.}

\item{omega_b_coefficients, omega_g_coefficients, omega_c_coefficients, omega_p_coefficients}{Coefficients for the variance parameters.}

\item{sigma}{Observation-error standard deviation.}

\item{link_dsld, link_ttg, link_identity, link_growth}{Link coefficients.}

\item{scaled_variance}{Whether to use multiplicative observation error.}
}
\value{
A \code{SimLongitudinalClaretBrunoCov} object.
}
\description{
Simulate a Claret-Bruno longitudinal model with covariates
}
