% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalGSFCov.R
\docType{class}
\name{SimLongitudinalGSFCov-class}
\alias{SimLongitudinalGSFCov-class}
\alias{.SimLongitudinalGSFCov}
\alias{SimLongitudinalGSFCov}
\title{Simulate a Generalized Stein-Fojo model with covariates}
\usage{
SimLongitudinalGSFCov(
  times = c(-100, -50, 0, 50, 100, 150, 250, 350, 450, 550)/365,
  mu_b_formula = ~study,
  omega_b_formula = ~study,
  mu_s_formula = ~arm,
  omega_s_formula = ~arm,
  mu_g_formula = ~arm,
  omega_g_formula = ~arm,
  mu_phi_formula = ~arm,
  omega_phi_formula = ~arm,
  mu_b_parametrization = "linear",
  omega_b_parametrization = "log-linear",
  mu_s_parametrization = "linear",
  omega_s_parametrization = "log-linear",
  mu_g_parametrization = "linear",
  omega_g_parametrization = "log-linear",
  mu_phi_parametrization = "linear",
  omega_phi_parametrization = "log-linear",
  mu_b_intercept = log(60),
  mu_b_coefficients = numeric(),
  omega_b_intercept = log(0.2),
  omega_b_coefficients = numeric(),
  mu_s_intercept = log(0.5),
  mu_s_coefficients = numeric(),
  omega_s_intercept = log(0.2),
  omega_s_coefficients = numeric(),
  mu_g_intercept = log(0.3),
  mu_g_coefficients = numeric(),
  omega_g_intercept = log(0.2),
  omega_g_coefficients = numeric(),
  mu_phi_intercept = qlogis(0.5),
  mu_phi_coefficients = numeric(),
  omega_phi_intercept = log(0.2),
  omega_phi_coefficients = numeric(),
  sigma = 0.01,
  link_dsld = 0,
  link_ttg = 0,
  link_identity = 0,
  link_growth = 0,
  link_shrinkage = 0,
  scaled_variance = FALSE
)
}
\arguments{
\item{times}{Observation times.}

\item{mu_b_formula, omega_b_formula, mu_s_formula, omega_s_formula, mu_g_formula, omega_g_formula, mu_phi_formula, omega_phi_formula}{One-sided covariate formulas.}

\item{mu_b_parametrization, omega_b_parametrization, mu_s_parametrization, omega_s_parametrization, mu_g_parametrization, omega_g_parametrization, mu_phi_parametrization, omega_phi_parametrization}{Predictor parametrizations.}

\item{mu_b_intercept, mu_b_coefficients, omega_b_intercept, omega_b_coefficients, mu_s_intercept, mu_s_coefficients, omega_s_intercept, omega_s_coefficients, mu_g_intercept, mu_g_coefficients, omega_g_intercept, omega_g_coefficients, mu_phi_intercept, mu_phi_coefficients, omega_phi_intercept, omega_phi_coefficients}{Predictor coefficients.}

\item{sigma}{Observation-error standard deviation.}

\item{link_dsld, link_ttg, link_identity, link_growth, link_shrinkage}{Link coefficients.}

\item{scaled_variance}{Whether to use multiplicative observation error.}
}
\value{
A \code{SimLongitudinalGSFCov} object.
}
\description{
Simulate a Generalized Stein-Fojo model with covariates
}
