% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalRandomSlope.R
\docType{class}
\name{SimLongitudinalRandomSlope-class}
\alias{SimLongitudinalRandomSlope-class}
\alias{.SimLongitudinalRandomSlope}
\alias{SimLongitudinalRandomSlope}
\title{Simulate Longitudinal Data from a Random Slope Model}
\usage{
SimLongitudinalRandomSlope(
  times = c(-100, -50, 0, 50, 100, 150, 250, 350, 450, 550),
  intercept = 50,
  slope_mu = c(0.01, 0.03),
  slope_sigma = 0.5,
  sigma = 2,
  link_dsld = 0,
  link_identity = 0
)
}
\arguments{
\item{times}{(\code{numeric})\cr the times to generate observations at.}

\item{intercept}{(\code{number})\cr the mean baseline value for each study.}

\item{slope_mu}{(\code{numeric})\cr the population slope for each treatment arm.}

\item{slope_sigma}{(\code{number})\cr the random slope standard deviation.}

\item{sigma}{(\code{number})\cr the variance of the longitudinal values.}

\item{link_dsld}{(\code{number})\cr the link coefficient for the DSLD contribution.}

\item{link_identity}{(\code{number})\cr the link coefficient for the identity contribution.}
}
\description{
Simulate Longitudinal Data from a Random Slope Model
}
\section{Slots}{

\describe{
\item{\code{intercept}}{(\code{numeric})\cr See arguments.}

\item{\code{slope_mu}}{(\code{numeric})\cr See arguments.}

\item{\code{slope_sigma}}{(\code{numeric})\cr See arguments.}

\item{\code{sigma}}{(\code{numeric})\cr See arguments.}

\item{\code{link_dsld}}{(\code{numeric})\cr See arguments.}

\item{\code{link_identity}}{(\code{numeric})\cr See arguments.}
}}

\seealso{
Other SimLongitudinal: 
\code{\link{SimLongitudinal-class}},
\code{\link{SimLongitudinalGSF-class}},
\code{\link{SimLongitudinalSteinFojo-class}}
}
\concept{SimLongitudinal}
