% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalRandomSlopeCov.R
\docType{class}
\name{SimLongitudinalRandomSlopeCov-class}
\alias{SimLongitudinalRandomSlopeCov-class}
\alias{.SimLongitudinalRandomSlopeCov}
\title{Simulate a random-slope longitudinal model with covariates}
\description{
Simulate a random-slope longitudinal model with covariates
}
