% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalSteinFojo.R
\docType{class}
\name{SimLongitudinalSteinFojo-class}
\alias{SimLongitudinalSteinFojo-class}
\alias{.SimLongitudinalSteinFojo}
\title{Construct a Simulation Function for Longitudinal Data from Stein-Fojo Model}
\arguments{
\item{sigma}{(\code{number})\cr the variance of the longitudinal values.}

\item{mu_s}{(\code{numeric})\cr the mean shrinkage rates for the two treatment arms.}

\item{mu_g}{(\code{numeric})\cr the mean growth rates for the two treatment arms.}

\item{mu_b}{(\code{numeric})\cr the mean baseline values for the two treatment arms.}

\item{omega_b}{(\code{number})\cr the baseline value standard deviation.}

\item{omega_s}{(\code{number})\cr the shrinkage rate standard deviation.}

\item{omega_g}{(\code{number})\cr the growth rate standard deviation.}

\item{link_dsld}{(\code{number})\cr the link coefficient for the derivative contribution.}

\item{link_ttg}{(\code{number})\cr the link coefficient for the time-to-growth contribution.}

\item{link_identity}{(\code{number})\cr the link coefficient for the SLD Identity contribution.}
}
\value{
A function with argument \code{lm_base} that can be used to simulate
longitudinal data from the corresponding Stein-Fojo model.
}
\description{
Construct a Simulation Function for Longitudinal Data from Stein-Fojo Model
}
