% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitudinalSteinFojoCov.R
\docType{class}
\name{SimLongitudinalSteinFojoCov-class}
\alias{SimLongitudinalSteinFojoCov-class}
\alias{.SimLongitudinalSteinFojoCov}
\title{Simulate a Stein-Fojo longitudinal model with covariates}
\description{
Simulate a Stein-Fojo longitudinal model with covariates
}
