% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSurvival.R
\name{SimSurvivalExponential}
\alias{SimSurvivalExponential}
\title{Construct a Log Hazard Function for the Exponential Model}
\usage{
SimSurvivalExponential(lambda, ...)
}
\arguments{
\item{lambda}{(\code{number})\cr the rate parameter.}
}
\value{
A function of \code{time} returning the log hazard.
}
\description{
Construct a Log Hazard Function for the Exponential Model
}
