% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSurvival.R
\name{SimSurvivalLogLogistic}
\alias{SimSurvivalLogLogistic}
\title{Construct a Log Hazard Function for the Log-Logistic Model}
\usage{
SimSurvivalLogLogistic(
  a,
  b,
  time_max = 2000,
  time_step = 1,
  lambda_censor = 1/3000,
  beta_cont = 0.2,
  beta_cat = c(A = 0, B = -0.4, C = 0.2)
)
}
\arguments{
\item{a}{(\code{number})\cr the scale parameter.}

\item{b}{(\code{number})\cr the shape parameter.}
}
\value{
A function of \code{time} returning the log hazard.
}
\description{
Construct a Log Hazard Function for the Log-Logistic Model
}
