% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSurvival.R
\name{SimSurvivalLogLogistic}
\alias{SimSurvivalLogLogistic}
\title{Construct a Log Hazard Function for the Log-Logistic Model}
\usage{
SimSurvivalLogLogistic(a, b, ...)
}
\arguments{
\item{a}{(\code{number})\cr the scale parameter.}

\item{b}{(\code{number})\cr the shape parameter.}
}
\value{
A function of \code{time} returning the log hazard.
}
\description{
Construct a Log Hazard Function for the Log-Logistic Model
}
