% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSurvival.R
\name{SimSurvivalWeibullPH}
\alias{SimSurvivalWeibullPH}
\title{Construct a Log Hazard Function for the Weibull Model}
\usage{
SimSurvivalWeibullPH(lambda, gamma, ...)
}
\arguments{
\item{lambda}{(\code{number})\cr the scale parameter.}

\item{gamma}{(\code{number})\cr the shape parameter.}
}
\value{
A function of \code{time} returning the log hazard.
}
\description{
Construct a Log Hazard Function for the Weibull Model
}
