% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.JointModelSamples}
\alias{simulate.JointModelSamples}
\title{Simulate Patients from Posterior Predictive Distribution}
\usage{
\method{simulate}{JointModelSamples}(
  object,
  newdata = NULL,
  ...,
  times,
  time_max,
  time_step,
  lambda_censor
)
}
\arguments{
\item{object}{A \link{JointModelSamples} object}

\item{newdata}{A data frame containing data in the same format as
the survival data.
@param ... Unused.}

\item{time_max}{(\code{number})\cr the maximum time to simulate to.}

\item{time_step}{(\code{number})\cr the time interval between evaluating the log-hazard function.}

\item{lambda_censor}{(\code{number})\cr the censoring rate.}
}
\description{
Simulate Patients from Posterior Predictive Distribution
}
\details{
Simulates a set of patients based on the covariates of those
used in the model fit or from \code{newdata}, which must contain the
same column names and factor levels.
}
