% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.JointModelSamples}
\alias{simulate.JointModelSamples}
\title{Simulate Patients from Posterior Predictive Distribution}
\usage{
\method{simulate}{JointModelSamples}(
  object,
  newdata = NULL,
  ...,
  times = c(0, 10, 50, 100),
  time_max = 2000,
  time_step = 1,
  lambda_censor = 1/3000,
  scaled_variance = TRUE
)
}
\arguments{
\item{object}{A \link{JointModelSamples} object}

\item{newdata}{A data frame containing data in the same format as
the \code{object@data@survival@data}. Importantly, it should contain the same covariates and factor levels
as the variables used in the survival formula \code{object@data@survival@formula} and
the same columns used for \code{study}, \code{id}, and \code{arm}.
@param ... Unused.}

\item{time_max}{(\code{number})\cr the maximum time to simulate to.}

\item{time_step}{(\code{number})\cr the time interval between evaluating the log-hazard function.}

\item{lambda_censor}{(\code{number})\cr the censoring rate, as the parameter of an exponential distribution.}

\item{scaled_variance}{Should variance be scaled by the expected value. Must be set the same as was used for
model fitting.}
}
\description{
Simulate Patients from Posterior Predictive Distribution
}
\details{
Simulates a set of patients based on the covariates of those
used in the model fit or from \code{newdata}, which must contain the
same column names and factor levels.
}
