% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_joint_data}
\alias{simulate_joint_data}
\title{Simulating Joint Longitudinal and Time-to-Event Data}
\usage{
simulate_joint_data(
  n_arm = c(50, 80),
  times = 1:2000,
  lambda_cen = 1/3,
  beta_cont = 0.2,
  beta_cat = c(A = 0, B = -0.4, C = 0.2),
  lm_fun,
  os_fun
)
}
\arguments{
\item{n_arm}{(\code{numeric})\cr numbers of patients per treatment arm.}

\item{times}{(\code{numeric})\cr time grid, e.g. specifying the days after randomization.}

\item{lambda_cen}{(\code{number})\cr rate of the exponential censoring distribution.}

\item{beta_cont}{(\code{number})\cr coefficient for the continuous covariate.}

\item{beta_cat}{(\code{numeric})\cr coefficients for the categorical covariate levels.}

\item{lm_fun}{(\code{function})\cr function of \code{lm_base} generating the longitudinal model outcomes.}

\item{os_fun}{(\code{function})\cr function of \code{lm_base} generating the survival model outcomes.}
}
\value{
List with simulated \code{lm} (longitudinal) and \code{os} (survival) data sets.
}
\description{
Simulating Joint Longitudinal and Time-to-Event Data
}
