% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{size}
\alias{size}
\title{\code{size}}
\usage{
size(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\value{
A \code{list} of parameter sizes.
}
\description{
\code{size}
}
\keyword{internal}
