% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Parameter.R, R/ParameterList.R
\name{size}
\alias{size}
\alias{size,Parameter-method}
\alias{size,ParameterList-method}
\title{\code{size}}
\usage{
size(object)

\S4method{size}{Parameter}(object)

\S4method{size}{ParameterList}(object)
}
\arguments{
\item{object}{where to get the parameter sizes from.}
}
\description{
Obtain the \code{list} of parameter sizes.
}
\keyword{internal}
