% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{size}
\alias{size}
\title{\code{size}}
\usage{
size(object)
}
\arguments{
\item{object}{where to get the parameter sizes from.}
}
\description{
Obtain the \code{list} of parameter sizes.
}
\keyword{internal}
