% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_generics.R
\name{standard-link-user}
\alias{standard-link-user}
\alias{linkNone}
\alias{linkTTG}
\alias{linkDSLD}
\alias{linkIdentity}
\title{Standard Links}
\usage{
linkNone()

linkTTG(prior, model = PromiseLongitudinalModel(), ...)

linkDSLD(prior, model = PromiseLongitudinalModel(), ...)

linkIdentity(prior, model = PromiseLongitudinalModel(), ...)
}
\arguments{
\item{prior}{(\code{\link{Prior}}) \cr A \code{\link{Prior}} object.}

\item{model}{(\code{\link{LongitudinalModel}}) \cr A \code{\link{LongitudinalModel}} object.}

\item{...}{Not used.}
}
\description{
These functions are used to enable the use of the corresponding link function between
the survival and longitudinal models in a joint model. Note that the exact implementation
of the link function is model specific, see
\code{vignette("Statistical Specifications", package = "jmpost")} for more details.
}
\section{Functions}{
\itemize{
\item \code{linkNone()}: No link (fit the survival and longitudinal models independently)

\item \code{linkTTG()}: Time to growth link

\item \code{linkDSLD()}: Derivative of the SLD over time link

\item \code{linkIdentity()}: Current SLD value link

}}
