% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModel.R
\docType{class}
\name{StanModel-class}
\alias{StanModel-class}
\alias{.StanModel}
\alias{StanModel}
\title{Stan Model Object and Constructor Function}
\usage{
StanModel(stan, parameters, name = "<Unnamed>")
}
\arguments{
\item{stan}{(\code{StanModule}) code containing the model-specific Stan code specification.}

\item{parameters}{(\code{ParameterList}) the parameter declaration and prior specification.}

\item{name}{(\code{character}) display name for the model object.}
}
\value{
A \code{StanModel} object.
}
\description{
Stan Model Object and Constructor Function
}
\section{Slots}{

\describe{
\item{\code{stan}}{(\code{StanModule})\cr See Arguments.}

\item{\code{parameters}}{(\code{ParameterList})\cr See Arguments.}

\item{\code{name}}{(\code{character})\cr display name for the model object.}
}}

\examples{
StanModel(StanModule(), ParameterList(), name = "Example model")
}
\seealso{
Other StanModel:
\code{\link[=as.list.StanModel]{as.list.StanModel()}}
}
\concept{StanModel}
