% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModel.R
\docType{class}
\name{StanModel-class}
\alias{StanModel-class}
\alias{.StanModel}
\alias{StanModel}
\alias{as.list,StanModel-method}
\title{\code{StanModel}}
\usage{
StanModel(stan, parameters, ...)

\S4method{as.list}{StanModel}(x)
}
\arguments{
\item{stan}{(\code{StanModule})\cr code containing the Stan code specification.}

\item{parameters}{(\code{ParameterList})\cr the parameter specification.}

\item{...}{additional arguments for \code{\link[=StanModel]{StanModel()}}.}

\item{x}{(\code{StanModel})\cr to be coerced to a list.}
}
\description{
\code{StanModel}
}
\section{Slots}{

\describe{
\item{\code{stan}}{(\code{StanModule})\cr code containing the Stan code specification.}

\item{\code{parameters}}{(\code{ParameterList})\cr the parameter specification.}
}}

