% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModel.R
\docType{class}
\name{StanModel-class}
\alias{StanModel-class}
\alias{.StanModel}
\alias{StanModel}
\title{Stan Model Object and Constructor Function}
\usage{
StanModel(stan, parameters, ...)
}
\arguments{
\item{stan}{(\code{StanModule})\cr code containing the Stan code specification.}

\item{parameters}{(\code{ParameterList})\cr the parameter specification.}
}
\description{
Stan Model Object and Constructor Function
}
\section{Slots}{

\describe{
\item{\code{stan}}{(\code{StanModule})\cr See Arguements.}

\item{\code{parameters}}{(\code{ParameterList})\cr See Arguements.}
}}

\seealso{
Other StanModel: 
\code{\link{as.list.StanModel}()}
}
\concept{StanModel}
