% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModel.R
\name{stanmodel_arguments}
\alias{stanmodel_arguments}
\title{\code{StanModel} Function Arguments}
\arguments{
\item{stan}{(\code{StanModule})\cr code containing the Stan code specification.}

\item{parameters}{(\code{ParameterList})\cr the parameter specification.}

\item{parameter}{(\code{ParameterList})\cr the (single) parameter specification.}

\item{name}{(\code{character})\cr display name for the model object.}

\item{...}{additional arguments for \code{\link[=StanModel]{StanModel()}}.}
}
\description{
The documentation lists all the conventional arguments for wrappers around
\code{\link[=StanModel]{StanModel()}}.
}
\keyword{internal}
