% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanModule.R
\docType{class}
\name{StanModule-class}
\alias{StanModule-class}
\alias{.StanModule}
\alias{StanModule}
\title{\code{StanModule} Object and Constructor Function}
\usage{
StanModule(x = "", ...)
}
\arguments{
\item{x}{(\code{string})\cr file path to a Stan program or a character vector
of Stan code to be parsed.}

\item{...}{additional arguments passed to the constructor.}
}
\description{
\code{StanModule} Object and Constructor Function
}
\section{Slots}{

\describe{
\item{\code{functions}}{(\code{character})\cr the \code{functions} block.}

\item{\code{data}}{(\code{character})\cr the \code{data} block.}

\item{\code{transformed_data}}{(\code{character})\cr the \code{transformed_data} block.}

\item{\code{parameters}}{(\code{character})\cr the \code{parameters} block.}

\item{\code{transformed_parameters}}{(\code{character})\cr the \code{transformed_parameters} block.}

\item{\code{model}}{(\code{character})\cr the \code{model} block.}

\item{\code{generated_quantities}}{(\code{character})\cr the \code{generated_quantities} block.}
}}

\seealso{
Other StanModule: 
\code{\link{as.character.StanModule}()},
\code{\link{as.list.StanModule}()},
\code{\link{as_print_string.StanModule}()}
}
\concept{StanModule}
