% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_module.R, R/hazard_link.R, R/JointModel.R
\docType{class}
\name{StanModule-class}
\alias{StanModule-class}
\alias{StanModule}
\alias{initialize,StanModule-method}
\alias{initialize,HazardLink-method}
\alias{initialize,JointModel-method}
\title{StanAll class object}
\usage{
\S4method{initialize}{StanModule}(
  .Object,
  ...,
  functions = "",
  data = "",
  parameters = "",
  transformed_parameters = "",
  model = "",
  generated_quantities = "",
  priors = list(),
  inits = list()
)

\S4method{initialize}{HazardLink}(.Object, ..., stan = StanModule(), contribution, parameters)

\S4method{initialize}{JointModel}(.Object, ..., long, os, util = TRUE, name = "stan_model.stan")
}
\description{
StanAll class object
}
\section{Slots}{

\describe{
\item{\code{functions}}{Character, the functions part of a stan model.}

\item{\code{data}}{Character, the data part of a stan model.}

\item{\code{parameters}}{Character, the parameters part of a stan model.}

\item{\code{transformed_parameters}}{Character, the transformed_parameters part of a stan model.}

\item{\code{model}}{Character, the model part of a stan model.}

\item{\code{priors}}{List, the priors part of a stan model.}

\item{\code{generated_quantities}}{Character, the generated_quantities part of a stan model.}

\item{\code{includes}}{Character}

\item{\code{inits}}{List with the initial values of the stan model.}
}}

