% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_module.R
\docType{class}
\name{StanModule-class}
\alias{StanModule-class}
\alias{StanModule}
\alias{initialize,StanModule-method}
\title{StanAll class object}
\usage{
\S4method{initialize}{StanModule}(
  .Object,
  ...,
  functions,
  data,
  parameters,
  transformed_parameters,
  priors,
  generated_quantities,
  inits
)
}
\arguments{
\item{functions}{TODO}

\item{data}{TODO}

\item{parameters}{TODO}

\item{transformed_parameters}{TODO}

\item{priors}{TODO}

\item{generated_quantities}{TODO}

\item{inits}{TODO}
}
\description{
StanAll class object
}
\section{Slots}{

\describe{
\item{\code{functions}}{Character, the functions part of a stan model.}

\item{\code{data}}{Character, the data part of a stan model.}

\item{\code{parameters}}{Character, the parameters part of a stan model.}

\item{\code{transformed_parameters}}{Character, the transformed_parameters part of a stan model.}

\item{\code{model}}{Character, the model part of a stan model.}

\item{\code{priors}}{List, the priors part of a stan model.}

\item{\code{generated_quantities}}{Character, the generated_quantities part of a stan model.}

\item{\code{includes}}{Character}

\item{\code{inits}}{List with the initial values of the stan model.}
}}

