% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\name{subset.DataJoint}
\alias{subset.DataJoint}
\title{Subsetting \code{DataJoint} as a \code{data.frame}}
\usage{
\method{subset}{DataJoint}(x, patients, ...)
}
\arguments{
\item{x}{(\code{DataJoint}) \cr object created by \code{\link[=DataJoint]{DataJoint()}}.}

\item{patients}{(\code{character} or \code{list})\cr patients that you wish to subset the \code{data.frame}
to contain. See details.}

\item{...}{Not used.}
}
\description{
Coerces the object into a \code{data.frame} containing just event times and status
filtering for specific patients. If \code{patients} is a list then an additional variable \code{group} will be added
onto the dataset specifying which group the row belongs to.
}
\examples{
\dontrun{
pts <- c("PT1", "PT3", "PT4")
subset(x, pts)

groups <- list(
    "g1" = c("PT1", "PT3", "PT4"),
    "g2" = c("PT2", "PT3")
)
subset(x, groups)
}
}
\seealso{
Other DataJoint: 
\code{\link{DataJoint-class}},
\code{\link{as.list.DataJoint}()}
}
\concept{DataJoint}
