% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataJoint.R
\name{subset_and_add_grouping}
\alias{subset_and_add_grouping}
\title{\code{subset_and_add_grouping}}
\usage{
subset_and_add_grouping(dat, groupings)
}
\arguments{
\item{dat}{(\code{data.frame}) \cr must have a column called \code{patient} which corresponds to the
values passed to \code{groupings}.}

\item{groupings}{(\code{character} or \code{list})\cr patients that you wish to subset the dataset
to contain. If \code{groupings} is a list then an additional variable \code{group} will be added
onto the dataset specifying which group the row belongs to.}
}
\description{
\code{subset_and_add_grouping}
}
\details{
Example of usage

\if{html}{\out{<div class="sourceCode">}}\preformatted{pts <- c("PT1", "PT3", "PT4")
subset_and_add_grouping(dat, pts)

groups <- list(
    "g1" = c("PT1", "PT3", "PT4"),
    "g2" = c("PT2", "PT3")
)
subset_and_add_grouping(dat, groups)
}\if{html}{\out{</div>}}
}
\keyword{internal}
