% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{summary,SurvivalQuantities-method}
\alias{summary,SurvivalQuantities-method}
\title{summary}
\usage{
\S4method{summary}{SurvivalQuantities}(object, conf.level = 0.95)
}
\description{
This method returns a \code{data.frame} of key quantities (survival / log-hazard / etc)
for selected patients at a given set of time points.
}
\seealso{
Other SurvivalQuantities: 
\code{\link{SurvivalQuantities-class}},
\code{\link{as.data.frame,SurvivalQuantities-method}},
\code{\link{autoplot,SurvivalQuantities-method}}

Other summary: 
\code{\link{summary}()}
}
\concept{SurvivalQuantities}
\concept{summary}
