% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{summary,SurvivalQuantities-method}
\alias{summary,SurvivalQuantities-method}
\title{summary}
\usage{
\S4method{summary}{SurvivalQuantities}(object, conf.level = 0.95)
}
\arguments{
\item{object}{(\code{\link{SurvivalQuantities}}) \cr Survival Quantities.}

\item{conf.level}{(\code{numeric}) \cr confidence level of the interval.}
}
\description{
This method returns a \code{data.frame} of key quantities (survival / log-hazard / etc)
for selected patients at a given set of time points.
}
\seealso{
Other \code{SurvivalQuantities}: 
\code{\link{SurvivalQuantities-class}},
\code{\link{as.data.frame,SurvivalQuantities-method}},
\code{\link{autoplot,SurvivalQuantities-method}}

Other \code{summary}: 
\code{\link{summary}()}
}
\concept{\code{SurvivalQuantities}}
\concept{\code{summary}}
