% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantities.R
\name{summary.Quantities}
\alias{summary.Quantities}
\title{summary}
\usage{
\method{summary}{Quantities}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{(\code{\link{Quantities}}) \cr generated quantities.}

\item{conf.level}{(\code{numeric}) \cr confidence level of the interval.}

\item{...}{not used.}
}
\value{
A \code{data.frame} with the following variables:
\itemize{
\item \code{median} (\code{numeric}) \cr the median value of the quantity.
\item \code{lower} (\code{numeric}) \cr the lower CI value of the quantity.
\item \code{upper} (\code{numeric}) \cr the upper CI value of the quantity.
\item \code{time} (\code{numeric}) \cr the time point which the quantity is for.
\item \code{group} (\code{character}) \cr which group the quantity belongs to.
\item \code{type} (\code{character}) \cr what type of quantity is it.
}
}
\description{
This method returns a summary statistic \code{data.frame} of the quantities. Note that this
is just an internal utility method in order to share common code between
\link{LongitudinalQuantities} and \link{SurvivalQuantities}
}
\seealso{
Other Quantities: 
\code{\link{Quantities-class}},
\code{\link{as.data.frame.Quantities}()},
\code{\link{as_print_string.Quantities}()}
}
\concept{Quantities}
\keyword{internal}
