% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{summary.SurvivalQuantities}
\alias{summary.SurvivalQuantities}
\title{summary}
\usage{
\method{summary}{SurvivalQuantities}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{(\code{SurvivalQuantities}) survival quantities.}

\item{conf.level}{(\code{numeric}) confidence level of the interval.}

\item{...}{not used.}
}
\description{
This method returns a \code{data.frame} of the longitudinal quantities.
}
\seealso{
Other SurvivalQuantities:
\code{\link{SurvivalQuantities-class}},
\code{\link[=as.data.frame.SurvivalQuantities]{as.data.frame.SurvivalQuantities()}},
\code{\link[=autoplot.SurvivalQuantities]{autoplot.SurvivalQuantities()}},
\code{\link[=brierScore.SurvivalQuantities]{brierScore.SurvivalQuantities()}}

Other summary:
\code{\link[=summary.LongitudinalQuantities]{summary.LongitudinalQuantities()}}
}
\concept{SurvivalQuantities}
\concept{summary}
