% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\name{summary.SurvivalQuantities}
\alias{summary.SurvivalQuantities}
\title{summary}
\usage{
\method{summary}{SurvivalQuantities}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{(\code{\link{SurvivalQuantities}}) \cr survival quantities.}

\item{conf.level}{(\code{numeric}) \cr confidence level of the interval.}

\item{...}{not used.}
}
\description{
This method returns a \code{data.frame} of key quantities (survival / log-hazard / etc)
}
\seealso{
Other SurvivalQuantities: 
\code{\link{SurvivalQuantities-class}},
\code{\link{as.data.frame.SurvivalQuantities}()},
\code{\link{as_print_string.SurvivalQuantities}()},
\code{\link{autoplot.SurvivalQuantities}()}

Other summary: 
\code{\link{summary.LongitudinalQuantities}()}
}
\concept{SurvivalQuantities}
\concept{summary}
