% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModelSamples.R
\name{survival}
\alias{survival}
\alias{survival,JointModelSamples-method}
\title{\code{survival}}
\usage{
survival(object, ...)

\S4method{survival}{JointModelSamples}(object, patients = NULL, ...)
}
\arguments{
\item{object}{samples to extract the survival function values from.}

\item{...}{additional options.}

\item{patients}{(\code{character} or \code{NULL})\cr optional subset of patients for
which the survival function samples should be extracted, the default \code{NULL}
meaning all patients.}
}
\description{
Obtain the survival function samples from \code{\link{JointModelSamples}}.
}
