% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalSamples.R
\name{survival_plot}
\alias{survival_plot}
\title{Survival Plot}
\usage{
survival_plot(
  data,
  add_ci = TRUE,
  add_wrap = TRUE,
  kmdf = NULL,
  y_label = expression(S(t)),
  x_label = expression(t)
)
}
\arguments{
\item{data}{(\code{data.frame})\cr A \code{data.frame} of summary statistics for a survival
curve to be plotted. See details.}

\item{add_ci}{(\code{logical})\cr Should confidence intervals be added? Default = \code{TRUE}.}

\item{add_wrap}{(\code{logical})\cr Should the plots be wrapped by \code{data$group}? Default = \code{TRUE}.}

\item{kmdf}{(\code{data.frame} or \code{NULL})\cr A \code{data.frame} of event times and status used to plot
overlaying KM curves. If \code{NULL} no KM curve will be plotted. See details.}

\item{y_label}{(\code{character} or \code{expression}) \cr Label to display on the y-axis.
Default = \code{expression(S(t))}}

\item{x_label}{(\code{character} or \code{expression}) \cr Label to display on the x-axis.}
}
\description{
Internal plotting function to create survival plots with KM curve overlays
This function predominately exists to extract core logic into its own function
to enable easier unit testing.
}
\details{
\subsection{\code{data}}{

Should contain the following columns:
\itemize{
\item \code{time} - Time point
\item \code{group} - The group in which the observation belongs to
\item \code{median} - The median value for the summary statistic
\item \code{upper} - The upper 95\% CI for the summary statistic
\item \code{lower} - The lower 95\% CI for the summary statistic
}
}

\subsection{\code{kmdf}}{

Should contain the following columns:
\itemize{
\item \code{time} - The time at which an event occurred
\item \code{event} - 1/0 status indicator for the event
\item \code{group} - Which group the event belongs to, should correspond to values in \code{data$group}
}
}
}
\keyword{internal}
