% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalGamma.R
\docType{class}
\name{SurvivalGamma-class}
\alias{SurvivalGamma-class}
\alias{.SurvivalGamma}
\alias{SurvivalGamma}
\title{\code{SurvivalGamma}}
\usage{
SurvivalGamma(
  k = prior_gamma(2, 0.5),
  theta = prior_gamma(2, 0.5),
  beta = prior_normal(0, 2)
)
}
\arguments{
\item{k}{(\code{Prior})\cr for the shape \code{k}.}

\item{theta}{(\code{Prior})\cr for the scale \code{theta}.}

\item{beta}{(\code{Prior})\cr for covariates coefficients \code{beta}.}
}
\description{
This class extends the general \code{\link{SurvivalModel}} class for using the
Gamma survival model.
}
