% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalLoglogistic.R
\docType{class}
\name{SurvivalLogLogistic-class}
\alias{SurvivalLogLogistic-class}
\alias{.SurvivalLogLogistic}
\alias{SurvivalLogLogistic}
\title{\code{SurvivalLogLogistic}}
\usage{
SurvivalLogLogistic(
  lambda = prior_lognormal(0, 5, init = 1/200),
  p = prior_gamma(2, 5, init = 0.5),
  beta = prior_normal(0, 5)
)
}
\arguments{
\item{lambda}{(\code{Prior})\cr for the inverse median \code{lambda}.}

\item{p}{(\code{Prior})\cr for the shape parameter \code{p}.}

\item{beta}{(\code{Prior})\cr for covariates coefficients \code{beta}.}
}
\description{
This class extends the general \code{\link{SurvivalModel}} class for using the
log-logistic survival model.
}
