% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalModel.R
\docType{class}
\name{SurvivalModel-class}
\alias{SurvivalModel-class}
\alias{.SurvivalModel}
\alias{SurvivalModel}
\title{\code{SurvivalModel}}
\usage{
SurvivalModel(stan = StanModule(), parameters = ParameterList(), ...)
}
\arguments{
\item{stan}{(\code{StanModule})\cr code containing the Stan code specification.}

\item{parameters}{(\code{ParameterList})\cr the parameter specification.}

\item{...}{additional arguments for \code{\link[=StanModel]{StanModel()}}.}
}
\description{
This class extends the general \code{\link{StanModel}} class to comprise the survival
model specification.
}
