% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\docType{class}
\name{SurvivalQuantities-class}
\alias{SurvivalQuantities-class}
\alias{.SurvivalQuantities}
\alias{SurvivalQuantities}
\title{title}
\usage{
SurvivalQuantities(quantities, groups, type, time_grid, data)
}
\arguments{
\item{object}{(\code{JointModelSamples}) \cr A \code{\link[=JointModelSamples-class]{JointModelSamples}} object}
}
\description{
#TODO
}
\seealso{
\link[=JointModelSamples-class]{JointModelSamples}

Other SurvivalQuantities: 
\code{\link{summary,SurvivalQuantities-method}}
}
\concept{SurvivalQuantities}
