% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalQuantities.R
\docType{class}
\name{SurvivalQuantities-class}
\alias{SurvivalQuantities-class}
\alias{.SurvivalQuantities}
\alias{SurvivalQuantities}
\title{\code{SurvivalQuantities} Object & Constructor Function}
\usage{
SurvivalQuantities(object, grid, type = c("surv", "haz", "loghaz", "cumhaz"))
}
\arguments{
\item{object}{(\code{\link{JointModelSamples}}) \cr samples as drawn from a Joint Model.}

\item{grid}{(\code{Grid}) \cr object that specifies which subjects and time points to calculate the
quantities for. See \link{Grid-Functions}.}

\item{type}{(\code{character})\cr quantity to be generated.
Must be one of \code{surv}, \code{haz}, \code{loghaz}, \code{cumhaz}.}
}
\description{
Constructor function to generate a \code{SurvivalQuantities} object.
}
\section{Slots}{

\describe{
\item{\code{quantities}}{(\code{Quantities})\cr The sampled quantities. Should contain 1 element per
element of \code{group}}

\item{\code{groups}}{(\code{list})\cr See argument section for details}

\item{\code{type}}{(\code{character})\cr See See argument section for details}

\item{\code{time_grid}}{(\code{numeric})\cr See argument section for details}

\item{\code{data}}{(\code{\link{DataJoint}})\cr The data that the Joint Model was fitted to to produce
the samples/quantities}
}}

\section{Group Specification}{

If \code{groups} is a character vector of subject IDs then the survival quantities will
only be calculated for those specific subjects.

If \code{groups} is a list then any elements with more than 1 subject ID will be grouped together
and their quantities will be calculated by taking a point-wise average.
For example: \code{groups = list("g1" = c("sub1", "sub2"), "g2" = c("sub3", "sub4"))} would result
in 2 groups being created whose values are the pointwise average
of \code{c("sub1", "sub2")} and \code{c("sub3", "sub4")} respectively.

If \code{groups=NULL} then all subjects from original dataset will be selected
}

\seealso{
Other SurvivalQuantities: 
\code{\link{as.data.frame.SurvivalQuantities}()},
\code{\link{autoplot.SurvivalQuantities}()},
\code{\link{brierScore.SurvivalQuantities}()},
\code{\link{summary.SurvivalQuantities}()}
}
\concept{SurvivalQuantities}
