% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalSamples.R
\docType{class}
\name{SurvivalSamples-class}
\alias{SurvivalSamples-class}
\alias{.SurvivalSamples}
\title{\code{SurvivalSamples}}
\description{
This class is an extension of \link[=JointModelSamples-class]{JointModelSamples} so that we
can define specific survival postprocessing methods for it.
}
\seealso{
Other SurvivalSamples: 
\code{\link{SurvivalSamples}},
\code{\link{autoplot,SurvivalSamples-method}},
\code{\link{predict,SurvivalSamples-method}}
}
\concept{SurvivalSamples}
