% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointModelQuantities.R
\name{validate_time_grid}
\alias{validate_time_grid}
\title{\code{validate_time_grid}}
\usage{
validate_time_grid(time_grid)
}
\arguments{
\item{time_grid}{(\code{numeric})\cr vector of time points}
}
\description{
This function performs basic assertions to ensure that a user-provided \code{time_grid}
meets the basic assumptions (ordered, finite, unique, not-NULL).
}
\keyword{internal}
