% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{write_stan}
\alias{write_stan}
\title{\code{write_stan}}
\usage{
write_stan(object, file_path)
}
\arguments{
\item{object}{the module.}

\item{file_path}{(\code{string})\cr output file.}
}
\description{
Write the Stan code for a Stan module.
}
