% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/JointModel.R
\name{write_stan}
\alias{write_stan}
\alias{write_stan.JointModel}
\title{\code{write_stan}}
\usage{
write_stan(object, destination, ...)

\method{write_stan}{JointModel}(object, destination)
}
\arguments{
\item{object}{the module.}

\item{destination}{(\code{character} or  \code{connection})\cr Where to write stan code to.}

\item{...}{Additional arguments}
}
\description{
Write the Stan code for a Stan module.
}
